### 2024-01-10 #6206 @editorialbot
[PRE REVIEW]: PyPEEC: A 3D Quasi-Magnetostatic Solver using an FFT-Accelerated PEEC Method with Voxelization
**Submitting author:** @otvam (<a href="http://orcid.org/0000-0003-0738-5823">Thomas Guillod</a>)
**Repository:** https://github.com/otvam/pypeec
**Branch with paper.md** (empty if default branch): paper_joss
**Version:** v4.8.0
