+++ docs/format/file_geometry_stl.yaml
@@ -18,8 +18,8 @@
     #   - required information, the dict cannot be empty
     #   - domain definition
     #       - list of STL meshes composing the domain
-    #       - scale: scaling factor for specifying the unit of the STL meshes
-    #       - offset: array with offsets for translating the STL meshes
+    #       - scale: scaling factor for specifying the unit of the STL meshes (before offsetting)
+    #       - offset: array with offsets for translating the STL meshes (after scaling)
     #       - filename: string with the STL filename
     "domain_stl":
         "dom_src": [{"scale": 1.0, "offset": [0.0, 0.0, 0.0], "filename": "stl/dom_src.stl"}]
