### 2024-01-15 b122592
update logs in the doc
