+++ docs/tutorial/log_solver.txt
@@ -1,129 +1,195 @@
-
-  ╔═══════════════════════════════════════════════════════╗
-  ║  ╔═══╗   ╔═══╦═══╦═══╦═══╗            3D PEEC Solver  ║
-  ║  ║╔═╗║   ║╔═╗║╔══╣╔══╣╔═╗║      Quasi-Magnetostatics  ║
-  ║  ║╚═╝╠╗ ╔╣╚═╝║╚══╣╚══╣║ ╚╝          Frequency Domain  ║
-  ║  ║╔══╣║ ║║╔══╣╔══╣╔══╣║ ╔╗          Voxel Geometries  ║
-  ║  ║║  ║╚═╝║║  ║╚══╣╚══╣╚═╝║       FFT/GPU Accelerated  ║
-  ║  ╚╝  ╚═╗ ╠╝  ╚═══╩═══╩═══╝    Python and Open Source  ║
-  ╚══════╦═╝ ╠════════════════════════════════════════════╝
-         ╚═══╝
-  ╔═══════════════════════════════════════════════════════╗
-  ║ Website - https://pypeec.otvam.ch                     ║
-  ║ Thomas Guillod - Dartmouth College                    ║
-  ║ Mozilla Public License Version 2.0                    ║
-  ╚═══════════════════════════════════════════════════════╝
-
 00:00:00.00 : MAIN     : INFO     : load the input data
-00:00:00.57 : MAIN     : INFO     : load the solver
-00:00:00.64 : MAIN     : INFO     : run the solver
-00:00:00.64 : SOLVER   : INFO     : check the input data
-00:00:00.64 : SOLVER   : INFO     : combine the input data
-00:00:00.64 : SOLVER   : INFO     : init : enter : timing
-00:00:00.64 : SOLVER   : INFO     :     voxel_geometry : enter : timing
-00:00:00.83 : SOLVER   : INFO     :     voxel_geometry : exit : 00:00:00.18
-00:00:00.83 : SOLVER   : INFO     :     system_tensor : enter : timing
-00:00:00.83 : GREEN    : DEBUG    :         analytical solution: 6D / 1
-00:00:00.85 : GREEN    : DEBUG    :         analytical solution: 6D / 4624
-00:00:01.02 : GREEN    : DEBUG    :         numerical approximation: 6D / 326432
-00:00:01.04 : SOLVER   : INFO     :     system_tensor : exit : 00:00:00.20
-00:00:01.04 : SOLVER   : INFO     :     problem_geometry : enter : timing
-00:00:01.08 : PROBLEM  : DEBUG    :         n_voxel_total = 331056
-00:00:01.08 : PROBLEM  : DEBUG    :         n_voxel_used = 21225
-00:00:01.08 : PROBLEM  : DEBUG    :         n_face_total = 993168
-00:00:01.08 : PROBLEM  : DEBUG    :         n_face_used = 42094
-00:00:01.08 : PROBLEM  : DEBUG    :         n_voxel_electric = 21225
-00:00:01.08 : PROBLEM  : DEBUG    :         n_voxel_magnetic = 0
-00:00:01.08 : PROBLEM  : DEBUG    :         n_face_electric = 42094
-00:00:01.08 : PROBLEM  : DEBUG    :         n_face_magnetic = 0
-00:00:01.08 : PROBLEM  : DEBUG    :         n_src_current = 0
-00:00:01.08 : PROBLEM  : DEBUG    :         n_src_voltage = 722
-00:00:01.08 : PROBLEM  : DEBUG    :         ratio_voxel = 6.41e-02
-00:00:01.08 : PROBLEM  : DEBUG    :         ratio_face = 4.24e-02
[…]
