+++ docs/content/examples.rst
@@ -25,7 +25,7 @@ Running the Examples
 * For running the examples from a **Jupyter notebook**, use ``notebook.ipynb``:
 
   * Jupyter should be installed (not included in the default environment).
-  * IPyWidgets and Trame are used for the rendering inside the notebook.
+  * IPyWidgets and Trame are used for the rendering inside Jupyter notebooks.
 
 STL Mesher Examples
 -------------------
