+++ docs/content/warnings.rst
@@ -11,6 +11,7 @@ PyPEEC Issues
 .. Warning::
     For problems with **magnetic domains**, the **preconditioner** is not optimal.
     This might lead to a slow convergence of the iterative matrix solver.
+    For such cases, using a segregated solver might be useful.
 
 .. Warning::
     For **large problems**, the code might allocate huge amounts of **memory**.
