### 2024-01-17 52f4407
bug for extracting solution terminals
