+++ examples/run_viewer.sh
@@ -12,7 +12,6 @@ source examples_config.sh
 
 # get the filenames
 FILE_VOXEL="$PATH_ROOT/$FOLDER_EXAMPLE/voxel.pck"
-FILE_POINT="$PATH_ROOT/$FOLDER_EXAMPLE/point.yaml"
 FILE_VIEWER="$PATH_ROOT/$FOLDER_CONFIG/viewer.yaml"
 
 # plot tag (from viewer.yaml)
@@ -21,7 +20,6 @@ TAG_PLOT="domain connection"
 # run
 pypeec viewer \
     --voxel $FILE_VOXEL \
-    --point $FILE_POINT \
     --viewer $FILE_VIEWER \
     --tag_plot $TAG_PLOT \
     --plot_mode qt
