+++ examples/run_plotter.sh
@@ -12,7 +12,6 @@ source examples_config.sh
 
 # get the filenames
 FILE_SOLUTION="$PATH_ROOT/$FOLDER_EXAMPLE/solution.pck"
-FILE_POINT="$PATH_ROOT/$FOLDER_EXAMPLE/point.yaml"
 FILE_PLOTTER="$PATH_ROOT/$FOLDER_CONFIG/plotter.yaml"
 
 # plot tag (from plotter.yaml)
@@ -21,7 +20,6 @@ TAG_PLOT="V_c_abs J_c_norm H_norm residuum"
 # run
 pypeec plotter \
     --solution $FILE_SOLUTION \
-    --point $FILE_POINT \
     --plotter $FILE_PLOTTER \
     --tag_plot $TAG_PLOT \
     --plot_mode qt
