### 2024-01-19 63e0047
clean voxel point generation
