### 2024-01-19 abf5e2d
check point cloud validity
