+++ pypeec/run/viewer.py
@@ -53,7 +53,7 @@ def _get_grid_voxel(data_geom):
     # convert the voxel geometry into PyVista grids
     grid = manage_voxel.get_grid(n, d, c)
     voxel = manage_voxel.get_voxel(grid, idx)
-    point = manage_voxel.get_point(pts_cloud, voxel)
+    point = manage_voxel.get_point(pts_cloud)
 
     # add the domain tag to the geometry
     voxel = manage_voxel.set_viewer_domain(voxel, idx, domain, connection)
