+++ examples/examples_voxel/core/geometry.yaml
@@ -41,6 +41,7 @@
     - [0.0e-3, -50.0e-3, -10.0e-3]
     - [0.0e-3, -50.0e-3, 10.0e-3]
     - [0.0e-3, -50.0e-3, 30.0e-3]
+"check_cloud": true
 "check_conflict": true
 "check_connection": true
 "domain_conflict": []
