+++ examples/examples_voxel/transformer/geometry.yaml
@@ -87,6 +87,7 @@
     - [0.004, 0.001714285714285715, 0.002]
     - [0.004, 0.0028571428571428576, 0.002]
     - [0.004, 0.004, 0.002]
+"check_cloud": true
 "check_conflict": true
 "check_connection": true
 "domain_conflict": []
