### 2024-01-19 2e5efd7
refactorize plot data management
