### 2024-01-19 3500c06
update documentation for the new point cloud format
