+++ docs/content/workflow.rst
@@ -88,7 +88,6 @@ The following input files (JSON or YAML format) are used:
 * ``file_geometry`` (JSON or YAML, input file): description of the geometry
 * ``file_problem`` (JSON or YAML, input file): description of the magnetic problem
 * ``file_tolerance`` (JSON or YAML, input file): description of the solver numerical options
-* ``file_point`` (JSON or YAML, input file): point cloud for evaluating the magnetic field
 * ``file_viewer`` (JSON or YAML, input file): options for the plots generated by the viewer
 * ``file_plotter`` (JSON or YAML, input file): options for the plots generated by the plotter
 * ``file_config`` (JSON or YAML, config file): global configuration file for PyPEEC
