+++ docs/format/file_geometry.yaml
@@ -19,6 +19,18 @@
     "use_resample": true
     "resampling_factor": [2, 2, 1]
 
+# 2D array containing the points used for magnetic field evaluation
+#   - the number of points (n_pts) can be zero.
+#   - the array has the following size: (n_pts, 3)
+#   - the points should be located outside the active voxels
+"pts_cloud":
+    - [-1.0, +1.0, +1.0],
+    - [+1.0, -1.0, +1.0],
+    - [+1.0, +1.0, -1.0],
+
+# check (or not) the validity of the point cloud
+"check_cloud": true
+
 # perform (or not) the conflict resolution between the voxels
 "check_conflict": true
 
