### 2024-01-21 4e94bab
transform PyVista mesh to numpy array for the Pickle file
