### 2024-01-22 93eaf8a
better exception handling
