### 2024-01-22 a17bfe1
cast timedelta to float
