### 2024-01-23 f01400d
new io module with JSON extensions
