+++ pypeec/io.py
@@ -21,6 +21,7 @@ import os.path
 import json
 import pickle
 import yaml
+import numpy as np
 from pypeec.error import FileError
 
 
@@ -95,9 +96,112 @@ class _YamlLoader(yaml.Loader):
             return content
 
 
+class _JsonNumPyEncoder(json.JSONEncoder):
+    """
+    This Python class offers extension to the JSON format (encoder).
+        - encode NumPy scalar types
+        - encode NumPy array types
+    """
+
+    def __init__(self, **kwargs):
+        """
+        Constructor
+        """
+
+        super().__init__(**kwargs)
+
+    def default(self, obj):
+        """
+        Function encoding NumPy types as dictionaries.
+        """
+
+        # if not numpy, default to the base encoder
+        if not hasattr(obj, "dtype") or not hasattr(obj, "ndim"):
+            return json.JSONEncoder.default(self, obj)
+
+        # handle numpy scalar
+        if np.isscalar(obj):
+            if np.iscomplexobj(obj):
+                return {
+                    "__np_scalar_complex__": None,
+                    "dtype": str(obj.dtype),
+                    "real": obj.real,
+                    "imag": obj.imag,
+                }
+            else:
+                return {
+                    "__np_scalar_direct__": None,
+                    "dtype": str(obj.dtype),
+                    "data": obj.item(),
+                }
+        else:
+            # handle numpy array
+            if np.iscomplexobj(obj):
+                return {
+                    "__np_array_complex__": None,
+                    "dtype": str(obj.dtype),
+                    "real": obj.real,
+                    "imag": obj.imag,
+                }
+            else:
+                return {
+                    "__np_array_direct__": None,
+                    "dtype": str(obj.dtype),
+                    "data": obj.tolist(),
+                }
+
+
+class _JsonNumPyDecoder(json.JSONDecoder):
+    """
+    This Python class offers extension to the JSON format (decoder).
+        - decode NumPy scalar types
+        - decode NumPy array types
+    """
+
+    def __init__(self, **kwargs):
+        """
+        Constructor
+        """
+
+        kwargs.setdefault("object_hook", self.encode)
+        super().__init__(**kwargs)
+
+    def encode(self, data):
+        """
+        Function decoding NumPy types from dictionaries.
+        """
+
+        # if not dict, do nothing
+        if not isinstance(data, dict):
+            return data
+
+        if "__np_scalar_complex__" in data:
+            dtype = np.dtype(data["dtype"])
+            real = dtype.type(data["real"])
+            imag = dtype.type(data["imag"])
+            val = real+1j*imag
+            return val
+        if "__np_scalar_direct__" in data:
+            dtype = np.dtype(data["dtype"])
+            val = dtype.type(data["data"])
+            return val
+        if "__np_array_complex__" in data:
+            dtype = np.dtype(data["dtype"])
+            real = np.array(data["real"], dtype=dtype)
+            imag = np.array(data["imag"], dtype=dtype)
+            val = real+1j*imag
+            return val
+        if "__np_array_direct__" in data:
+            dtype = np.dtype(data["dtype"])
+            val = np.array(data["data"], dtype=dtype)
+            return val
+
+        return data
+
+
 def _load_yaml(filename):
     """
-    Load a YAML file.
+    Load a YAML file (with extensions).
     """
 
     try:
@@ -111,15 +215,18 @@ def _load_yaml(filename):
     return data
 
 
-def _load_json(filename):
+def _load_json(filename, has_extensions):
     """
-    Load a JSON file.
+    Load a JSON file (with or without extensions).
     """
 
     try:
[…]
