### 2024-01-23 b1988e7
simplify some data for serialization
