### 2024-01-23 4f31334
remove the single precision data type (very complex and not really useful)
