+++ pypeec/lib_solver/equation_system.py
@@ -79,10 +79,6 @@ __license__ = "Mozilla Public License Version 2.0"
 import numpy as np
 import scipy.sparse as sps
 from pypeec.lib_matrix import matrix_factorization
-from pypeec import config
-
-# get config
-NP_TYPES = config.NP_TYPES
 
 
 def _get_system_size(A_net_c, A_net_m, A_src):
@@ -122,12 +118,12 @@ def _get_system_scaler(freq, n_vc, n_fc, n_vm, n_fm, n_src):
 
     # get the scaling vectors
     if freq == 0:
-        scaler_c = np.ones(n_fc+n_vc+n_src, dtype=NP_TYPES.COMPLEX)
-        scaler_m = np.ones(n_fm+n_vm, dtype=NP_TYPES.COMPLEX)
+        scaler_c = np.ones(n_fc+n_vc+n_src, dtype=np.complex_)
+        scaler_m = np.ones(n_fm+n_vm, dtype=np.complex_)
     else:
-        scaler_c = np.ones(n_fc+n_vc+n_src, dtype=NP_TYPES.COMPLEX)
-        scaler_fm = s*np.ones(n_fm, dtype=NP_TYPES.COMPLEX)
-        scaler_vm = np.ones(n_vm, dtype=NP_TYPES.COMPLEX)
+        scaler_c = np.ones(n_fc+n_vc+n_src, dtype=np.complex_)
+        scaler_fm = s*np.ones(n_fm, dtype=np.complex_)
+        scaler_vm = np.ones(n_vm, dtype=np.complex_)
         scaler_m = np.concatenate((scaler_fm, scaler_vm))
 
     return scaler_c, scaler_m
@@ -147,12 +143,12 @@ def _get_coupling_electric(sol_m, freq, n_vc, n_fc, n_fm, n_src, K_op_c):
 
     # compute the couplings
     if freq == 0:
-        cpl_fc = np.zeros(n_fc, dtype=NP_TYPES.COMPLEX)
+        cpl_fc = np.zeros(n_fc, dtype=np.complex_)
     else:
         cpl_fc = K_op_c(I_fm)
 
-    cpl_vc = np.zeros(n_vc, dtype=NP_TYPES.COMPLEX)
-    cpl_src = np.zeros(n_src, dtype=NP_TYPES.COMPLEX)
+    cpl_vc = np.zeros(n_vc, dtype=np.complex_)
+    cpl_src = np.zeros(n_src, dtype=np.complex_)
 
     # assemble the vectors
     cpl_c = np.concatenate((cpl_fc, cpl_vc, cpl_src))
@@ -174,7 +170,7 @@ def _get_coupling_magnetic(sol_c, n_fc, n_vm, K_op_m):
 
     # compute the couplings
     cpl_fm = -K_op_m(I_fc)
-    cpl_vm = np.zeros(n_vm, dtype=NP_TYPES.COMPLEX)
+    cpl_vm = np.zeros(n_vm, dtype=np.complex_)
 
     # assemble the vectors
     cpl_m = np.concatenate((cpl_fm, cpl_vm))
@@ -211,18 +207,18 @@ def _get_cond_fact_electric(freq, A_net_c, R_c, L_c, A_src):
     # assemble the matrices
     A_12_mat = -A_net_c.transpose()
     A_21_mat = A_net_c
-    A_22_mat = sps.csc_matrix((n_vc, n_vc), dtype=NP_TYPES.INT)
+    A_22_mat = sps.csc_matrix((n_vc, n_vc), dtype=np.int_)
 
     # expand for the source matrices
-    A_add = sps.csc_matrix((n_fc, n_src), dtype=NP_TYPES.INT)
[…]
