+++ pypeec/lib_solver/system_tensor.py
@@ -11,10 +11,6 @@ __license__ = "Mozilla Public License Version 2.0"
 import numpy as np
 import numpy.linalg as lna
 from pypeec.lib_matrix import green_function
-from pypeec import config
-
-# get config
-NP_TYPES = config.NP_TYPES
 
 
 def _get_coupling(d, idx, method, dimension):
@@ -25,7 +21,7 @@ def _get_coupling(d, idx, method, dimension):
 
     # extract the voxel data
     (dx, dy, dz) = d
-    d = np.array(d, NP_TYPES.FLOAT)
+    d = np.array(d, np.float_)
 
     # dimension permutation:
     #   - the 5D integral is solved for the xy faces
@@ -88,9 +84,9 @@ def _get_voxel_indices(n):
     (nx, ny, nz) = n
 
     # get the indices array
-    idx_x = np.arange(nx, dtype=NP_TYPES.INT)
-    idx_y = np.arange(ny, dtype=NP_TYPES.INT)
-    idx_z = np.arange(nz, dtype=NP_TYPES.INT)
+    idx_x = np.arange(nx, dtype=np.int_)
+    idx_y = np.arange(ny, dtype=np.int_)
+    idx_z = np.arange(nz, dtype=np.int_)
     [idx_x, idx_y, idx_z] = np.meshgrid(idx_x, idx_y, idx_z, indexing="ij")
 
     # flatten the indices into vectors
@@ -127,7 +123,7 @@ def get_green_self(d):
     The self-coefficient is used for the preconditioner.
     """
 
-    idx = np.array([[0, 0, 0]], dtype=NP_TYPES.INT)
+    idx = np.array([[0, 0, 0]], dtype=np.int_)
     G_self = green_function.get_green_ana(d, idx, "6D")
     G_self = G_self[0]
 
@@ -161,7 +157,7 @@ def get_green_tensor(n, d, green_simplify):
     idx_num = np.invert(idx_ana)
 
     # init the result vector
-    G_mutual = np.empty(nv, dtype=NP_TYPES.FLOAT)
+    G_mutual = np.empty(nv, dtype=np.float_)
 
     # analytical solution
     G_mutual[idx_ana] = green_function.get_green_ana(d, idx[idx_ana], "6D")
@@ -205,7 +201,7 @@ def get_coupling_tensor(n, d, coupling_simplify, has_coupling):
     idx_num = np.invert(idx_ana)
 
     # init the result vector
-    K_tsr = np.empty((nv, 3), dtype=NP_TYPES.FLOAT)
+    K_tsr = np.empty((nv, 3), dtype=np.float_)
 
     # analytical solution
     K_tsr[idx_ana, 0] = _get_coupling(d, idx[idx_ana], "ana", "yz")
