### 2024-01-23 02ca088
JSON serialization is working
