### 2024-01-23 506c9d5
JSON serialization documentation
