+++ docs/content/workflow.rst
@@ -92,7 +92,7 @@ The following input files (JSON or YAML format) are used:
 * ``file_plotter`` (JSON or YAML, input file): options for the plots generated by the plotter
 * ``file_config`` (JSON or YAML, config file): global configuration file for PyPEEC
 
-The following files (Pickle format) are generated:
+The following files (JSON or Pickle format) are generated:
 
-* ``file_voxel`` (Pickle, output file): file with the definition of the voxel structure
-* ``file_solution`` (Pickle, output file): file with the solution of the magnetic problem
+* ``file_voxel`` (JSON or Pickle, output file): file with the definition of the voxel structure
+* ``file_solution`` (JSON or Pickle, output file): file with the solution of the magnetic problem
