### 2024-01-23 0a705df
JSON gzip serialization
