### 2024-01-23 08ca88b
default use of json/gzip in place of pickle
