+++ examples/run_mesher.py
@@ -20,7 +20,7 @@ FOLDER_EXAMPLE = examples_config.FOLDER_EXAMPLE
 if __name__ == "__main__":
     # get the filenames
     file_geometry = os.path.join(PATH_ROOT, FOLDER_EXAMPLE, "geometry.yaml")
-    file_voxel = os.path.join(PATH_ROOT, FOLDER_EXAMPLE, "voxel.pck")
+    file_voxel = os.path.join(PATH_ROOT, FOLDER_EXAMPLE, "voxel.gz")
 
     # run
     try:
