### 2024-01-23 7e5bf97
bug for empty dimension serialization
