### 2024-01-23 8d78236
switch tests to JSON serialization
