+++ pypeec/lib_check/check_data_config.py
@@ -6,7 +6,6 @@ __author__ = "Thomas Guillod"
 __copyright__ = "Thomas Guillod - Dartmouth College"
 __license__ = "Mozilla Public License Version 2.0"
 
-import numpy as np
 from pypeec.lib_check import datachecker
 
 
