### 2024-01-23 c8c889c
migrate from setup.cfg to pyproject.toml
