+++ setup.cfg
@@ -1,106 +1,0 @@
-# Configuration file for defining the package data.
-# The bootstrapping data is located "pyproject.toml".
-#
-# WARNING: Only the required packages are listed as the package dependencies.
-#          The following optional libraries are not installed:
-#              - FFTW, UMFPACK, PyAMG, and MKL/PARDISO
-#              - Jupyter, IPyWidgets, and Trame
-#              - CuPy and CUDA
-#
-# Thomas Guillod - Dartmouth College
-# Mozilla Public License Version 2.0
-
-##############################################################################
-[metadata]
-##############################################################################
-
-# base data
-name = pypeec
-description = PyPEEC - 3D PEEC Solver
-
-# license
-license = MPL-2.0
-license_files = LICENSE.txt
-
-# contributors
-author = Thomas Guillod
-author_email = guillod@otvam.ch
-maintainer = Thomas Guillod
-maintainer_email = guillod@otvam.ch
-
-# websites
-url = https://pypeec.otvam.ch
-project_urls =
-	Repository = https://github.com/otvam/pypeec
-	Releases = https://github.com/otvam/pypeec/releases
-	Issues = https://github.com/otvam/pypeec/issues
-	Tags = https://github.com/otvam/pypeec/tags
-
-# description
-long_description = file: README.md
-long_description_content_type = text/markdown
-
-# package indexing
-keywords =
-    PyPEEC
-    FFT
-    PEEC
-    3D
-    voxel
-    conductor
-    electric
-    magnetic
-    field simulation
-    maxwell equations
-    frequency domain
-    power electronics
-classifiers =
-    Development Status :: 5 - Production/Stable
-    License :: OSI Approved :: Mozilla Public License 2.0 (MPL 2.0)
-    Intended Audience :: Science/Research
-    Operating System :: OS Independent
-    Programming Language :: Python :: 3
-    Topic :: Scientific/Engineering
-    Natural Language :: English
-    Environment :: Console
-
-##############################################################################
-[options]
-##############################################################################
-
-# package list
-packages =
-    pypeec
-    pypeec.lib_mesher
-    pypeec.lib_solver
-    pypeec.lib_visualization
-    pypeec.lib_check
-    pypeec.lib_matrix
-    pypeec.run
-    pypeec.utils
-    pypeec.data
-
-# dependencies
-python_requires = >=3.9
-install_requires =
-    PyYAML >= 6.0
-    joblib >= 1.3
-    numpy >= 1.24
-    scipy >= 1.9
-    matplotlib >= 3.7
-    shapely >= 2.0
-    rasterio >= 1.3
-    Pillow >= 10.0
-    vtk >= 9.2
-    pyvista >= 0.40
-    pyvistaqt >= 0.11
-    PyQt5 >= 5.15
-
-# define package data
-[options.package_data]
-pypeec = data/*.png, data/*.txt, data/*.yaml, data/*.zip
-
-# define console script
-[options.entry_points]
-console_scripts =
-    pypeec = pypeec.script:run_script
