+++ pyproject.toml
@@ -1,5 +1,8 @@
-# Configuration file for bootstraping the package.
-# The detailed configuration is located "setup.cfg".
+# Configuration file for creating the package.
+#   - standard metadata
+#   - dynamic version number
+#   - command line script
+#   - package data
 #
 # Thomas Guillod - Dartmouth College
 # Mozilla Public License Version 2.0
@@ -11,3 +14,87 @@ requires = ["setuptools", "wheel", "setuptools_scm"]
 [tool.setuptools_scm]
 write_to = "pypeec/data/version.txt"
 version_scheme = "guess-next-dev"
+
+[project]
+name = "pypeec"
+description = "PyPEEC - 3D PEEC Solver"
+license = {text = "MPL-2.0"}
+authors = [{name = "Thomas Guillod", email = "guillod@otvam.ch"}]
+maintainers = [{name = "Thomas Guillod", email = "guillod@otvam.ch"}]
+keywords = [
+    "PyPEEC",
+    "FFT",
+    "PEEC",
+    "3D",
+    "voxel",
+    "conductor",
+    "electric",
+    "magnetic",
+    "field simulation",
+    "maxwell equations",
+    "frequency domain",
+    "power electronics",
+]
+classifiers = [
+    "Development Status :: 5 - Production/Stable",
+    "License :: OSI Approved :: Mozilla Public License 2.0 (MPL 2.0)",
+    "Intended Audience :: Science/Research",
+    "Operating System :: OS Independent",
+    "Programming Language :: Python :: 3",
+    "Topic :: Scientific/Engineering",
+    "Natural Language :: English",
+    "Environment :: Console",
+]
+dependencies = [
+    "PyYAML >= 6.0",
+    "joblib >= 1.3",
+    "numpy >= 1.24",
+    "scipy >= 1.9",
+    "matplotlib >= 3.7",
+    "shapely >= 2.0",
+    "rasterio >= 1.3",
+    "Pillow >= 10.0",
+    "vtk >= 9.2",
+    "pyvista >= 0.40",
+    "pyvistaqt >= 0.11",
+    "PyQt5 >= 5.15",
+]
+requires-python = ">=3.9"
+dynamic = ["version"]
+
+[project.urls]
+Homepage = "https://pypeec.otvam.ch"
+Repository = "https://github.com/otvam/pypeec"
+Releases = "https://github.com/otvam/pypeec/releases"
+Issues = "https://github.com/otvam/pypeec/issues"
+Tags = "https://github.com/otvam/pypeec/tags"
+
+[project.readme]
+file = "README.md"
+content-type = "text/markdown"
+
+[project.scripts]
+pypeec = "pypeec.script:run_script"
+
+[tool.setuptools]
+packages = [
+    "pypeec",
+    "pypeec.lib_mesher",
+    "pypeec.lib_solver",
+    "pypeec.lib_visualization",
+    "pypeec.lib_check",
+    "pypeec.lib_matrix",
+    "pypeec.run",
+    "pypeec.utils",
+    "pypeec.data",
+]
+license-files = ["LICENSE.txt"]
+include-package-data = true
+
+[tool.setuptools.package-data]
+pypeec = [
+    "data/*.png",
+    "data/*.txt",
+    "data/*.yaml",
+    "data/*.zip",
+]
