### 2024-01-24 6d68626
simplify notebook.ipynb
