+++ examples/examples_config.sh
@@ -11,6 +11,7 @@ PATH_ROOT="."
 FOLDER_CONFIG="config"
 
 # name of the different examples
+#   tutorial
 #   examples_voxel
 #       examples_voxel/slab
 #       examples_voxel/transformer
