### 2024-01-24 4ea3292
handle all the exceptions in main.py
