### 2024-01-25 #6206 @likeajumprope
yes I can take this one
