### 2024-01-25 cd3e95e
custom logging level for blocks and exceptions
