### 2024-01-26 #6206 @likeajumprope
> Failed to discover a valid open source license

Editorialbot is incorrect here, work covered by CC BY 4.0
