### 2024-01-26 #6206 @likeajumprope
> > > Failed to discover a valid open source license
> >
> >
> > Editorialbot is incorrect here, work covered by CC BY 4.0
>
> So, CC BY is not a software license - is that just for the written work?
>
> I do see the MPL 2 license in the main repo: https://github.com/otvam/pypeec/blob/main/LICENSE.txt
>
> `editorialbot`'s issue here is that it's looking at just the paper branch `paper_joss`, which doesn't have the software license. (That's why I ran the repository checks from the `main` branch above.)
>
> > > Failed to discover a Statement of need section in paper
> >
> >
> > Editorialbot is incorrect here, statement of need is present.
>
> Yeah, that's strange - I see it too. Safe to ignore :)

Ahh thanks for the feedback!!
