### 2024-01-26 b2a2308
Paper is now referring to the version 4.12 PyPEEC
