+++ .github/workflows/run_platform.yml
@@ -30,9 +30,9 @@ jobs:
       matrix:
         os-type: ["ubuntu-latest", "macos-latest", "windows-latest"]
     steps:
-      - uses: actions/checkout@v3
+      - uses: actions/checkout@v4
       - name: Set up Python / ${{ matrix.os-type }}
-        uses: actions/setup-python@v4
+        uses: actions/setup-python@v5
         with:
           python-version: "3.10"
       - name: Install dependencies  / ${{ matrix.os-type }}
