+++ .github/workflows/run_package.yml
@@ -1,0 +1,38 @@
+# Run the tests after each push/pull event.
+# The tests are run with a single Python version on Linux.
+#
+# The environment variable "ALLOW_PLOTTING" is used to prevent warnings without display server.
+#
+# Thomas Guillod - Dartmouth College
+# Mozilla Public License Version 2.0
+
+name: Package Tests
+
+on:
+  workflow_dispatch:
+
+env:
+  ALLOW_PLOTTING: "true"
+  PYTHONIOENCODING: "utf8"
+  PYTHONUNBUFFERED: "1"
+
+permissions:
+  contents: read
+
+jobs:
+  build:
+    runs-on: "ubuntu-latest"
+    timeout-minutes: 10
+    steps:
+      - uses: actions/checkout@v4
+      - name: Set up Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: "3.10"
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+      - name: Run the tests
+        shell: bash
+        run: |
+          source run_tests.sh
