### 2024-01-27 02e4e82
update test conditions
