+++ .github/workflows/run_package.yml
@@ -34,6 +34,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
+          pip install .
       - name: Run the tests
         shell: bash
         run: |
