+++ .github/workflows/run_package.yml
@@ -12,8 +12,6 @@
 name: Test Package
 
 on:
-  push:
-  pull_request:
   workflow_dispatch:
   schedule:
     - cron: "0 2 * * 1"
