+++ .github/workflows/run_standard.yml
@@ -19,14 +19,6 @@ on:
   schedule:
     - cron: "0 2 * * *"
 
-env:
-  ALLOW_PLOTTING: "true"
-  PYTHONIOENCODING: "utf8"
-  PYTHONUNBUFFERED: "1"
-
-permissions:
-  contents: read
-
 jobs:
   build:
     runs-on: "ubuntu-latest"
