### 2024-01-28 1231dc8
add env to bash
