### 2024-01-28 5b6d93c
add env to bash
