### 2024-01-28 c716e16
add env to bash
