+++ run_tests.sh
@@ -30,6 +30,11 @@ function test_collect {
   fi
 }
 
+# global variables
+export ALLOW_PLOTTING="true"
+export PYTHONIOENCODING="utf8"
+export PYTHONUNBUFFERED="1"
+
 # init status
 ret=0
 
