+++ .github/workflows/run_platform.yml
@@ -45,5 +45,4 @@ jobs:
       - name: Run the tests / ${{ matrix.os-type }}
         shell: bash
         run: |
-          export ALLOW_PLOTTING="true"
           source run_tests.sh
