### 2024-02-02 #6206 @maple367
> 👋 @akm12k16 @LyceanEM @maple367 could you help us reviewing this paper?

I can help with the physical numerical methods section. Is it ok?
