### 2024-02-03 1f71acd
move sitemap to release script
