+++ run_build.sh
@@ -41,15 +41,6 @@ function build_docs {
   mkdir docs/_templates
   mkdir html
 
-  # timestamp for sitemap
-  LASTMOD=$(date '+%Y-%m-%d')
-
-  # export timestamp
-  export LASTMOD
-
-  # substitute the timestamp
-  cat docs/sitemap.xml | envsubst > html/sitemap.xml
-
   # build documentation
   sphinx-build -b html docs html
 
