+++ README.md
@@ -1,5 +1,7 @@
 # PyPEEC - 3D PEEC Solver
 
+![PyPEEC Banner](https://pypeec.otvam.ch/_images/banner.png)
+
 > * **Website: [pypeec.otvam.ch](https://pypeec.otvam.ch)**
 > * **Repository: [github.com/otvam/pypeec](https://github.com/otvam/pypeec)**
 > * **Package: [pypi.org/project/pypeec](https://pypi.org/project/pypeec)**
