+++ README.md
@@ -2,9 +2,11 @@
 
 ![PyPEEC Banner](https://pypeec.otvam.ch/_images/banner.png)
 
-> * **Website: [pypeec.otvam.ch](https://pypeec.otvam.ch)**
-> * **Repository: [github.com/otvam/pypeec](https://github.com/otvam/pypeec)**
-> * **Package: [pypi.org/project/pypeec](https://pypi.org/project/pypeec)**
+---
+* **Website: [pypeec.otvam.ch](https://pypeec.otvam.ch)**
+* **Repository: [github.com/otvam/pypeec](https://github.com/otvam/pypeec)**
+* **Package: [pypi.org/project/pypeec](https://pypi.org/project/pypeec)**
+---
 
 ## Summary
 
