### 2024-02-04 aba7c48
preparation for anisotropy
