+++ examples/tutorial/problem.yaml
@@ -7,11 +7,25 @@
 # Mozilla Public License Version 2.0
 
 "material_def":
-    "trace_via": {"domain_list": ["trace", "via"], "material_type": "electric", "var_type": "lumped"}
-    "src_sink": {"domain_list": ["src", "sink"], "material_type": "electric", "var_type": "lumped"}
+    "trace_via":
+        "domain_list": ["trace", "via"]
+        "material_type": "electric"
+        "orientation_type": "isotropic"
+        "var_type": "lumped"
+    "src_sink":
+        "domain_list": ["src", "sink"]
+        "material_type": "electric"
+        "orientation_type": "isotropic"
+        "var_type": "lumped"
 "source_def":
-    "src": {"domain_list": ["src"], "source_type": "voltage", "var_type": "lumped"}
-    "sink": {"domain_list": ["sink"], "source_type": "voltage", "var_type": "lumped"}
+    "src":
+        "domain_list": ["src"]
+        "source_type": "voltage"
+        "var_type": "lumped"
+    "sink":
+        "domain_list": ["sink"]
+        "source_type": "voltage"
+        "var_type": "lumped"
 "sweep_config":
     "sim_dc": null
     "sim_ac": null
