### 2024-02-04 a548cb0
anisotropy code
