### 2024-02-04 7adc926
anisotropy examples update
