+++ examples/examples_stl/transformer/problem.yaml
@@ -7,13 +7,33 @@
 # Mozilla Public License Version 2.0
 
 "material_def":
-    "pri": {"domain_list": ["pri_src", "pri_sink", "pri_coil"], "material_type": "electric", "var_type": "lumped"}
-    "sec": {"domain_list": ["sec_src", "sec_sink", "sec_coil"], "material_type": "electric", "var_type": "lumped"}
+    "pri":
+        "domain_list": ["pri_src", "pri_sink", "pri_coil"]
+        "material_type": "electric"
+        "orientation_type": "isotropic"
+        "var_type": "lumped"
+    "sec":
+        "domain_list": ["sec_src", "sec_sink", "sec_coil"]
+        "material_type": "electric"
+        "orientation_type": "isotropic"
+        "var_type": "lumped"
 "source_def":
-    "pri_src": {"domain_list": ["pri_src"], "source_type": "voltage", "var_type": "lumped"}
-    "pri_sink": {"domain_list": ["pri_sink"], "source_type": "voltage", "var_type": "lumped"}
-    "sec_src": {"domain_list": ["sec_src"], "source_type": "voltage", "var_type": "lumped"}
-    "sec_sink": {"domain_list": ["sec_sink"], "source_type": "voltage", "var_type": "lumped"}
+    "pri_src":
+        "domain_list": ["pri_src"]
+        "source_type": "voltage"
+        "var_type": "lumped"
+    "pri_sink":
+        "domain_list": ["pri_sink"]
+        "source_type": "voltage"
+        "var_type": "lumped"
+    "sec_src":
+        "domain_list": ["sec_src"]
+        "source_type": "voltage"
+        "var_type": "lumped"
+    "sec_sink":
+        "domain_list": ["sec_sink"]
+        "source_type": "voltage"
+        "var_type": "lumped"
 "sweep_config":
     "sim_default": null
 "sweep_param":
