+++ examples/examples_stl/inductor_core/problem.yaml
@@ -7,11 +7,25 @@
 # Mozilla Public License Version 2.0
 
 "material_def":
-    "coil": {"domain_list": ["src", "sink", "coil"], "material_type": "electric", "var_type": "lumped"}
-    "core": {"domain_list": ["core"], "material_type": "magnetic", "var_type": "lumped"}
+    "coil":
+        "domain_list": ["src", "sink", "coil"]
+        "material_type": "electric"
+        "orientation_type": "isotropic"
+        "var_type": "lumped"
+    "core":
+        "domain_list": ["core"]
+        "material_type": "magnetic"
+        "orientation_type": "isotropic"
+        "var_type": "lumped"
 "source_def":
-    "src": {"domain_list": ["src"], "source_type": "voltage", "var_type": "lumped"}
-    "sink": {"domain_list": ["sink"], "source_type": "voltage", "var_type": "lumped"}
+    "src":
+        "domain_list": ["src"]
+        "source_type": "voltage"
+        "var_type": "lumped"
+    "sink":
+        "domain_list": ["sink"]
+        "source_type": "voltage"
+        "var_type": "lumped"
 "sweep_config":
     "sim_default": null
 "sweep_param":
