+++ examples/examples_shape/trace/problem.yaml
@@ -7,11 +7,24 @@
 # Mozilla Public License Version 2.0
 
 "material_def":
-    "coil": {"domain_list": ["src_front", "src_back", "sink", "copper"], "material_type": "electric", "var_type": "lumped"}
+    "coil":
+        "domain_list": ["src_front", "src_back", "sink", "copper"]
+        "material_type": "electric"
+        "orientation_type": "isotropic"
+        "var_type": "lumped"
 "source_def":
-    "src_front": {"domain_list": ["src_front"], "source_type": "voltage", "var_type": "lumped"}
-    "src_back": {"domain_list": ["src_back"], "source_type": "voltage", "var_type": "lumped"}
-    "sink": {"domain_list": ["sink"], "source_type": "voltage", "var_type": "lumped"}
+    "src_front":
+        "domain_list": ["src_front"]
+        "source_type": "voltage"
+        "var_type": "lumped"
+    "src_back":
+        "domain_list": ["src_back"]
+        "source_type": "voltage"
+        "var_type": "lumped"
+    "sink":
+        "domain_list": ["sink"]
+        "source_type": "voltage"
+        "var_type": "lumped"
 "sweep_config":
     "sim_default": null
 "sweep_param":
