+++ examples/examples_voxel/transformer/problem.yaml
@@ -7,12 +7,29 @@
 # Mozilla Public License Version 2.0
 
 "material_def":
-    "pri": {"domain_list": ["pri" ,"src", "sink"], "material_type": "electric", "var_type": "lumped"}
-    "sec": {"domain_list": ["sec", "short"], "material_type": "electric", "var_type": "lumped"}
+    "pri":
+        "domain_list": ["pri" ,"src", "sink"]
+        "material_type": "electric"
+        "orientation_type": "isotropic"
+        "var_type": "lumped"
+    "sec":
+        "domain_list": ["sec", "short"]
+        "material_type": "electric"
+        "orientation_type": "isotropic"
+        "var_type": "lumped"
 "source_def":
-    "src": {"domain_list": ["src"], "source_type": "current", "var_type": "lumped"}
-    "sink": {"domain_list": ["sink"], "source_type": "voltage", "var_type": "lumped"}
-    "short": {"domain_list": ["short"], "source_type": "voltage", "var_type": "lumped"}
+    "src":
+        "domain_list": ["src"]
+        "source_type": "current"
+        "var_type": "lumped"
+    "sink":
+        "domain_list": ["sink"]
+        "source_type": "voltage"
+        "var_type": "lumped"
+    "short":
+        "domain_list": ["short"]
+        "source_type": "voltage"
+        "var_type": "lumped"
 "sweep_config":
     "sim_dc": null
     "sim_ac": "sim_dc"
