+++ examples/examples_shape/coplanar/problem.yaml
@@ -7,14 +7,38 @@
 # Mozilla Public License Version 2.0
 
 "material_def":
-    "copper": {"domain_list": ["copper"], "material_type": "electric", "var_type": "lumped"}
-    "front": {"domain_list": ["src_a", "sink_a"], "material_type": "electric", "var_type": "lumped"}
-    "back": {"domain_list": ["src_b", "sink_b"], "material_type": "electric", "var_type": "lumped"}
+    "copper":
+        "domain_list": ["copper"]
+        "material_type": "electric"
+        "orientation_type": "isotropic"
+        "var_type": "lumped"
+    "front":
+        "domain_list": ["src_a", "sink_a"]
+        "material_type": "electric"
+        "orientation_type": "isotropic"
+        "var_type": "lumped"
+    "back":
+        "domain_list": ["src_b", "sink_b"]
+        "material_type": "electric"
+        "orientation_type": "isotropic"
+        "var_type": "lumped"
 "source_def":
-    "src_a": {"domain_list": ["src_a"], "source_type": "voltage", "var_type": "lumped"}
-    "sink_a": {"domain_list": ["sink_a"], "source_type": "voltage", "var_type": "lumped"}
-    "src_b": { "domain_list": [ "src_b" ], "source_type": "voltage", "var_type": "lumped" }
-    "sink_b": {"domain_list": ["sink_b"], "source_type": "voltage", "var_type": "lumped"}
+    "src_a":
+        "domain_list": ["src_a"]
+        "source_type": "voltage"
+        "var_type": "lumped"
+    "sink_a":
+        "domain_list": ["sink_a"]
+        "source_type": "voltage"
+        "var_type": "lumped"
+    "src_b":
+        "domain_list": ["src_b"]
+        "source_type": "voltage"
+        "var_type": "lumped"
+    "sink_b":
+        "domain_list": ["sink_b"]
+        "source_type": "voltage"
+        "var_type": "lumped"
 "sweep_config":
     "sim_default": null
 "sweep_param":
