### 2024-02-04 9c1a9f4
anisotropy documentation
