+++ .github/workflows/run_standard.yml
@@ -8,7 +8,7 @@
 # Thomas Guillod - Dartmouth College
 # Mozilla Public License Version 2.0
 
-name: Standard Tests
+name: Run Standard Tests
 
 on:
   push:
