+++ docs/api/api.rst
@@ -1,0 +1,30 @@
+API Doc
+=======
+
+Main Modules
+------------
+
+.. autosummary::
+    :toctree: .
+
+    pypeec.main
+    pypeec.script
+    pypeec.error
+    pypeec.config
+    pypeec.log
+    pypeec.io
+
+Main Packages
+-------------
+
+============================  =====================================================
+``pypeec``                    Main PyPEEC package.
+``pypeec.run``                Mesher, solver, viewer, and plotter.
+``pypeec.utils``              Various independent utils for data manipulation.
+``pypeec.lib_check``          Library for checking the integrity of the input data.
+``pypeec.lib_matrix``         Library for handling and manipulating matrices.
+``pypeec.lib_mesher``         Modules used by the mesher.
+``pypeec.lib_solver``         Modules used by the solver.
+``pypeec.lib_visualization``  Modules used by the viewer and plotter.
+``pypeec.data``               Modules containing the static data.
+============================  =====================================================
