+++ docs/website/robots.txt
@@ -1,0 +1,4 @@
+User-agent: *
+Allow: /
+
+Sitemap: https://pypeec.otvam.ch/sitemap.xml
