+++ run_tests.sh
@@ -17,7 +17,7 @@ function test_run {
   echo "TEST: $1"
   echo "======================================================================"
 
-  python -W ignore:DeprecationWarning -m unittest -v "tests/$1.py"
+  python -W ignore:DeprecationWarning -m unittest "tests/$1.py" -v
   ret=$(( ret || $? ))
 }
 
@@ -42,11 +42,11 @@ export PYTHONUNBUFFERED="1"
 ret=0
 
 # run test
-test_run run_test_tutorial
-test_run run_test_voxel
-test_run run_test_shape
-test_run run_test_png
-test_run run_test_stl
+test_run test_tutorial
+test_run test_voxel
+test_run test_shape
+test_run test_png
+test_run test_stl
 
 # collect results
 test_collect
