### 2024-02-06 f8b82c9
add code coverage checks
