+++ conda_dev.yaml
@@ -18,6 +18,7 @@ dependencies:
   - pip=23.1.2
   - twine=4.0.2
   - python-build=0.10.0
+  - coverage=7.4.1
   - setuptools_scm=7.1.0
   - sphinx_rtd_theme=1.2.2
   - sphinx=6.2.1
