### 2024-02-07 79d3c40
script callable with list arguments
