### 2024-02-07 c34834a
tests for the command line script
