+++ tests/test_shape.py
@@ -28,4 +28,4 @@ main.run_hide_logo()
 
 # add the tests
 for name in name_list:
-    test_workflow.set_test(TestShape, name)
+    test_workflow.set_test(TestShape, name, name, False)
