+++ tests/test_png.py
@@ -28,4 +28,4 @@ main.run_hide_logo()
 
 # add the tests
 for name in name_list:
-    test_workflow.set_test(TestPng, name)
+    test_workflow.set_test(TestPng, name, name, False)
