+++ docs/content/technical.rst
@@ -34,7 +34,7 @@ Supported Platforms
 The main **target platform** of PyPEEC is **Linux** on x86/x64:
 
 * Linux / RedHat 7.9 on x86/x64
-* Linux / Debian 12.2 on x86/x64
+* Linux / Debian 12.4 on x86/x64
 * Linux / Ubuntu 20.04 on x86/x64
 * Linux / Ubuntu 22.04 on x86/x64
 
@@ -45,11 +45,12 @@ The following **platforms** and systems have been  **partially tested**:
 * MS Windows / Pro 10 on x86/x64
 * MS Windows / Pro 11 on x86/x64
 
-The following **GPUs** have been tested:
+The following **GPUs** have been tested (CUDA / CuPy compatible):
 
-* GPU / NVIDIA RTX 2070 Max-Q
+* GPU / NVIDIA RTX 2070
 * GPU / NVIDIA T4 Tensor
 * GPU / NVIDIA Tesla K80
+* GPU / NVIDIA RTX 3090
 
 The following **platforms** are passing the **automated tests**:
 
