### 2024-02-15 be2f3b8
prepare for the removal of PAUSE_GUI
