+++ pypeec/lib_check/datachecker.py
@@ -8,7 +8,6 @@ __license__ = "Mozilla Public License Version 2.0"
 
 import os.path
 import numpy as np
-from pypeec.error import CheckError
 
 
 def check_dict(name, data, key_list=None, sub_type=None, can_be_empty=True, can_be_none=False):
@@ -22,22 +21,22 @@ def check_dict(name, data, key_list=None, sub_type=None, can_be_empty=True, can_
 
     # check type
     if not isinstance(data, dict):
-        raise CheckError("%s: should be a dict" % name)
+        raise AssertionError("%s: should be a dict" % name)
     if (not can_be_empty) and (len(data) == 0):
-        raise CheckError("%s: should not be empty" % name)
+        raise AssertionError("%s: should not be empty" % name)
 
     # check value
     for tag, value in data.items():
         if not isinstance(tag, str):
-            raise CheckError("%s: all the keys should be strings" % name)
+            raise AssertionError("%s: all the keys should be strings" % name)
         if len(tag) == 0:
-            raise CheckError("%s: keys cannot be empty" % name)
+            raise AssertionError("%s: keys cannot be empty" % name)
         if (sub_type is not None) and not isinstance(value, sub_type):
-            raise CheckError("%s: invalid type" % name)
+            raise AssertionError("%s: invalid type" % name)
     if key_list is not None:
         for tag in key_list:
             if tag not in data:
-                raise CheckError("%s: dict is incomplete: %s" % (name, tag))
+                raise AssertionError("%s: dict is incomplete: %s" % (name, tag))
 
 
 def check_list(name, data, sub_type=None, can_be_empty=True, can_be_none=False):
@@ -51,14 +50,14 @@ def check_list(name, data, sub_type=None, can_be_empty=True, can_be_none=False):
 
     # check type
     if not isinstance(data, list):
-        raise CheckError("%s: should be a list" % name)
+        raise AssertionError("%s: should be a list" % name)
     if (not can_be_empty) and (len(data) == 0):
-        raise CheckError("%s: should not be empty" % name)
+        raise AssertionError("%s: should not be empty" % name)
 
     # check value
     for value in data:
         if (sub_type is not None) and not isinstance(value, sub_type):
-            raise CheckError("%s: invalid type" % name)
+            raise AssertionError("%s: invalid type" % name)
 
 
 def check_float_pts(name, data, size=None, is_positive=False, can_be_zero=True, can_be_empty=True, can_be_none=False):
@@ -75,7 +74,7 @@ def check_float_pts(name, data, size=None, is_positive=False, can_be_zero=True, 
 
     # check type and value
     if (not can_be_empty) and (len(data) == 0):
-        raise CheckError("%s: cannot be empty" % name)
+        raise AssertionError("%s: cannot be empty" % name)
     for data_tmp in data:
         check_float_array(name, data_tmp, size, is_positive, can_be_zero, can_be_empty)
 
@@ -94,7 +93,7 @@ def check_integer_pts(name, data, size=None, is_positive=False, can_be_zero=True
 
     # check type and value
     if (not can_be_empty) and (len(data) == 0):
-        raise CheckError("%s: cannot be empty" % name)
+        raise AssertionError("%s: cannot be empty" % name)
     for data_tmp in data:
         check_integer_array(name, data_tmp, size, is_positive, can_be_zero, can_be_empty)
 
@@ -113,19 +112,19 @@ def check_float_array(name, data, size=None, is_positive=False, can_be_zero=True
 
     # check type and value
     if len(data.shape) != 1:
-        raise CheckError("%s: invalid shape" % name)
+        raise AssertionError("%s: invalid shape" % name)
     if size is not None:
         if not (len(data) == size):
[…]
