### 2024-02-15 d397d59
get/set timestamp and indentation for parallel processes
