+++ pypeec/run/solver.py
@@ -392,8 +392,12 @@ def run(data_voxel, data_problem, data_tolerance, is_truncated=False):
     with log.BlockTimer(LOGGER, "init"):
         (data_init, data_internal) = _run_solver_init(data_solver, is_truncated)
 
+    # get the log global parameters
+    (timestamp_tmp, level_tmp) = log.get_global()
+
     # function for solving a single sweep
     def fct_compute(tag, param, init):
+        log.set_global(timestamp_tmp, level_tmp)
         with log.BlockTimer(LOGGER, "run sweep: " + tag):
             (output, init) = _run_solver_sweep(data_solver, data_internal, param, init, is_truncated)
         return output, init
