### 2024-02-15 4d5ee6c
factorize the parallel loop
