### 2024-02-15 a4c9609
collect log warning flag for the parallel loop
