+++ pypeec/log.py
@@ -26,6 +26,9 @@ EXCEPTION_TRACE = config.LOGGING_OPTIONS.EXCEPTION_TRACE
 USE_COLOR = config.LOGGING_OPTIONS.USE_COLOR
 DEF_COLOR = config.LOGGING_OPTIONS.DEF_COLOR
 
+# flag used to signal a problem
+GLOBAL_WARNING = False
+
 # global timestamp (constant over the complete run)
 GLOBAL_TIMESTAMP = time.time()
 
@@ -263,8 +266,6 @@ def log_exception(logger, ex, level="ERROR"):
         Logging level to be used.
     """
 
-    print("===================================")
-
     # remove the expression context
     ex.__context__ = None
 
@@ -279,8 +280,6 @@ def log_exception(logger, ex, level="ERROR"):
         logger.log(level, "exception : " + name, exc_info=ex)
     else:
         logger.log(level, "exception : " + name + "\n" + str(ex))
-
-    print("===================================")
 
 
 def get_timer():
@@ -356,6 +355,33 @@ def get_global():
     """
 
     return GLOBAL_TIMESTAMP, GLOBAL_LEVEL
+
+
+def set_warning(warning):
+    """
+    Set a global warning flag (used to signal issues).
+
+    Parameters
+    ----------
+    warning : boolean
+        Warning flag value to be set.
+    """
+
+    global GLOBAL_WARNING
+    GLOBAL_WARNING = warning
+
+
+def get_warning():
+    """
+    Get a global warning flag (used to signal issues).
+
+    Returns
+    -------
+    warning : boolean
+        Current value of the warning flag.
+    """
+
+    return GLOBAL_WARNING
 
 
 def get_logger(name):
