### 2024-02-15 e7b2055
factorization failure raise error (before was a warning)
