+++ examples/examples_voxel/core/problem.yaml
@@ -49,7 +49,7 @@
         "material_val":
             "terminal": {"rho_re": 1.0e-8, "rho_im": 0.0}
             "cond": {"rho_re": [1.0e-8, 1.0e-8, 1.0e-8, 1.0e-8, 1.0e-8], "rho_im": [0.0, 0.0, 0.0, 0.0, 0.0]}
-            "core": { "chi_re": [ 100.0, 100.0, 50.0 ], "chi_im": [ 50.0, 50.0, 25.0 ] }
+            "core": {"chi_re": [100.0, 100.0, 50.0], "chi_im": [50.0, 50.0, 25.0]}
             "empty": {"rho_re": 1.0e-8, "rho_im": 0.0}
         "source_val":
             "src": {"I_re": 1.0, "I_im": 0.0, "Y_re": 0.0, "Y_im": 0.0}
@@ -60,7 +60,7 @@
         "material_val":
             "terminal": {"rho_re": 1.0e-8, "rho_im": 0.0}
             "cond": {"rho_re": [1.0e-8, 1.0e-8, 1.0e-8, 1.0e-8, 1.0e-8], "rho_im": [0.0, 0.0, 0.0, 0.0, 0.0]}
-            "core": { "chi_re": [ 100.0, 100.0, 50.0 ], "chi_im": [ 50.0, 50.0, 25.0 ] }
+            "core": {"chi_re": [100.0, 100.0, 50.0], "chi_im": [50.0, 50.0, 25.0]}
             "empty": {"rho_re": 1.0e-8, "rho_im": 0.0}
         "source_val":
             "src": {"I_re": 1.0, "I_im": 0.0, "Y_re": 0.0, "Y_im": 0.0}
