### 2024-02-19 d426b65
simplify the core example
