+++ examples/examples_voxel/core/problem.yaml
@@ -16,11 +16,11 @@
         "domain_list": ["cond"]
         "material_type": "electric"
         "orientation_type": "isotropic"
-        "var_type": "distributed"
+        "var_type": "lumped"
     "core":
         "domain_list": [ "core" ]
         "material_type": "magnetic"
-        "orientation_type": "anisotropic"
+        "orientation_type": "isotropic"
         "var_type": "lumped"
     "empty":
         "domain_list": ["empty"]
@@ -48,8 +48,8 @@
         "freq": 0.0
         "material_val":
             "terminal": {"rho_re": 1.0e-8, "rho_im": 0.0}
-            "cond": {"rho_re": [1.0e-8, 1.0e-8, 1.0e-8, 1.0e-8, 1.0e-8], "rho_im": [0.0, 0.0, 0.0, 0.0, 0.0]}
-            "core": {"chi_re": [100.0, 100.0, 50.0], "chi_im": [50.0, 50.0, 25.0]}
+            "cond": {"rho_re": 1.0e-8, "rho_im": 0.0}
+            "core": {"chi_re": 100.0, "chi_im": 50.0}
             "empty": {"rho_re": 1.0e-8, "rho_im": 0.0}
         "source_val":
             "src": {"I_re": 1.0, "I_im": 0.0, "Y_re": 0.0, "Y_im": 0.0}
@@ -59,8 +59,8 @@
         "freq": 1.0e+3
         "material_val":
             "terminal": {"rho_re": 1.0e-8, "rho_im": 0.0}
-            "cond": {"rho_re": [1.0e-8, 1.0e-8, 1.0e-8, 1.0e-8, 1.0e-8], "rho_im": [0.0, 0.0, 0.0, 0.0, 0.0]}
-            "core": {"chi_re": [100.0, 100.0, 50.0], "chi_im": [50.0, 50.0, 25.0]}
+            "cond": {"rho_re": 1.0e-8, "rho_im": 0.0}
+            "core": {"chi_re": 100.0, "chi_im": 50.0}
             "empty": {"rho_re": 1.0e-8, "rho_im": 0.0}
         "source_val":
             "src": {"I_re": 1.0, "I_im": 0.0, "Y_re": 0.0, "Y_im": 0.0}
