### 2024-02-19 64c7e2f
example for anisotropic conductor
