+++ examples/examples_voxel/anisotropic/geometry.yaml
@@ -1,0 +1,29 @@
+# Definition of a "file_geometry" data.
+#
+# This file is used by the mesher.
+# Definition of the voxel structure.
+#
+# Thomas Guillod - Dartmouth College
+# Mozilla Public License Version 2.0
+
+"mesh_type": "voxel"
+"data_voxelize":
+    "param":
+        "n": [5, 5, 1]
+        "d": [10.0e-3, 10.0e-3, 10.0e-3]
+        "c": [0.0, 0.0, 0.0]
+    "domain_index":
+        "cond": [1, 2, 3, 4, 9, 14, 19]
+        "src": [0]
+        "sink": [24]
+"resampling":
+    "use_reduce": false
+    "use_resample": false
+    "resampling_factor": [1, 1, 1]
+"pts_cloud": []
+"check_cloud": true
+"check_conflict": true
+"check_connection": true
+"domain_conflict": []
+"domain_connection":
+    "connect": {"domain_group": [["cond"], ["src"], ["sink"]], "connected": true}
