+++ examples/examples_voxel/anisotropic/problem.yaml
@@ -1,0 +1,33 @@
+# Definition of a "file_problem" data.
+#
+# This file is used by the solver.
+# Definition of the magnetic problem to be solved.
+#
+# Thomas Guillod - Dartmouth College
+# Mozilla Public License Version 2.0
+
+"material_def":
+    "cond":
+        "domain_list": ["cond", "src", "sink"]
+        "material_type": "electric"
+        "orientation_type": "anisotropic"
+        "var_type": "lumped"
+"source_def":
+    "src":
+        "domain_list": ["src"]
+        "source_type": "current"
+        "var_type": "lumped"
+    "sink":
+        "domain_list": ["sink"]
+        "source_type": "voltage"
+        "var_type": "lumped"
+"sweep_config":
+    "sim_default": null
+"sweep_param":
+    "sim_default":
+        "freq": 0.0
+        "material_val":
+            "cond": {"rho_re": [1.0e-8, 1.0e-7, 1.0e-6], "rho_im": [0.0, 0.0, 0.0]}
+        "source_val":
+            "src": {"I_re": 1.0, "I_im": 0.0, "Y_re": 80.0e+3, "Y_im": 0.0}
+            "sink": {"V_re": 0.0, "V_im": 0.0, "Z_re": 6.0e-6, "Z_im": 0.0}
