+++ examples/examples_voxel/iron/problem.yaml
@@ -1,0 +1,45 @@
+# Definition of a "file_problem" data.
+#
+# This file is used by the solver.
+# Definition of the magnetic problem to be solved.
+#
+# Thomas Guillod - Dartmouth College
+# Mozilla Public License Version 2.0
+
+"material_def":
+    "terminal":
+        "domain_list": ["src", "sink"]
+        "material_type": "electric"
+        "orientation_type": "isotropic"
+        "var_type": "lumped"
+    "cond":
+        "domain_list": ["cond"]
+        "material_type": "electric"
+        "orientation_type": "isotropic"
+        "var_type": "lumped"
+    "core":
+        "domain_list": ["core"]
+        "material_type": "magnetic"
+        "orientation_type": "isotropic"
+        "var_type": "lumped"
+"source_def":
+    "src":
+        "domain_list": ["src"]
+        "source_type": "voltage"
+        "var_type": "lumped"
+    "sink":
+        "domain_list": ["sink"]
+        "source_type": "voltage"
+        "var_type": "lumped"
+"sweep_config":
+    "sim_default": null
+"sweep_param":
+    "sim_default":
+        "freq": 10.0e+3
+        "material_val":
+            "terminal": {"rho_re": 1.0e-8, "rho_im": 0.0}
+            "cond": {"rho_re": 1.0e-8, "rho_im": 0.0}
+            "core": {"chi_re": 100.0, "chi_im": 50.0}
+        "source_val":
+            "src": {"V_re": +1.0, "V_im": 0.0, "Z_re": 0.1, "Z_im": 0.0}
+            "sink": {"V_re": -1.0, "V_im": 0.0, "Z_re": 0.1, "Z_im": 0.0}
