### 2024-02-19 547ff10
checks for electromagnetic materials
