+++ examples/examples_voxel/iron/geometry.yaml
@@ -13,14 +13,15 @@
         "d": [10.0e-3, 10.0e-3, 10.0e-3]
         "c": [0.0, 0.0, 0.0]
     "domain_index":
-        "core": [75, 76, 77, 78, 79, 80, 84, 85, 89, 90, 94, 95, 96, 97, 98, 99]
+        "core": [75, 76, 78, 79, 80, 84, 85, 89, 90, 94, 95, 96, 98, 99]
+        "gnd": [97]
         "cond": [37, 62, 87, 112, 137]
         "src": [12]
         "sink": [162]
 "resampling":
     "use_reduce": true
     "use_resample": true
-    "resampling_factor": [3, 3, 3]
+    "resampling_factor": [1, 1, 1]
 "pts_cloud": []
 "check_cloud": true
 "check_conflict": true
