### 2024-02-19 fe27aa4
solver support for electromagnetic materials
