### 2024-02-19 422b52b
plotter support for electromagnetic materials
