### 2024-02-19 97c0b6c
new example for iron core
