+++ examples/examples_voxel/iron/geometry.yaml
@@ -1,32 +1,0 @@
-# Definition of a "file_geometry" data.
-#
-# This file is used by the mesher.
-# Definition of the voxel structure.
-#
-# Thomas Guillod - Dartmouth College
-# Mozilla Public License Version 2.0
-
-"mesh_type": "voxel"
-"data_voxelize":
-    "param":
-        "n": [5, 5, 7]
-        "d": [10.0e-3, 10.0e-3, 10.0e-3]
-        "c": [0.0, 0.0, 0.0]
-    "domain_index":
-        "core": [75, 76, 78, 79, 80, 84, 85, 89, 90, 94, 95, 96, 98, 99]
-        "gnd": [97]
-        "cond": [37, 62, 87, 112, 137]
-        "src": [12]
-        "sink": [162]
-"resampling":
-    "use_reduce": true
-    "use_resample": true
-    "resampling_factor": [1, 1, 1]
-"pts_cloud": []
-"check_cloud": true
-"check_conflict": true
-"check_connection": true
-"domain_conflict": []
-"domain_connection":
-    "connect": {"domain_group": [["cond"], ["src"], ["sink"]], "connected": true}
-    "core": {"domain_group": [["cond"], ["core"]], "connected": false}
