+++ examples/examples_png/iron/geometry.yaml
@@ -1,0 +1,41 @@
+# Definition of a "file_geometry" data.
+#
+# This file is used by the mesher.
+# Definition of the voxel structure.
+#
+# Thomas Guillod - Dartmouth College
+# Mozilla Public License Version 2.0
+
+"mesh_type": "png"
+"data_voxelize":
+    "param":
+        "d": [2.0e-3, 2.0e-3, 2.0e-3]
+        "c": [0.0, 0.0, 0.0]
+        "size": [19, 19]
+    "domain_color":
+        "core": [[0, 0, 0, 255]]
+        "wire": [[255, 0, 0, 255]]
+        "gnd": [[0, 255, 0, 255]]
+        "src": [[255, 255, 0, 255]]
+        "sink": [[0, 255, 255, 255]]
+    "layer_stack":
+        - {"n_layer": 1, "filename_list": !path ["png/src.png"]}
+        - {"n_layer": 2, "filename_list": !path ["png/wire.png"]}
+        - {"n_layer": 3, "filename_list": !path ["png/core.png"]}
+        - {"n_layer": 1, "filename_list": !path ["png/gnd.png"]}
+        - {"n_layer": 3, "filename_list": !path ["png/core.png"]}
+        - {"n_layer": 2, "filename_list": !path ["png/wire.png"]}
+        - {"n_layer": 1, "filename_list": !path ["png/sink.png"]}
+"resampling":
+    "use_reduce": false
+    "use_resample": false
+    "resampling_factor": [1, 1, 1]
+"pts_cloud": []
+"check_cloud": true
+"check_conflict": true
+"check_connection": true
+"domain_conflict": []
+"domain_connection":
+    "cond": {"domain_group": [["wire"], ["src"], ["sink"]], "connected": true}
+    "core": {"domain_group": [["core"], ["gnd"]], "connected": true}
+    "insulation": {"domain_group": [["wire"], ["core"]], "connected": false}
