+++ examples/examples_png/shield/problem.yaml
@@ -6,7 +6,6 @@
 # Thomas Guillod - Dartmouth College
 # Mozilla Public License Version 2.0
 
-"freq": 0.0
 "material_def":
     "coil":
         "domain_list": ["src", "sink", "coil"]
