### 2024-02-19 6b93003
documentation for the new examples
