### 2024-02-19 4e93735
documentation with the electromagnetic materials
