### 2024-02-27 #6206 @likeajumprope
> > 👋 @akm12k16 @LyceanEM @maple367 could you help us reviewing this paper?
>
> I can help with the physical numerical methods section. Is it ok?

Yes that would be fantastic. Could you give me a quick heads up if you are still available?
