### 2024-03-03 1e9b7b9
add magnetic losses for the iron core
