### 2024-03-03 8cdc4d9
test setup with env variables
