### 2024-03-04 a428318
allow disabling of shapely simplify
