+++ docs/format/file_geometry_shape.yaml
@@ -8,7 +8,7 @@
     #   - dy: voxel dimension for the y direction
     #   - dz: voxel dimension for the z direction
     #   - cz: coordinates of the voxel structure center for the z direction
-    #   - tol: tolerance for simplifying the shapes
+    #   - tol: tolerance for simplifying the shapes (null for disabling simplification)
     #   - xy_min: array with the lower corner coordinates of the voxel structure (x and y directions)
     #   - xy_max: array with the upper corner coordinates of the voxel structure (x and y directions)
     #   - alternatively, the xy_min/xy_max can be set to null and the shape bounds are kept
