### 2024-03-05 e5ff322
allow topology changes for shapely simplify
