### 2024-03-14 5154e98
update conda env (for pydiso/mkl)
