### 2024-03-14 adf2c12
add metadata in __init__.py
