+++ requirements.txt
@@ -17,14 +17,14 @@ numpy>=1.24
 scipy>=1.9
 joblib>=1.3
 
-# Qt bindings
-PyQt5>=5.15
-
-# geometry manipulation and plotting libraries
+# geometry manipulation
 vtk>=9.2
 pyvista>=0.40
-pyvistaqt>=0.11
-matplotlib>=3.7
 shapely>=2.0
 rasterio>=1.3
 Pillow>=10.0
+
+# plotting libraries
+PyQt5>=5.15
+pyvistaqt>=0.11
+matplotlib>=3.7
