+++ conda_dev.yaml
@@ -29,16 +29,16 @@ dependencies:
   - numpy=1.24.1
   - scipy=1.9.3
   - joblib=1.3.0
-  # Qt bindings
-  - pyqt=5.15.7
-  # geometry manipulation and plotting libraries
+  # geometry manipulation
   - vtk=9.2.6
   - pyvista=0.40.0
-  - pyvistaqt=0.11.0
-  - matplotlib=3.7.1
   - shapely=2.0.1
   - rasterio=1.3.8
   - pillow=10.0.0
+  # plotting libraries
+  - pyqt=5.15.7
+  - pyvistaqt=0.11.0
+  - matplotlib=3.7.1
   # optional Jupyter libs
   - jupyterlab=4.0.4
   - ipywidgets=8.1.0
