### 2024-03-18 b59697f
simplify utils for symmetry
