### 2024-03-21 86b6705
yaml merge command for list and dict
