### 2024-03-21 9899316
reduce frequency of cron run to the weekly run
