+++ .github/workflows/run_standard.yml
@@ -14,8 +14,6 @@ on:
   push:
   pull_request:
   workflow_dispatch:
-  schedule:
-    - cron: "0 2 * * *"
 
 jobs:
   build:
