### 2024-03-21 9cc8c1e
io.py documentation
