### 2024-03-24 5e5b359
add version and date to the doc
