### 2024-03-25 #6206 @likeajumprope
> > Hi @thelfer, would you be able to help with this one?
>
> @likeajumprope I see that you made this proposal three weeks ago, but I did not get any notification. Maybe my e-mail have treated this request as a spam. Anyway my apologies for this delay. I have no experience with magnetostatic and will be very busy in the next three weeks. However, the project seems interesting as I work with lot of FFT solvers in mechanics. What would be deadline ?

Hi @thelfer, we are still trying to find reviewers so we could be a bit more lenient with the deadline. If you could provide the review within 1 month -ish that would be fine. Your help would be greatly appreciated
