+++ docs/tutorial/log_solver.txt
@@ -1,195 +1,195 @@
 00:00:00.00 : MAIN     : INFO     : load the input data
-00:00:00.68 : MAIN     : INFO     : load the solver
-00:00:00.91 : MAIN     : INFO     : run the solver
-00:00:00.91 : SOLVER   : INFO     : check the voxel data
-00:00:00.91 : SOLVER   : INFO     : check the input data
-00:00:00.92 : SOLVER   : INFO     : combine the input data
-00:00:00.92 : SOLVER   : INFO     : init : enter : timing
-00:00:00.95 : SOLVER   : INFO     :     voxel_geometry : enter : timing
-00:00:01.16 : SOLVER   : INFO     :     voxel_geometry : exit : 00:00:00.21
-00:00:01.16 : SOLVER   : INFO     :     system_tensor : enter : timing
-00:00:01.16 : GREEN    : DEBUG    :         analytical solution: 6D / 1
-00:00:01.20 : GREEN    : DEBUG    :         analytical solution: 6D / 4624
-00:00:01.39 : GREEN    : DEBUG    :         numerical approximation: 6D / 326432
-00:00:01.42 : SOLVER   : INFO     :     system_tensor : exit : 00:00:00.25
-00:00:01.42 : SOLVER   : INFO     :     problem_geometry : enter : timing
-00:00:01.48 : PROBLEM  : DEBUG    :         n_voxel_total = 331056
-00:00:01.48 : PROBLEM  : DEBUG    :         n_voxel_used = 21215
-00:00:01.48 : PROBLEM  : DEBUG    :         n_face_total = 993168
-00:00:01.48 : PROBLEM  : DEBUG    :         n_face_used = 42062
-00:00:01.48 : PROBLEM  : DEBUG    :         n_voxel_electric = 21215
-00:00:01.48 : PROBLEM  : DEBUG    :         n_voxel_magnetic = 0
-00:00:01.48 : PROBLEM  : DEBUG    :         n_face_electric = 42062
-00:00:01.48 : PROBLEM  : DEBUG    :         n_face_magnetic = 0
-00:00:01.48 : PROBLEM  : DEBUG    :         n_src_current = 0
-00:00:01.48 : PROBLEM  : DEBUG    :         n_src_voltage = 722
-00:00:01.48 : PROBLEM  : DEBUG    :         ratio_voxel = 6.41e-02
-00:00:01.48 : PROBLEM  : DEBUG    :         ratio_face = 4.24e-02
-00:00:01.48 : SOLVER   : INFO     :     problem_geometry : exit : 00:00:00.06
-00:00:01.48 : SOLVER   : INFO     :     system_matrix : enter : timing
-00:00:01.49 : FFT      : DEBUG    :         multiplication: inductance
-00:00:01.49 : FFT      : DEBUG    :             tensor size: (132, 132, 19)
-00:00:01.49 : FFT      : DEBUG    :             tensor footprint: 40.41 MB
-00:00:01.58 : SOLVER   : INFO     :     system_matrix : exit : 00:00:00.09
-00:00:01.58 : SOLVER   : INFO     : init : exit : 00:00:00.65
-00:00:01.58 : SOLVER   : INFO     : run sweep: sim_dc : enter : timing
-00:00:01.58 : SOLVER   : INFO     :     problem_value : enter : timing
-00:00:01.62 : SOLVER   : INFO     :     problem_value : exit : 00:00:00.03
-00:00:01.62 : SOLVER   : INFO     :     equation_system : enter : timing
-00:00:01.65 : FACTOR   : DEBUG    :         factorization: electric
-00:00:01.65 : FACTOR   : DEBUG    :             matrix size: (21937, 21937)
[…]
