### 2024-03-26 e0d3d20
update documentation with warning about voxel structures
