### 2024-03-30 eb91033
bash bug with exit code
