+++ run_build.sh
@@ -50,14 +50,14 @@ function build_docs {
   # check external links
   sphinx-build -b linkcheck docs html
 
+  # update status
+  ret=$(( ret || $? ))
+
   # clean data
   rm -rf html/.buildinfo
   rm -rf html/.doctrees
   rm -rf html/output.json
   rm -rf html/output.txt
-
-  # update status
-  ret=$(( ret || $? ))
 }
 
 function build_package {
