### 2024-04-01 0b9c33c
simplify utils for impedance matrix
