### 2024-04-02 a0b50c4
options for stopping the iterative solver when the complex power has converged
