+++ pypeec/lib_solver/equation_solver.py
@@ -17,12 +17,32 @@ from pypeec import log
 LOGGER = log.get_logger("EQUATION")
 
 
+class PowerConvergenceError(RuntimeError):
+    """
+    Simple exception for signaling the convergence of the complex power.
+    """
+
+    def __init__(self, status, sol):
+        """
+        Constructor.
+        Init the exception.
+        Save the solution.
+        """
+
+        # create the exception
+        super().__init__("complex power has converged")
+
+        # save the solution
+        self.status = status
+        self.sol = sol
+
+
 class _IterCounter:
     """
     Simple class used as a callback to monitor the solver iterations.
     """
 
-    def __init__(self, fct_conv):
+    def __init__(self, fct_conv, power_options):
         """
         Constructor.
         Init the counters.
@@ -30,16 +50,22 @@ class _IterCounter:
 
         # assign data
         self.fct_conv = fct_conv
+        self.n_min = power_options["n_min"]
+        self.rel_tol = power_options["rel_tol"]
+        self.abs_tol = power_options["abs_tol"]
 
         # init data
         self.n_iter = 0
         self.power_vec = []
         self.power_final = None
         self.power_init = None
+        self.sol = None
 
     def get_callback_run(self, sol):
         """
         Callback displaying and saving the iteration.
+        Check the convergence on the complex power.
+        If convergence is achieved, stop the solver and save the solution.
         """
 
         # update the iteration
@@ -54,6 +80,21 @@ class _IterCounter:
 
         # log the results
         LOGGER.debug(f"i = {iter_tmp:d} / {power_tmp:.2e} VA")
+
+        # check for convergence
+        if self.n_iter >= np.maximum(2, self.n_min):
+            # get complex power
+            power_ref = self.power_vec[-1]
+            power_cmp = self.power_vec[-2]
+            power_err = power_ref-power_cmp
+
+            # get convergence status
+            power_thr = np.maximum(self.rel_tol*np.abs(power_ref), self.abs_tol)
+            status = np.abs(power_err) <= power_thr
+
+            # if convergence is achieved, stop the solver and save the solution
+            if status:
+                raise PowerConvergenceError(status, sol)
 
     def get_callback_init(self, sol):
         """
@@ -82,6 +123,13 @@ class _IterCounter:
 
         # log the results
         LOGGER.debug(f"final / {power_tmp:.2e} VA")
+
+    def get_sol(self):
+        """
+        Get the temporary solution.
+        """
+
+        return self.sol
 
     def get_n_iter(self):
         """
@@ -162,6 +210,61 @@ class _OpCounter:
         return self.n_sys_eval
 
 
+def _fct_pcd_all(rhs_tmp, rhs, fct_pcd):
+    """
+    Function describing the preconditioner for the coupling system.
+    """
+
+    # extract
+    (rhs_c, rhs_m) = rhs
+    (fct_pcd_c, fct_pcd_m) = fct_pcd
+
+    # get problem size
+    n_dof_c = len(rhs_c)
+    n_dof_m = len(rhs_m)
+
+    # split vector
+    rhs_c = rhs_tmp[0:n_dof_c]
+    rhs_m = rhs_tmp[n_dof_c:n_dof_c+n_dof_m]
+
+    # solve the preconditioner
+    sol_c = fct_pcd_c(rhs_c)
+    sol_m = fct_pcd_m(rhs_m)
+
[…]
