### 2024-04-02 a19d409
check for the tolerance format
