### 2024-04-02 e1a34c7
finer options for checking the solver convergence
