### 2024-04-02 446fe14
new tests with the power convergence
