### 2024-04-02 7e6bbdb
disable switch for convergence stop
