+++ examples/config/tolerance.yaml
@@ -82,7 +82,8 @@
 
     # criteria for interrupting the solver if the complex power has converged
     "power_options":
-        "n_min": 2                     # minimum number of iterations
+        "stop": true                   # use stop criteria on the complex compler
+        "n_min": 2                     # minimum number of iterations of the solver
         "rel_tol": 1.0e-4              # relative tolerance for the complex power
         "abs_tol": 1.0e-10             # absolute tolerance for the complex power
 
