### 2024-04-02 e260120
ensure convergence over several iterations
