+++ examples/config/tolerance.yaml
@@ -83,7 +83,8 @@
     # criteria for interrupting the solver if the complex power has converged
     "power_options":
         "stop": true                   # use stop criteria on the complex compler
-        "n_min": 2                     # minimum number of iterations of the solver
+        "n_min": 3                     # minimum number of iterations of the solver
+        "n_cmp": 2                     # number of iterations to consider for the convergence
         "rel_tol": 1.0e-4              # relative tolerance for the complex power
         "abs_tol": 1.0e-10             # absolute tolerance for the complex power
 
