### 2024-04-02 9fddb2d
scipy version is changing gmres results => adapt test tolerance
