### 2024-04-02 d9b3ecc
documentation of the file format
