### 2024-04-03 ba15b00
remove warning flag as it is not thread safe
