+++ pypeec/lib_solver/equation_solver.py
@@ -561,7 +561,6 @@ def get_solver(sol_init, fct_cpl, fct_sys, fct_pcd, rhs, fct_conv, solver_option
             LOGGER.debug("convergence achieved")
         else:
             LOGGER.warning("convergence issues")
-            log.set_warning(True)
 
     return sol, res, conv, status, solver_status
 
@@ -611,7 +610,6 @@ def get_condition(S_mat_c, S_mat_m, conditions_options):
                 LOGGER.debug("matrix condition is good")
             else:
                 LOGGER.warning("matrix condition is problematic")
-                log.set_warning(True)
         else:
             LOGGER.debug("matrix condition check is disabled")
 
