### 2024-04-03 eb61e8d
simplify main solver
