### 2024-04-03 1cea3f2
simplify log timestamp calculation
