### 2024-04-03 be2b4ba
refactor variable names
