+++ pypeec/run/mesher.py
@@ -223,7 +223,7 @@ def _get_data(data_geom, timestamp, is_truncated):
     """
 
     # end message
-    (seconds, span, date) = log.get_duration(timestamp)
+    (seconds, duration, date) = log.get_duration(timestamp)
 
     # get status
     is_successful = True
@@ -231,7 +231,7 @@ def _get_data(data_geom, timestamp, is_truncated):
     # extract the solution
     data_voxel = {
         "date": date,
-        "span": span,
+        "duration": duration,
         "seconds": seconds,
         "is_truncated": is_truncated,
         "is_successful": is_successful,
