### 2024-04-03 5639f47
solver return flag bug
