### 2024-04-10 c9a1f18
io.py small cleaning
