### 2024-04-11 13640ae
move SWEEP_POOL from config to tolerance.yaml
