### 2024-04-11 cf2c0ce
remove PROBLEM_MAX_SIZE config
