+++ pypeec/data/config.yaml
@@ -43,11 +43,6 @@ LOGGING_OPTIONS:
         CL_CRITICAL: '[38;5;196m' # critical level color code
         CL_RESET: '[0m' # terminal reset color sequence
 
-# maximum number of voxels
-PROBLEM_MAX_SIZE:
-    VOXEL_TOTAL: null  # maximum size of the voxel structure (null for unlimited)
-    VOXEL_USED: null  # maximum number of used voxels (null for unlimited)
-
 # delay (in seconds) between the Qt plots
 #   - this parameter should not be required (null for disable)
 #   - however, it seems to mitigate a race condition in PyVista/PyVistaQt
