### 2024-04-11 7aa5ec2
log config is autonomous, remove config module
