### 2024-04-11 9c58681
documentation of the new config format
