+++ docs/api/pypeec.config.rst
@@ -1,5 +1,0 @@
-﻿pypeec.config
-=============
-
-.. automodule:: pypeec.config
-    :members:
