### 2024-04-11 efc1e6d
PYTHONLOGGER env var for the logging
