+++ docs/content/format.rst
@@ -56,6 +56,7 @@ Other File Formats
 ------------------
 
 * The logger configuration file format is documented in ``pypeec/data/logger.yaml``.
+* The logger configuration can be changed with the ``PYTHONLOGGER`` environment variable.
 * The ``file_tolerance`` format is documented in ``examples/config/tolerance.yaml``.
 * The ``file_viewer`` format is documented in ``examples/config/viewer.yaml``.
 * The ``file_plotter`` format is documented in ``examples/config/plotter.yaml``.
