### 2024-04-11 ff77754
simplify PyPEEC logo display (avoid side-effect)
