### 2024-04-11 80b93bc
better yaml error message
