### 2024-04-12 e666336
bug for main file (double logo display)
