+++ pypeec/main.py
@@ -24,6 +24,98 @@ LOGGER = log.get_logger("MAIN")
 SHOW_LOGO = True
 
 
+def _run_mesher(data_geometry, **kwargs):
+    """
+    Load, run, and catch exceptions for the mesher.
+    """
+
+    # execute workflow
+    try:
+        # load the tool
+        LOGGER.info("load the mesher")
+        from pypeec.run import mesher
+
+        # run the tool
+        LOGGER.info("run the mesher")
+        data_voxel = mesher.run(data_geometry, **kwargs)
+    except Exception as ex:
+        log.log_exception(LOGGER, ex)
+        LOGGER.error("invalid termination")
+        raise ex
+    else:
+        LOGGER.info("successful termination")
+
+    return data_voxel
+
+
+def _run_viewer(data_voxel, data_viewer, **kwargs):
+    """
+    Load, run, and catch exceptions for the viewer.
+    """
+
+    # execute workflow
+    try:
+        # load the tool
+        LOGGER.info("load the viewer")
+        from pypeec.run import viewer
+
+        # run the tool
+        LOGGER.info("run the viewer")
+        viewer.run(data_voxel, data_viewer, **kwargs)
+    except Exception as ex:
+        log.log_exception(LOGGER, ex)
+        LOGGER.error("invalid termination")
+        raise ex
+    else:
+        LOGGER.info("successful termination")
+
+
+def _run_solver(data_voxel, data_problem, data_tolerance, **kwargs):
+    """
+    Load, run, and catch exceptions for the plotter.
+    """
+
+    # execute workflow
+    try:
+        # load the tool
+        LOGGER.info("load the solver")
+        from pypeec.run import solver
+
+        # run the tool
+        LOGGER.info("run the solver")
+        data_solution = solver.run(data_voxel, data_problem, data_tolerance, **kwargs)
+    except Exception as ex:
+        log.log_exception(LOGGER, ex)
+        LOGGER.error("invalid termination")
+        raise ex
+    else:
+        LOGGER.info("successful termination")
+
+    return data_solution
+
+
+def _run_plotter(data_solution, data_plotter, **kwargs):
+    """
+    Load, run, and catch exceptions for the mesher.
+    """
+
+    # execute workflow
+    try:
+        # load the tool
+        LOGGER.info("load the plotter")
+        from pypeec.run import plotter
+
+        # run the tool
+        LOGGER.info("run the plotter")
+        plotter.run(data_solution, data_plotter, **kwargs)
+    except Exception as ex:
+        log.log_exception(LOGGER, ex)
+        LOGGER.error("invalid termination")
+        raise ex
+    else:
+        LOGGER.info("successful termination")
+
+
 def run_display_logo():
     """
     Display the logo as a splash screen.
@@ -111,20 +203,7 @@ def run_mesher_data(data_geometry, **kwargs):
     run_display_logo()
 
     # execute workflow
-    try:
-        # load the tool
-        LOGGER.info("load the mesher")
-        from pypeec.run import mesher
-
-        # run the tool
-        LOGGER.info("run the mesher")
-        data_voxel = mesher.run(data_geometry, **kwargs)
-    except Exception as ex:
-        log.log_exception(LOGGER, ex)
-        LOGGER.error("invalid termination")
-        raise ex
-    else:
-        LOGGER.info("successful termination")
+    data_voxel = _run_mesher(data_geometry, **kwargs)
 
     return data_voxel
 
@@ -161,7 +240,7 @@ def run_mesher_file(file_geometry, file_voxel, **kwargs):
         raise ex
 
     # run the tool
[…]
