### 2024-04-12 ec255f5
if else cleanup
