### 2024-04-12 3f50f2a
add voxel points to solver output
