### 2024-04-16 3112f88
allow empty shapes for the mesher
