### 2024-04-18 otvam/pypeec#5 @thelfer
[joss review] python requirements ?
[This is issue is part of the review of the paper submited to JOSS.]

First, I would recommend an `INSTALL` file at the root of the sources. On the first contact with the library, I didn't see any installation instructions and had to figure out that I had to use pip from the "Project links" section of the `README`.
The installation instructions are somehow hidden in the tutorial, which is is not so obvious. At least the `INSTALL` file could link to the tutorial, IMHO. Also the root of the directory contains file related to `conda`, which is not mentioned in the installation section.

Regarding first contact with `pypeec`, there are a few scripts called `run...` at the root of the sources which seems tools for developers. The prefix `run` seems misleading and users may be tempted to have a look at them. That is only an **totally discardable** advice, but maybe such scripts would be better in a dedicated subdirectory (something like `tools` for instance).

I followed your installation guide and it failed as follows:

~~~~
(pypeec-env) $ python3 -m pip install pypeec
Collecting pypeec
  Could not find a version that satisfies the requirement pypeec (from versions: )
No matching distribution found for pypeec
(pypeec-env)$ python3
Python 3.7.3 (default, Oct 11 2023, 09:51:27)
[GCC 8.3.0] on linux
~~~~

I guess that my version of python is too old. Would you tell me some minimal requirements for `pypeec` ?
