### 2024-04-19 otvam/pypeec#5 @otvam
Thank you very much for accepting this review and for the comments!

Concerning, the installation instruction, I will add a "INSTALL.md" file at the root of the repository. You are totally right that the "run_*.sh" are maintenance scripts that are only useful for the maintainers, so I will move them to a "scripts" directory.

For the installation, the supported Python version are ["3.9", "3.10", "3.11", "3.12"]. I know that several Linux distributions are still shipping older versions but some libraries used by PyPEEC requires a more recent version.

Probably, conda is the easiest way to create a environment with a compatible Python interpreter:

```
conda create -n pypeec python=3.10         # create the conda env with Python
conda activate pypeec                      # activate the environment
conda install -c conda-forge libstdcxx-ng  # the env might be missing this system library for the plots
python -m pip install pypeec               # install pypeec and the dependencies
pypeec -v                                  # check the installation
```
