### 2024-04-19 otvam/pypeec#5 @thelfer
No problem.  Thanks for taking my remarks into account.

> For the installation, the supported Python version are ["3.9", "3.10", "3.11", "3.12"]. I know that several Linux distributions are still shipping older versions but some libraries used by PyPEEC requires a more recent version.

This shall be cleary stated in the `INSTALL.md` file. I did not check yet what is you CI procedure if any, but tested configurations is a very important information for end users (i.e. I would have known that I needed a more recent python version).

> Probably, conda is the easiest way to create a environment with a compatible Python interpreter:

Thanks. My point was not to use conda, but that conda's usage was not documented. And I naively though that a conda package was available, i.e. that pip was not the only way of installing `PyPEEC`.
