### 2024-04-19 4eba4e6
create scripts directory for bash scripts
