### 2024-04-19 3ac1705
update doc and actions paths for the scripts
