+++ .github/workflows/run_standard.yml
@@ -32,4 +32,4 @@ jobs:
       - name: Run the tests
         shell: bash
         run: |
-          source run_tests.sh
+          ./scripts/run_tests.sh
