+++ .github/workflows/run_platform.yml
@@ -35,4 +35,4 @@ jobs:
       - name: Run the tests / ${{ matrix.os-type }}
         shell: bash
         run: |
-          source run_tests.sh
+          ./scripts/run_tests.sh
