+++ .github/workflows/run_package.yml
@@ -31,4 +31,4 @@ jobs:
       - name: Run the tests
         shell: bash
         run: |
-          source run_tests.sh
+          ./scripts/run_tests.sh
