+++ .github/workflows/run_version.yml
@@ -35,4 +35,4 @@ jobs:
       - name: Run the tests / ${{ matrix.python-version }}
         shell: bash
         run: |
-          source run_tests.sh
+          ./scripts/run_tests.sh
