+++ docs/content/tutorial.rst
@@ -1,28 +1,14 @@
 Tutorial
 ========
 
-Installing the Code
--------------------
-
-.. code-block:: bash
-
-    # Installation of PyPEEC with pip
-    #    - The Python executable can be "python" or "python3"
-    #    - Alternatively, Conda can be used for the environment
-
-    # Create a Python Virtual Environment
-    python -m venv venv
-
-    # Activate the Python Virtual Environment
-    source venv/bin/activate
-
-    # Install PyPEEC from PyPi
-    python -m pip install pypeec
-
 Extracting the Data
 -------------------
 
 .. code-block:: bash
+
+    # Install PyPEEC
+    #   - With a Python Environment
+    #   - With a Conda Environment
 
     # Check the PyPEEC version
     pypeec --version
