+++ .github/workflows/run_package.yml
@@ -31,4 +31,5 @@ jobs:
       - name: Run the tests
         shell: bash
         run: |
+          chmod +x scripts/run_tests.sh
           ./scripts/run_tests.sh
