+++ .github/workflows/run_standard.yml
@@ -32,4 +32,5 @@ jobs:
       - name: Run the tests
         shell: bash
         run: |
+          chmod +x scripts/run_tests.sh
           ./scripts/run_tests.sh
