+++ requirements.txt
@@ -10,21 +10,21 @@
 # Mozilla Public License Version 2.0
 
 # YAML parsing library
-PyYAML>=6.0
+PyYAML~=6.0
 
 # numerical libraries
-numpy>=1.24
-scipy>=1.9
-joblib>=1.3
+numpy~=1.24
+scipy~=1.9
+joblib~=1.3
 
 # geometry manipulation
-vtk>=9.2
-pyvista>=0.40
-shapely>=2.0
-rasterio>=1.3
-Pillow>=10.0
+vtk~=9.2
+pyvista~=0.40
+shapely~=2.0
+rasterio~=1.3
+Pillow~=10.0
 
 # plotting libraries
-PyQt5>=5.15
-pyvistaqt>=0.11
-matplotlib>=3.7
+PyQt5~=5.15
+pyvistaqt~=0.11
+matplotlib~=3.7
