+++ INSTALL.md
@@ -1,4 +1,11 @@
-# PyPEEC - Installation
+**# PyPEEC - Installation
+
+## Dependencies Description
+
+* `requirements.txt` - list of Python dependencies (pinned versions)
+* `pyproject.toml` - package definition with dependencies (minimum versions)
+* `conda_base.yaml` - Conda file with the minimum requirements for PyPEEC
+* `conda_dev.yaml` - Conda file including the optional and development packages
 
 ## Using a Python Environment
 
@@ -6,6 +13,7 @@
 # Install a Python interpreter
 #   - Python website: https://www.python.org
 #   - Supported versions: 3.9, 3.10, 3.11, and 3.12
+#   - Python executable: "python" or "python3"
 
 # Create a Python Virtual Environment
 python -m venv venv
