+++ .github/workflows/run_conda.yml
@@ -26,7 +26,7 @@ jobs:
         with:
           python-version: "3.10"
           activate-environment: pypeec
-          environment-file: conda_base.yml
+          environment-file: conda_base.yaml
       - name: Run the tests
         shell: bash
         run: |
