+++ .github/workflows/run_conda.yml
@@ -27,6 +27,7 @@ jobs:
           python-version: "3.10"
           activate-environment: pypeec
           environment-file: conda_base.yaml
+          auto-activate-base: false
       - name: Run the tests
         shell: bash
         run: |
