### 2024-04-19 45f713e
conda bash options
