+++ .github/workflows/run_conda.yml
@@ -29,7 +29,7 @@ jobs:
           environment-file: conda_base.yaml
           auto-activate-base: false
       - name: Run the tests
-        shell: bash
+        shell: bash -el {0}
         run: |
           conda info
           chmod +x scripts/run_tests.sh
