+++ .github/workflows/run_conda.yml
@@ -17,17 +17,26 @@ on:
 
 jobs:
   build:
+    strategy:
+      fail-fast: false
+      matrix:
+        conda:
+          - name: "base"
+            file: "conda_base.yaml"
+          - name: "dev"
+            file: "conda_dev.yaml"
+    name: OS ${{ matrix.matrix.name }}
     runs-on: "ubuntu-latest"
-    timeout-minutes: 10
+    timeout-minutes: 20
     steps:
       - uses: actions/checkout@v4
-      - name: Set up Python
+      - name: Set up Conda
         uses: conda-incubator/setup-miniconda@v3
         with:
           python-version: "3.10"
-          activate-environment: pypeec
-          environment-file: conda_base.yaml
           auto-activate-base: false
+          activate-environment: pypeec
+          environment-file: ${{ matrix.matrix.file }}
       - name: Run the tests
         shell: bash -el {0}
         run: |
