+++ .github/workflows/run_standard.yml
@@ -25,7 +25,7 @@ jobs:
             cmd: "pip install -r requirements.txt"
           - name: "package"
             cmd: "pip install ."
-    name: OS ${{ matrix.method.name }}
+    name: Standard / ${{ matrix.method.name }}
     runs-on: "ubuntu-latest"
     timeout-minutes: 10
     steps:
