+++ .github/workflows/run_package.yml
@@ -1,35 +1,0 @@
-# Run the tests with the Python package (on Linux / default Python version).
-#
-# Run with the following conditions:
-#   - dispatch
-#   - weekly
-#
-# Thomas Guillod - Dartmouth College
-# Mozilla Public License Version 2.0
-
-name: Test Python Package
-
-on:
-  workflow_dispatch:
-  schedule:
-    - cron: "0 2 * * 1"
-
-jobs:
-  build:
-    runs-on: "ubuntu-latest"
-    timeout-minutes: 10
-    steps:
-      - uses: actions/checkout@v4
-      - name: Set up Python
-        uses: actions/setup-python@v5
-        with:
-          python-version: "3.10"
-      - name: Install dependencies
-        run: |
-          python -m pip install --upgrade pip
-          pip install .
-      - name: Run the tests
-        shell: bash
-        run: |
-          chmod +x scripts/run_tests.sh
-          ./scripts/run_tests.sh
