+++ .github/workflows/run_platform.yml
@@ -16,23 +16,26 @@ on:
 
 jobs:
   build:
-    runs-on: ${{ matrix.os-type }}
-    timeout-minutes: 20
     strategy:
       fail-fast: false
       matrix:
-        os-type: ["ubuntu-latest", "macos-latest", "windows-latest"]
+        system: ["ubuntu-latest",  "macos-latest", "windows-latest"]
+        version: ["3.9", "3.10", "3.11", "3.12"]
+
+    name: Platform / ${{ matrix.system }} / ${{ matrix.version }}
+    runs-on: ${{ matrix.system }}
+    timeout-minutes: 20
     steps:
       - uses: actions/checkout@v4
-      - name: Set up Python / ${{ matrix.os-type }}
+      - name: Set up Python
         uses: actions/setup-python@v5
         with:
-          python-version: "3.10"
-      - name: Install dependencies  / ${{ matrix.os-type }}
+          python-version: ${{ matrix.version }}
+      - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
           pip install -r requirements.txt
-      - name: Run the tests / ${{ matrix.os-type }}
+      - name: Run the tests
         shell: bash
         run:
           chmod +x scripts/run_tests.sh
