+++ .github/workflows/run_version.yml
@@ -1,39 +1,0 @@
-# Run the tests with different Python versions (on Linux).
-#
-# Run with the following conditions:
-#   - dispatch
-#   - weekly
-#
-# Thomas Guillod - Dartmouth College
-# Mozilla Public License Version 2.0
-
-name: Test Python Versions
-
-on:
-  workflow_dispatch:
-  schedule:
-    - cron: "0 2 * * 1"
-
-jobs:
-  build:
-    runs-on: "ubuntu-latest"
-    timeout-minutes: 20
-    strategy:
-      fail-fast: false
-      matrix:
-        python-version: ["3.9", "3.10", "3.11", "3.12"]
-    steps:
-      - uses: actions/checkout@v4
-      - name: Set up Python / ${{ matrix.python-version }}
-        uses: actions/setup-python@v5
-        with:
-          python-version: ${{ matrix.python-version }}
-      - name: Install dependencies  / ${{ matrix.python-version }}
-        run: |
-          python -m pip install --upgrade pip
-          pip install -r requirements.txt
-      - name: Run the tests / ${{ matrix.python-version }}
-        shell: bash
-        run: |
-          chmod +x scripts/run_tests.sh
-          ./scripts/run_tests.sh
