+++ .github/workflows/run_standard.yml
@@ -1,9 +1,14 @@
-# Run the standard tests (on Linux / default Python version).
-#
-# Run with the following conditions:
+# Run the Standard tests with the following conditions.
 #   - push/pull
 #   - dispatch
-#   - daily
+#
+# The test are run on:
+#   - Linux / Ubuntu
+#   - Python 3.10
+#
+# Two different kinds of tests are made:
+#   - using the "requirements.txt" file
+#   - using the package dependencies
 #
 # Thomas Guillod - Dartmouth College
 # Mozilla Public License Version 2.0
@@ -25,21 +30,24 @@ jobs:
             cmd: "pip install -r requirements.txt"
           - name: "package"
             cmd: "pip install ."
+        system: ["ubuntu-latest"]
+        version: ["3.10"]
+
     name: Standard / ${{ matrix.method.name }}
-    runs-on: "ubuntu-latest"
-    timeout-minutes: 10
+    runs-on: ${{ matrix.system }}
+    timeout-minutes: 20
     steps:
       - uses: actions/checkout@v4
       - name: Set up Python
         uses: actions/setup-python@v5
         with:
-          python-version: "3.10"
+          python-version: ${{ matrix.version }}
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
           ${{ matrix.method.cmd }}
       - name: Run the tests
-        shell: bash
+        shell: bash -el {0}
         run: |
           chmod +x scripts/run_tests.sh
           ./scripts/run_tests.sh
