+++ .github/workflows/run_conda.yml
@@ -16,6 +16,7 @@
 name: Run Conda Tests
 
 on:
+  push:
   workflow_dispatch:
   schedule:
     - cron: "0 2 * * 1"
@@ -30,10 +31,10 @@ jobs:
             file: "conda_base.yaml"
           - name: "dev"
             file: "conda_dev.yaml"
-        system: ["ubuntu-latest"]
-        version: ["3.10"]
+        system: ["ubuntu-latest",  "macos-latest", "windows-latest"]
+        version: ["3.9", "3.10", "3.11", "3.12"]
 
-    name: Conda / ${{ matrix.conda.name }}
+    name: Conda / ${{ matrix.conda.name }} / ${{ matrix.system }} / ${{ matrix.version }}
     runs-on: ${{ matrix.system }}
     timeout-minutes: 20
     steps:
