+++ .github/workflows/run_standard.yml
@@ -33,7 +33,7 @@ jobs:
         system: ["ubuntu-latest"]
         version: ["3.10"]
 
-    name: Standard / ${{ matrix.method.name }}
+    name: Standard / ${{ matrix.method.name }} / ${{ matrix.system }} / ${{ matrix.version }}
     runs-on: ${{ matrix.system }}
     timeout-minutes: 20
     steps:
