+++ .github/workflows/run_platform.yml
@@ -26,10 +26,15 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
+        method:
+          - name: "requirements"
+            cmd: "pip install -r requirements.txt"
+          - name: "package"
+            cmd: "pip install ."
         system: ["ubuntu-latest",  "macos-latest", "windows-latest"]
         version: ["3.9", "3.10", "3.11", "3.12"]
 
-    name: Platform / ${{ matrix.system }} / ${{ matrix.version }}
+    name: Platform / ${{ matrix.method.name }} / ${{ matrix.system }} / ${{ matrix.version }}
     runs-on: ${{ matrix.system }}
     timeout-minutes: 20
     steps:
@@ -41,7 +46,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install -r requirements.txt
+          ${{ matrix.method.cmd }}
       - name: Run the tests
         shell: bash -el {0}
         run:
