### 2024-04-19 6e90541
better test triggers
