+++ .github/workflows/run_conda.yml
@@ -1,4 +1,6 @@
 # Run the Conda tests with the following conditions.
+#   - push (main)
+#   - pull request
 #   - dispatch
 #   - weekly
 #
@@ -16,9 +18,13 @@
 name: Run Conda Tests
 
 on:
+  pull_request:
   workflow_dispatch:
   schedule:
     - cron: "0 2 * * 1"
+  push:
+    branches:
+      - 'main'
 
 jobs:
   build:
