### 2024-04-21 86393eb
log when exiting the plot code
