### 2024-04-21 otvam/pypeec#5 @otvam
Yes, ["3.9", "3.10", "3.11", "3.12"] are the versions tested during the CI.

The version requirements are now clearly stated in the INSTALL.md and in the documentation.

The problem to create a conda package is that the VTK (C++ library for 3D object manipulation) conda package is currently broken for win64 (#6). I might have found a workaround and will try to create a Conda package.
