### 2024-04-22 otvam/pypeec#5 @thelfer
@otvam Thanks for the precision.
Concerning conda, don't be mistaken. My remark originates from the fact that there was no INSTALL file, so I had to somehow guess how to install it. I saw files related to conda at the root directory and though "maybe there is a conda package ?". For the review, a `conda` package is not a matter, having clear install instructions is.
So don't bother with a `conda` package for the review (but feel free to do it if you find it important for the project)
