### 2024-04-22 2cc9888
add conda links
