+++ docs/content/about.rst
@@ -7,10 +7,7 @@ Links
 ==========  ========================================
 Website     https://pypeec.otvam.ch
 Repository  https://github.com/otvam/pypeec
-Releases    https://github.com/otvam/pypeec/releases
-Tags        https://github.com/otvam/pypeec/tags
 Issues      https://github.com/otvam/pypeec/issues
-PyPi        https://pypi.org/project/pypeec
 ==========  ========================================
 
 Author
@@ -26,14 +23,14 @@ Website      https://otvam.ch
 Release
 -------
 
-=======  ========================================
-Package  **pypeec**
-Version  |ver|
-Date     |date|
-PyPi     https://pypi.org/project/pypeec
-GitHub   https://github.com/otvam/pypeec/releases
-=======  ========================================
-
+==========  ========================================
+Package     **pypeec**
+Version     |ver| / |date|
+PyPi        https://pypi.org/project/pypeec
+Releases    https://github.com/otvam/pypeec/releases
+Conda       https://anaconda.org/conda-forge/pypeec
+GitHub      https://github.com/otvam/pypeec/releases
+==========  ========================================
 Credits
 -------
 
