+++ docs/content/about.rst
@@ -30,6 +30,7 @@ GitHub      https://github.com/otvam/pypeec/releases
 Conda       https://anaconda.org/conda-forge/pypeec
 PyPi        https://pypi.org/project/pypeec
 ==========  ========================================
+
 Credits
 -------
 
