### 2024-04-22 b029631
scripts path bug
