### 2024-04-22 8a6d1eb
better conda url
