+++ INSTALL.md
@@ -13,7 +13,7 @@ The optional HPC libraries are usually easier to install though Conda.
 
 ```bash
 # Install a Python interpreter
-#   - Website: https://www.python.org
+#   - Website: https://www.python.org/downloads
 #   - Supported versions: 3.9, 3.10, 3.11, and 3.12
 #   - Python executable: "python" or "python3"
 
@@ -34,7 +34,7 @@ pypeec --version
 
 ```bash
 # Install a Conda distribution
-#   - Website: https://github.com/conda-forge/miniforge
+#   - Website: https://conda-forge.org/download
 #   - Miniforge is sufficient for installing PyPEEC
 #   - Using mamba as a dependency solver is faster
 
