+++ docs/website/README.md
@@ -2,7 +2,6 @@
 
 > * **Website: [pypeec.otvam.ch](https://pypeec.otvam.ch)**
 > * **Repository: [github.com/otvam/pypeec](https://github.com/otvam/pypeec)**
-> * **Package: [pypi.org/project/pypeec](https://pypi.org/project/pypeec)**
 
 > * This repository contains the PyPEEC website and documentation.
 > * The content is automatically generated from the main repository.
