### 2024-04-23 f28de04
add pot core example
