+++ examples/examples_png/inductor_pot/geometry.yaml
@@ -1,0 +1,89 @@
+# Definition of a "file_geometry" data.
+#
+# This file is used by the mesher.
+# Definition of the voxel structure.
+#
+# Thomas Guillod - Dartmouth College
+# Mozilla Public License Version 2.0
+
+"mesh_type": "png"
+"data_voxelize":
+    "param":
+        "d": [500.0e-6, 500.0e-6, 500.0e-6]
+        "c": [0.0e-3, 0.0e-3, 0.0e-6]
+        "size": [37, 37]
+    "domain_color":
+        "coil": [[0, 0, 0, 255]]
+        "core": [[0, 0, 255, 255]]
+        "src": [[255, 0, 0, 255]]
+        "sink": [[0, 255, 0, 255]]
+    "layer_stack":
+        - {"n_layer": 2, "filename_list": !path ["png/cover.png"]}
+        - {"n_layer": 2, "filename_list": !path ["png/core.png"]}
+        - {"n_layer": 2, "filename_list": !path ["png/coil.png"]}
+        - {"n_layer": 1, "filename_list": !path ["png/gap.png"]}
+        - {"n_layer": 2, "filename_list": !path ["png/coil.png"]}
+        - {"n_layer": 2, "filename_list": !path ["png/core.png"]}
+        - {"n_layer": 2, "filename_list": !path ["png/cover.png"]}
+"resampling":
+    "use_reduce": false
+    "use_resample": false
+    "resampling_factor": [1, 1, 1]
+"pts_cloud":
+    - [-0.01, -0.01, -0.008]
+    - [-0.01, -0.01, 0.008]
+    - [-0.005, -0.01, -0.008]
+    - [-0.005, -0.01, 0.008]
+    - [0.0, -0.01, -0.008]
+    - [0.0, -0.01, 0.008]
+    - [0.005, -0.01, -0.008]
+    - [0.005, -0.01, 0.008]
+    - [0.01, -0.01, -0.008]
+    - [0.01, -0.01, 0.008]
+    - [-0.01, -0.005, -0.008]
+    - [-0.01, -0.005, 0.008]
+    - [-0.005, -0.005, -0.008]
+    - [-0.005, -0.005, 0.008]
+    - [0.0, -0.005, -0.008]
+    - [0.0, -0.005, 0.008]
+    - [0.005, -0.005, -0.008]
+    - [0.005, -0.005, 0.008]
+    - [0.01, -0.005, -0.008]
+    - [0.01, -0.005, 0.008]
+    - [-0.01, 0.0, -0.008]
+    - [-0.01, 0.0, 0.008]
+    - [-0.005, 0.0, -0.008]
+    - [-0.005, 0.0, 0.008]
+    - [0.0, 0.0, -0.008]
+    - [0.0, 0.0, 0.008]
+    - [0.005, 0.0, -0.008]
+    - [0.005, 0.0, 0.008]
+    - [0.01, 0.0, -0.008]
+    - [0.01, 0.0, 0.008]
+    - [-0.01, 0.005, -0.008]
+    - [-0.01, 0.005, 0.008]
+    - [-0.005, 0.005, -0.008]
+    - [-0.005, 0.005, 0.008]
+    - [0.0, 0.005, -0.008]
+    - [0.0, 0.005, 0.008]
+    - [0.005, 0.005, -0.008]
+    - [0.005, 0.005, 0.008]
+    - [0.01, 0.005, -0.008]
+    - [0.01, 0.005, 0.008]
+    - [-0.01, 0.01, -0.008]
+    - [-0.01, 0.01, 0.008]
+    - [-0.005, 0.01, -0.008]
+    - [-0.005, 0.01, 0.008]
+    - [0.0, 0.01, -0.008]
[…]
