+++ examples/examples_voxel/slab/geometry.yaml
@@ -21,22 +21,22 @@
     "use_resample": true
     "resampling_factor": [2, 2, 1]
 "pts_cloud":
-    - [30.0e-3, 0.0e-3, -30.0e-3]
-    - [30.0e-3, 0.0e-3, -10.0e-3]
-    - [30.0e-3, 0.0e-3, 10.0e-3]
-    - [30.0e-3, 0.0e-3, 30.0e-3]
-    - [-30.0e-3, 0.0e-3, -30.0e-3]
-    - [-30.0e-3, 0.0e-3, -10.0e-3]
-    - [-30.0e-3, 0.0e-3, 10.0e-3]
-    - [-30.0e-3, 0.0e-3, 30.0e-3]
-    - [0.0e-3, 30.0e-3, -30.0e-3]
-    - [0.0e-3, 30.0e-3, -10.0e-3]
-    - [0.0e-3, 30.0e-3, 10.0e-3]
-    - [0.0e-3, 30.0e-3, 30.0e-3]
-    - [0.0e-3, -30.0e-3, -30.0e-3]
-    - [0.0e-3, -30.0e-3, -10.0e-3]
-    - [0.0e-3, -30.0e-3, 10.0e-3]
-    - [0.0e-3, -30.0e-3, 30.0e-3]
+    - [+30.00e-3, +00.00e-3, -30.00e-3]
+    - [+30.00e-3, +00.00e-3, -10.00e-3]
+    - [+30.00e-3, +00.00e-3, +10.00e-3]
+    - [+30.00e-3, +00.00e-3, +30.00e-3]
+    - [-30.00e-3, +00.00e-3, -30.00e-3]
+    - [-30.00e-3, +00.00e-3, -10.00e-3]
+    - [-30.00e-3, +00.00e-3, +10.00e-3]
+    - [-30.00e-3, +00.00e-3, +30.00e-3]
+    - [+00.00e-3, +30.00e-3, -30.00e-3]
+    - [+00.00e-3, +30.00e-3, -10.00e-3]
+    - [+00.00e-3, +30.00e-3, +10.00e-3]
+    - [+00.00e-3, +30.00e-3, +30.00e-3]
+    - [+00.00e-3, -30.00e-3, -30.00e-3]
+    - [+00.00e-3, -30.00e-3, -10.00e-3]
+    - [+00.00e-3, -30.00e-3, +10.00e-3]
+    - [+00.00e-3, -30.00e-3, +30.00e-3]
 "check_cloud": true
 "check_conflict": true
 "check_connection": true
