+++ examples/examples_png/inductor_pot/geometry.yaml
@@ -30,56 +30,448 @@
     "use_resample": false
     "resampling_factor": [1, 1, 1]
 "pts_cloud":
-    - [-0.01, -0.01, -0.008]
-    - [-0.01, -0.01, 0.008]
-    - [-0.005, -0.01, -0.008]
-    - [-0.005, -0.01, 0.008]
-    - [0.0, -0.01, -0.008]
-    - [0.0, -0.01, 0.008]
-    - [0.005, -0.01, -0.008]
-    - [0.005, -0.01, 0.008]
-    - [0.01, -0.01, -0.008]
-    - [0.01, -0.01, 0.008]
-    - [-0.01, -0.005, -0.008]
-    - [-0.01, -0.005, 0.008]
-    - [-0.005, -0.005, -0.008]
-    - [-0.005, -0.005, 0.008]
-    - [0.0, -0.005, -0.008]
-    - [0.0, -0.005, 0.008]
-    - [0.005, -0.005, -0.008]
-    - [0.005, -0.005, 0.008]
-    - [0.01, -0.005, -0.008]
-    - [0.01, -0.005, 0.008]
-    - [-0.01, 0.0, -0.008]
-    - [-0.01, 0.0, 0.008]
-    - [-0.005, 0.0, -0.008]
-    - [-0.005, 0.0, 0.008]
-    - [0.0, 0.0, -0.008]
-    - [0.0, 0.0, 0.008]
-    - [0.005, 0.0, -0.008]
-    - [0.005, 0.0, 0.008]
-    - [0.01, 0.0, -0.008]
-    - [0.01, 0.0, 0.008]
-    - [-0.01, 0.005, -0.008]
-    - [-0.01, 0.005, 0.008]
-    - [-0.005, 0.005, -0.008]
-    - [-0.005, 0.005, 0.008]
-    - [0.0, 0.005, -0.008]
-    - [0.0, 0.005, 0.008]
-    - [0.005, 0.005, -0.008]
-    - [0.005, 0.005, 0.008]
-    - [0.01, 0.005, -0.008]
-    - [0.01, 0.005, 0.008]
-    - [-0.01, 0.01, -0.008]
-    - [-0.01, 0.01, 0.008]
-    - [-0.005, 0.01, -0.008]
-    - [-0.005, 0.01, 0.008]
-    - [0.0, 0.01, -0.008]
-    - [0.0, 0.01, 0.008]
-    - [0.005, 0.01, -0.008]
-    - [0.005, 0.01, 0.008]
-    - [0.01, 0.01, -0.008]
-    - [0.01, 0.01, 0.008]
+    - [-11.75e-3, -11.75e-3, -05.75e-3]
+    - [-11.75e-3, -11.75e-3, -03.83e-3]
+    - [-11.75e-3, -11.75e-3, -01.92e-3]
+    - [-11.75e-3, -11.75e-3, +00.00e-3]
+    - [-11.75e-3, -11.75e-3, +01.92e-3]
+    - [-11.75e-3, -11.75e-3, +03.83e-3]
+    - [-11.75e-3, -11.75e-3, +05.75e-3]
+    - [-09.40e-3, -11.75e-3, -05.75e-3]
+    - [-09.40e-3, -11.75e-3, -03.83e-3]
+    - [-09.40e-3, -11.75e-3, -01.92e-3]
+    - [-09.40e-3, -11.75e-3, +00.00e-3]
+    - [-09.40e-3, -11.75e-3, +01.92e-3]
+    - [-09.40e-3, -11.75e-3, +03.83e-3]
[…]
