+++ examples/tutorial/geometry.yaml
@@ -97,15 +97,15 @@
     "use_resample": false
     "resampling_factor": [1, 1, 1]
 "pts_cloud":
-    - [1.0e-3, 1.0e-3, 0.0e-3]
-    - [1.0e-3, 2.0e-3, 0.0e-3]
-    - [1.0e-3, 3.0e-3, 0.0e-3]
-    - [2.0e-3, 1.0e-3, 0.0e-3]
-    - [2.0e-3, 2.0e-3, 0.0e-3]
-    - [2.0e-3, 3.0e-3, 0.0e-3]
-    - [3.0e-3, 1.0e-3, 0.0e-3]
-    - [3.0e-3, 2.0e-3, 0.0e-3]
-    - [3.0e-3, 3.0e-3, 0.0e-3]
+    - [+01.00e-3, +01.00e-3, +00.00e-3]
+    - [+01.00e-3, +02.00e-3, +00.00e-3]
+    - [+01.00e-3, +03.00e-3, +00.00e-3]
+    - [+02.00e-3, +01.00e-3, +00.00e-3]
+    - [+02.00e-3, +02.00e-3, +00.00e-3]
+    - [+02.00e-3, +03.00e-3, +00.00e-3]
+    - [+03.00e-3, +01.00e-3, +00.00e-3]
+    - [+03.00e-3, +02.00e-3, +00.00e-3]
+    - [+03.00e-3, +03.00e-3, +00.00e-3]
 "check_cloud": true
 "check_conflict": true
 "check_connection": true
