+++ examples/examples_voxel/core/geometry.yaml
@@ -25,22 +25,21 @@
     "use_resample": false
     "resampling_factor": [1, 1, 1]
 "pts_cloud":
-    - [50.0e-3, 0.0e-3, -30.0e-3]
-    - [50.0e-3, 0.0e-3, -10.0e-3]
-    - [50.0e-3, 0.0e-3, 10.0e-3]
-    - [50.0e-3, 0.0e-3, 30.0e-3]
-    - [-50.0e-3, 0.0e-3, -30.0e-3]
-    - [-50.0e-3, 0.0e-3, -10.0e-3]
-    - [-50.0e-3, 0.0e-3, 10.0e-3]
-    - [-50.0e-3, 0.0e-3, 30.0e-3]
-    - [0.0e-3, 50.0e-3, -30.0e-3]
-    - [0.0e-3, 50.0e-3, -10.0e-3]
-    - [0.0e-3, 50.0e-3, 10.0e-3]
-    - [0.0e-3, 50.0e-3, 30.0e-3]
-    - [0.0e-3, -50.0e-3, -30.0e-3]
-    - [0.0e-3, -50.0e-3, -10.0e-3]
-    - [0.0e-3, -50.0e-3, 10.0e-3]
-    - [0.0e-3, -50.0e-3, 30.0e-3]
+    - [+50.00e-3, +00.00e-3, -30.00e-3]
+    - [+50.00e-3, +00.00e-3, -10.00e-3]
+    - [+50.00e-3, +00.00e-3, +10.00e-3]
+    - [+50.00e-3, +00.00e-3, +30.00e-3]
+    - [-50.00e-3, +00.00e-3, -30.00e-3]
+    - [-50.00e-3, +00.00e-3, -10.00e-3]
+    - [-50.00e-3, +00.00e-3, +10.00e-3]
+    - [-50.00e-3, +00.00e-3, +30.00e-3]
+    - [+00.00e-3, +50.00e-3, -30.00e-3]
+    - [+00.00e-3, +50.00e-3, -10.00e-3]
+    - [+00.00e-3, +50.00e-3, +10.00e-3]
+    - [+00.00e-3, +50.00e-3, +30.00e-3]
+    - [+00.00e-3, -50.00e-3, -30.00e-3]
+    - [+00.00e-3, -50.00e-3, -10.00e-3]
+    - [+00.00e-3, -50.00e-3, +10.00e-3]
 "check_cloud": true
 "check_conflict": true
 "check_connection": true

[…]
