+++ examples/config/tolerance.yaml
@@ -88,8 +88,8 @@ sweep_pool: null
     # criteria for interrupting the solver if the complex power has converged
     "power_options":
         "stop": true                   # use (or not) stop criteria on the complex power
-        "n_min": 3                     # minimum number of iterations of the iterative solver
-        "n_cmp": 2                     # number of iterations to consider for the convergence
+        "n_min": 4                     # minimum number of iterations of the iterative solver
+        "n_cmp": 3                     # number of iterations to consider for the convergence
         "rel_tol": 1.0e-4              # relative tolerance for the complex power
         "abs_tol": 1.0e-10             # absolute tolerance for the complex power
 
