### 2024-04-24 0f98a06
add pot core examples to the doc
