### 2024-04-24 9c9804e
equation variable documentation
