+++ pypeec/lib_solver/equation_system.py
@@ -63,7 +63,7 @@ The matrices have the following units (for the AC system):
     - R_m => 1/Henry
     - P_m => 1/Henry
 
-For the DC problem (zero frequency), division per zero are occurring.
+For the DC problem (zero frequency), multiplication per zero are occurring.
 Therefore, the problem is formulated slightly differently for this case.
 
 For the preconditioner, the diagonal of the inductance and potential matrix is used.
@@ -120,34 +120,6 @@ def _get_system_size(A_net_c, A_net_m, A_src):
     return n_vc, n_fc, n_vm, n_fm, n_src
 
 
-def _get_system_scaler(freq, n_vc, n_fc, n_vm, n_fm, n_src):
-    """
-    Get a scaling vector for the solution.
-
-    The scaling vector is required for the following reasons:
-        - for avoid division per zero, the DC and AC system are slightly different
-        - the DC system is using the magnetic flux as a variable
-        - the AC system is using the derivative of the magnetic flux as a variable
-        - the solution is using the magnetic flux as a variable
-        - the scaling vector is used to match this discrepancy
-    """
-
-    # get the angular frequency
-    s = 1j*2*np.pi*freq
-
-    # get the scaling vectors
-    if freq == 0:
-        scaler_c = np.ones(n_fc+n_vc+n_src, dtype=np.complex_)
-        scaler_m = np.ones(n_fm+n_vm, dtype=np.complex_)
-    else:
-        scaler_c = np.ones(n_fc+n_vc+n_src, dtype=np.complex_)
-        scaler_fm = s*np.ones(n_fm, dtype=np.complex_)
-        scaler_vm = np.ones(n_vm, dtype=np.complex_)
-        scaler_m = np.concatenate((scaler_fm, scaler_vm))
-
-    return scaler_c, scaler_m
-
-
 def _get_coupling_electric(sol_m, freq, n_vc, n_fc, n_fm, n_src, K_op_c):
     """
     Compute the magnetic to electric couplings.
@@ -160,11 +132,14 @@ def _get_coupling_electric(sol_m, freq, n_vc, n_fc, n_fm, n_src, K_op_c):
     # extract the face current
     I_fm = sol_m[0:n_fm]
 
+    # get the angular frequency
+    s = 1j*2*np.pi*freq
+
     # compute the couplings
     if freq == 0:
         cpl_fc = np.zeros(n_fc, dtype=np.complex_)
     else:
-        cpl_fc = K_op_c(I_fm)
+        cpl_fc = s*K_op_c(I_fm)
 
     cpl_vc = np.zeros(n_vc, dtype=np.complex_)
     cpl_src = np.zeros(n_src, dtype=np.complex_)
@@ -245,7 +220,7 @@ def _get_cond_fact_electric(freq, A_net_c, R_c, L_c, A_src):
     return Y_mat, S_mat, A_12_mat, A_21_mat
 
 
-def _get_cond_fact_magnetic(freq, A_net_m, R_m, P_m):
+def _get_cond_fact_magnetic(A_net_m, R_m, P_m):
     """
     Compute the sparse matrices using for the magnetic preconditioner.
 
@@ -258,16 +233,9 @@ def _get_cond_fact_magnetic(freq, A_net_m, R_m, P_m):
     # get the system size
     (n_vm, n_fm) = A_net_m.shape
 
-    # get the angular frequency
-    s = 1j*2*np.pi*freq
-
     # get the magnetic admittance (avoid singularity for DC solution)
-    if freq == 0:
[…]
