### 2024-04-25 987b6b6
equation_system.py documentation
