+++ pypeec/lib_solver/equation_system.py
@@ -9,32 +9,32 @@ The problem contains n_src_c current source voxels and n_src_v voltage source vo
 
 The equations are set in the following order:
     [
-        n_fc    : the electric KVL equations
-        n_vc    : the electric KCL equations
-        n_src_c : the current source voxels equations (source equation with internal admittance)
-        n_src_v : the voltage source voxels equations (source equation with internal impedance)
-        n_fm    : the magnetic KVL equations
-        n_vm    : the magnetic KCL equations
+        n_fc        the electric KVL equations
+        n_vc        the electric KCL equations
+        n_src_c     the current source voxels equations (source equation with internal admittance)
+        n_src_v     the voltage source voxels equations (source equation with internal impedance)
+        n_fm        the magnetic KVL equations
+        n_vm        the magnetic KCL equations
     ]
 
 The complete solution vector is:
     [
-        I_fc    : n_fc    : electric face currents    : A
-        V_vc    : n_vc    : electric voxel potentials : V
-        I_src_c : n_src_c : current source currents   : A
-        I_src_v : n_src_v : voltage source currents   : A
-        I_fm    : n_fm    : magnetic face fluxes      : V*s
-        V_vm    : n_vm    : magnetic voxel potentials : A
+        I_fc        n_fc        electric face currents        A
+        V_vc        n_vc        electric voxel potentials     V
+        I_src_c     n_src_c     current source currents       A
+        I_src_v     n_src_v     voltage source currents       A
+        I_fm        n_fm        magnetic face fluxes          V*s
+        V_vm        n_vm        magnetic voxel potentials     A
     ]
 
 The complete right-hand size vector is:
     [
-        n_fc    : zero excitation            : V
-        n_vc    : zero excitation            : A
-        n_src_c : current source excitations : A
-        n_src_v : voltage source excitations : V
-        n_fm    : zero excitation            : A
-        n_vm    : zero excitation            : A
+        n_fc        zero excitation                V
+        n_vc        zero excitation                A
+        n_src_c     current source excitations     A
+        n_src_v     voltage source excitations     V
+        n_fm        zero excitation                A
+        n_vm        zero excitation                A
     ]
 
 The complete equation matrix is:
@@ -57,11 +57,23 @@ The units of the equation matrix is:
         0              0               0                  0                  A/V/s            1
     ]
 
-The matrices have the following units:
-    - R_c => Ohm
-    - L_c => Henry
-    - R_m => 1/Henry = A/V/s
-    - P_m => 1/Henry = A/V/s
+The matrices have the following description and units:
+    [
+        R_c               resistance matrix                             Ohm
+        L_c               inductance matrix                             Henry
+        R_m               magnetic resistance matrix                    1/Henry
+        P_m               magnetic potential matrix                     1/Henry
+        A_net_c           electric voxel incidence matrix               1
+        A_net_m           magnetic voxel incidence matrix               1
+        K_c               magnetic to electric coupling matrix          1
+        K_m               electric to magnetic coupling matrix          1
+        A_vc_src_c        current source current coupling               1
+        A_vc_src_v        voltage source current coupling               1
+        A_src_c_vc        admittance matrix for the current sources     1/Ohm
+        A_src_v_vc        voltage source voltage coupling               1
[…]
