+++ examples/run_solver.py
@@ -20,8 +20,8 @@ FOLDER_EXAMPLE = examples_config.FOLDER_EXAMPLE
 if __name__ == "__main__":
     # get the filenames
     file_problem = os.path.join(PATH_ROOT, FOLDER_EXAMPLE, "problem.yaml")
-    file_voxel = os.path.join(PATH_ROOT, FOLDER_EXAMPLE, "voxel.gz")
-    file_solution = os.path.join(PATH_ROOT, FOLDER_EXAMPLE, "solution.gz")
+    file_voxel = os.path.join(PATH_ROOT, FOLDER_EXAMPLE, "voxel.json.gz")
+    file_solution = os.path.join(PATH_ROOT, FOLDER_EXAMPLE, "solution.json.gz")
     file_tolerance = os.path.join(PATH_ROOT, FOLDER_CONFIG, "tolerance.yaml")
 
     # run
