+++ docs/content/tutorial.rst
@@ -33,10 +33,10 @@ Mesher: Building and Meshing the Geometry
 
     # Run the mesher
     #   - geometry.yaml - contains the geometry description (input)
-    #   - voxel.gz - contains the meshed voxel structure (output)
+    #   - voxel.json.gz - contains the meshed voxel structure (output)
     pypeec mesher \
         --geometry tutorial/geometry.yaml \
-        --voxel tutorial/voxel.gz
+        --voxel tutorial/voxel.json.gz
 
 .. literalinclude:: ../tutorial/log_mesher.txt
    :language: text
@@ -47,11 +47,11 @@ Viewer: Plot the Geometry and Mesh
 .. code-block:: bash
 
     # Run the viewer
-    #   - voxel.gz - contains the meshed voxel structure (input)
+    #   - voxel.json.gz - contains the meshed voxel structure (input)
     #   - viewer.yaml - contains the plot configuration (input)
     #   - tag_plot - list of plots to be shown (defined in viewer.yaml)
     pypeec viewer \
-        --voxel tutorial/voxel.gz \
+        --voxel tutorial/voxel.json.gz \
         --viewer config/viewer.yaml \
         --tag_plot domain connection
 
@@ -69,15 +69,15 @@ Solver: Solve the PEEC Problem
 .. code-block:: bash
 
     # Run the solver
-    #   - voxel.gz - contains the meshed voxel structure (input)
+    #   - voxel.json.gz - contains the meshed voxel structure (input)
     #   - problem.yaml - contains the magnetic problem description (input)
     #   - tolerance.yaml - contains the solver numerical tolerances (input)
-    #   - solution.gz - contains the problem solution (output)
+    #   - solution.json.gz - contains the problem solution (output)
     pypeec solver \
-        --voxel tutorial/voxel.gz \
+        --voxel tutorial/voxel.json.gz \
         --problem tutorial/problem.yaml \
         --tolerance config/tolerance.yaml \
-        --solution tutorial/solution.gz
+        --solution tutorial/solution.json.gz
 
 .. literalinclude:: ../tutorial/log_solver.txt
    :language: text
@@ -88,11 +88,11 @@ Plotter: Plot the Solution
 .. code-block:: bash
 
     # Run the plotter
-    #   - solution.gz - contains the problem solution (input)
+    #   - solution.json.gz - contains the problem solution (input)
     #   - plotter.yaml - contains the plot configuration (input)
     #   - tag_plot - list of plots to be shown (defined in plotter.yaml)
     pypeec plotter \
-        --solution tutorial/solution.gz \
+        --solution tutorial/solution.json.gz \
         --plotter config/plotter.yaml \
         --tag_plot V_c_abs J_c_norm H_norm residuum
 
