### 2024-05-11 0f97cd4
prevent nagware-like logo display
