+++ tests/test_png.py
@@ -25,9 +25,6 @@ name_list = [
     "examples_png/gerber",
 ]
 
-# show the logo
-main.run_hide_logo()
-
 # add the tests
 for name in name_list:
     test_workflow.set_test(TestPng, name, name, False)
