+++ tests/test_stl.py
@@ -22,9 +22,6 @@ name_list = [
     "examples_stl/transformer",
 ]
 
-# show the logo
-main.run_hide_logo()
-
 # add the tests
 for name in name_list:
     test_workflow.set_test(TestStl, name, name, False)
