### 2024-05-16 abfc186
refactor logging with unique name
