+++ pypeec/main.py
@@ -18,7 +18,7 @@ from pypeec import log
 from pypeec import io
 
 # create the logger
-LOGGER = log.get_logger("MAIN")
+LOGGER = log.get_logger("pypeec")
 
 
 def run_display_logo():
