+++ pypeec/lib_mesher/mesher_stl.py
@@ -15,7 +15,7 @@ import pyvista as pv
 from pypeec import log
 
 # get a logger
-LOGGER = log.get_logger("STL")
+LOGGER = log.get_logger("pypeec")
 
 # prevent VTK to mess up the output
 vtk.vtkObject.GlobalWarningDisplayOff()
