+++ pypeec/run/viewer.py
@@ -27,7 +27,7 @@ from pypeec.lib_check import check_data_options
 from pypeec import log
 
 # get a logger
-LOGGER = log.get_logger("VIEWER")
+LOGGER = log.get_logger("pypeec")
 
 
 def _get_grid_voxel(data_geom):
