+++ pypeec/lib_solver/extract_solution.py
@@ -22,7 +22,7 @@ import numpy.linalg as lna
 from pypeec import log
 
 # get a logger
-LOGGER = log.get_logger("SOLUTION")
+LOGGER = log.get_logger("pypeec")
 
 
 def _get_biot_savart(pts, pts_net, J_src, vol):
