### 2024-05-16 b6986d4
recursive enabling/disabling
