### 2024-05-16 a849bac
logger with full name (and short names)
