+++ pypeec/lib_mesher/mesher_voxel.py
@@ -12,7 +12,7 @@ import numpy as np
 from pypeec import log
 
 # get a logger
-LOGGER = log.get_logger("pypeec")
+LOGGER = log.get_logger(__name__, "pypeec")
 
 
 def get_mesh(param, domain_index):
