+++ pypeec/lib_matrix/multiply_direct.py
@@ -10,7 +10,7 @@ import numpy as np
 from pypeec import log
 
 # get a logger
-LOGGER = log.get_logger("pypeec")
+LOGGER = log.get_logger(__name__, "pypeec")
 
 
 def _get_voxel_indices(nx, ny, nz):
