+++ pypeec/lib_solver/problem_value.py
@@ -12,7 +12,7 @@ import numpy as np
 from pypeec import log
 
 # get a logger
-LOGGER = log.get_logger("pypeec")
+LOGGER = log.get_logger(__name__, "pypeec")
 
 
 def get_material_vector(material_val, material_idx):
