+++ pypeec/run/plotter.py
@@ -32,7 +32,7 @@ from pypeec.lib_check import check_data_options
 from pypeec import log
 
 # get a logger
-LOGGER = log.get_logger("pypeec")
+LOGGER = log.get_logger(__name__, "pypeec")
 
 
 def _get_grid_voxel(data_init, data_sweep):
