### 2024-05-16 66c60e8
finer control of the logging level
