### 2024-05-16 4c16b7d
logging format documentation
