### 2024-05-18 44ead63
option to hide the logo (problem with non-unicode terminals)
