+++ pypeec/main.py
@@ -26,13 +26,17 @@ def run_display_logo():
     Display the logo as a splash screen.
     """
 
+    # load logo data
+    with importlib.resources.open_text("pypeec.data", "pypeec.txt") as file_logo:
+        data = file_logo.read()
+
+    # display logo
     try:
-        with importlib.resources.open_text("pypeec.data", "pypeec.txt") as file_logo:
-            data = file_logo.read()
-            print("", flush=True, file=sys.stderr)
-            print(data, flush=True, file=sys.stderr)
-            print("", flush=True, file=sys.stderr)
-    except UnicodeError:
+        data.encode(sys.stderr.encoding)
+        print("", flush=True, file=sys.stderr)
+        print(data, flush=True, file=sys.stderr)
+        print("", flush=True, file=sys.stderr)
+    except UnicodeEncodeError:
         pass
 
 
