### 2024-05-18 8eb05ab
quiet mode for tests
