### 2024-05-18 0541716
ad/dc simulations for skin and proximity effect examples
