+++ examples/examples_shape/busbar/problem.yaml
@@ -22,9 +22,17 @@
         "source_type": "voltage"
         "var_type": "lumped"
 "sweep_config":
-    "sim_default": null
+    "sim_dc": null
+    "sim_ac": null
 "sweep_param":
-    "sim_default":
+    "sim_dc":
+        "freq": 0.0
+        "material_val":
+            "coil": {"rho_re": 1.75e-8, "rho_im": 0.0}
+        "source_val":
+            "src": {"V_re": +1.0, "V_im": 0.0, "Z_re": 1.0e-3, "Z_im": 0.0}
+            "sink": {"V_re": -1.0, "V_im": 0.0, "Z_re": 1.0e-3, "Z_im": 0.0}
+    "sim_ac":
         "freq": 1.0e+6
         "material_val":
             "coil": {"rho_re": 1.75e-8, "rho_im": 0.0}
