+++ examples/examples_shape/busbar/geometry.yaml
@@ -30,8 +30,8 @@
                     "buffer": null
                     "coord_shell": [[0.0e-3, 0.0e-3], [2.5e-3, 0.0e-3], [2.5e-3, 2.5e-3], [0.0e-3, 2.5e-3]]
                     "coord_holes":
-                        - [[0.2e-3, 0.2e-3], [0.4e-3, 0.2e-3], [0.4e-3, 0.4e-3], [0.2e-3, 0.4e-3]]
-                        - [[2.1e-3, 2.1e-3], [2.3e-3, 2.1e-3], [2.3e-3, 2.3e-3], [2.1e-3, 2.3e-3]]
+                        - [[0.1e-3, 0.1e-3], [0.2e-3, 0.1e-3], [0.2e-3, 0.2e-3], [0.1e-3, 0.2e-3]]
+                        - [[0.3e-3, 0.3e-3], [0.4e-3, 0.3e-3], [0.4e-3, 0.4e-3], [0.3e-3, 0.4e-3]]
             -
                 "shape_layer": ["front"]
                 "shape_operation": "add"
