+++ .github/workflows/run_conda.yml
@@ -1,4 +1,4 @@
-# Run the Conda tests with the following conditions.
+# Run the "Conda tests" with the following conditions.
 #   - dispatch
 #   - weekly
 #
