+++ .github/workflows/run_standard.yml
@@ -1,4 +1,4 @@
-# Run the Standard tests with the following conditions.
+# Run the "Standard tests" with the following conditions.
 #   - push (all)
 #   - pull request
 #   - dispatch
@@ -6,10 +6,6 @@
 # The test are run on:
 #   - Linux / Ubuntu
 #   - Python 3.10
-#
-# Two different kinds of tests are made:
-#   - using the "requirements.txt" file
-#   - using the package dependencies
 #
 # Thomas Guillod - Dartmouth College
 # Mozilla Public License Version 2.0
@@ -22,31 +18,19 @@ on:
   workflow_dispatch:
 
 jobs:
-  build:
-    strategy:
-      fail-fast: false
-      matrix:
-        method:
-          - name: "requirements"
-            cmd: "pip install -r requirements.txt"
-          - name: "package"
-            cmd: "pip install ."
-        system: ["ubuntu-latest"]
-        version: ["3.10"]
-
-    name: Standard / ${{ matrix.method.name }} / ${{ matrix.system }} / ${{ matrix.version }}
-    runs-on: ${{ matrix.system }}
+    name: Standard
+    runs-on: "ubuntu-latest"
     timeout-minutes: 20
     steps:
       - uses: actions/checkout@v4
       - name: Set up Python
         uses: actions/setup-python@v5
         with:
-          python-version: ${{ matrix.version }}
+          python-version: "3.10"
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          ${{ matrix.method.cmd }}
+          pip install -r requirements.txt
       - name: Run the tests
         shell: bash -el {0}
         run: |
