+++ .github/workflows/run_standard.yml
@@ -19,7 +19,6 @@ on:
 
 jobs:
   build:
-    name: Standard
     runs-on: "ubuntu-latest"
     timeout-minutes: 20
     steps:
