+++ docs/format/file_geometry_shape.yaml
@@ -8,7 +8,8 @@
     #   - dy: voxel dimension for the y direction
     #   - dz: voxel dimension for the z direction
     #   - cz: coordinates of the voxel structure center for the z direction
-    #   - tol: tolerance for simplifying the shapes (null for disabling simplification)
+    #   - simplify: tolerance for simplifying the shapes (null for disabling simplification)
+    #   - construct: grid size for constructing the shapes (null for default grid size)
     #   - xy_min: array with the lower corner coordinates of the voxel structure (x and y directions)
     #   - xy_max: array with the upper corner coordinates of the voxel structure (x and y directions)
     #   - alternatively, the xy_min/xy_max can be set to null and the shape bounds are kept
@@ -17,7 +18,8 @@
         "dy": 50.0e-6
         "dz": 50.0e-6
         "cz": 0.0e-6
-        "tol": 1.0e-6
+        "simplify": 1.0e-6
+        "construct": 1.0e-6
         "xy_min": [-20e-3, -20e-3]
         "xy_max": [+20e-3, +20e-3]
 
