+++ pypeec/lib_check/check_data_geometry_shape.py
@@ -137,7 +137,7 @@ def _check_param(param):
     datachecker.check_float("dx", param["dx"], is_positive=True, can_be_zero=False)
     datachecker.check_float("dy", param["dy"], is_positive=True, can_be_zero=False)
     datachecker.check_float("dz", param["dz"], is_positive=True, can_be_zero=False)
-    datachecker.check_float("cz", param["cz"], is_positive=False, can_be_zero=True)
+    datachecker.check_float("cz", param["cz"], is_positive=False, can_be_zero=True, can_be_none=True)
     datachecker.check_float("simplify", param["simplify"], is_positive=True, can_be_zero=False, can_be_none=True)
     datachecker.check_float("construct", param["construct"], is_positive=True, can_be_zero=False, can_be_none=True)
     datachecker.check_float_array("xy_min", param["xy_min"], size=2, is_positive=False, can_be_zero=True, can_be_none=True)
