### 2024-05-25 #6644 @otvam
Hi,

I just wanted to ask if there is any progress for the review. The issue about the installation of PyPEEC has been solved (see https://github.com/otvam/pypeec/issues/5).
