### 2024-05-25 otvam/pypeec#5 @otvam
The conda package is now available through conda-forge!
It was overdue because the HPC libraries used by PyPEEC are much easier to install with conda.
