### 2024-05-28 d8dfe01
remove log color support (non portable) / proper exception logging
