+++ pypeec/log.py
@@ -112,8 +112,6 @@ def _check_config(data):
         "FORMAT",
         "INDENTATION",
         "EXCEPTION_TRACE",
-        "USE_COLOR",
-        "DEF_COLOR",
         "LEVEL_DEF",
         "LEVEL_NAME",
     ]
@@ -123,7 +121,6 @@ def _check_config(data):
     _check_string("LEVEL_DEF", data["LEVEL_DEF"])
     _check_integer("INDENTATION", data["INDENTATION"])
     _check_boolean("EXCEPTION_TRACE", data["EXCEPTION_TRACE"])
-    _check_boolean("USE_COLOR", data["USE_COLOR"])
 
     # check sub dict
     key_list = [
@@ -140,41 +137,11 @@ def _check_config(data):
     _check_integer("TIMESTAMP_TRC", data["FORMAT"]["TIMESTAMP_TRC"])
     _check_integer("DURATION_TRC", data["FORMAT"]["DURATION_TRC"])
 
-    # check sub dict
-    key_list = [
-        "CL_DEBUG",
-        "CL_INFO",
-        "CL_WARNING",
-        "CL_ERROR",
-        "CL_CRITICAL",
-        "CL_RESET",
-    ]
-    _check_dict("DEF_COLOR", data["DEF_COLOR"], key_list)
-    _check_string("CL_DEBUG", data["DEF_COLOR"]["CL_DEBUG"])
-    _check_string("CL_INFO", data["DEF_COLOR"]["CL_INFO"])
-    _check_string("CL_WARNING", data["DEF_COLOR"]["CL_WARNING"])
-    _check_string("CL_ERROR", data["DEF_COLOR"]["CL_ERROR"])
-    _check_string("CL_CRITICAL", data["DEF_COLOR"]["CL_CRITICAL"])
-    _check_string("CL_RESET", data["DEF_COLOR"]["CL_RESET"])
-
     # check the logging levels
     _check_dict("LEVEL_NAME", data["LEVEL_NAME"], [])
     for name, level in data["LEVEL_NAME"].items():
         _check_string("LEVEL_NAME", name)
         _check_string("LEVEL_NAME", level)
-
-
-def _get_fmt(color, reset):
-    """
-    Get a logging formatter.
-    """
-
-    if (color is None) and (reset is None):
-        fmt = logging.Formatter(FORMAT["LOGGER"])
-    else:
-        fmt = logging.Formatter("\x1b" + color + FORMAT["LOGGER"] + "\x1b" + reset)
-
-    return fmt
 
 
 def _get_format_timestamp(timestamp):
@@ -211,43 +178,40 @@ class _DeltaTimeFormatter(logging.Formatter):
     Class for adding elapsed time to a logger.
     """
 
-    def __init__(self, name, tag):
+    def __init__(self, tag):
         """
         Constructor.
         Create a timer.
         """
 
         # call parent constructor
-        super().__init__()
+        super().__init__(fmt=FORMAT["LOGGER"])
 
         # assign
-        self.name = name
-        if tag is None:
-            self.tag = name
-        else:
-            self.tag = tag
+        self.tag = tag
 
-        # define the color formatters
-        self.fmt_color = {
-            logging.DEBUG: _get_fmt(DEF_COLOR["CL_DEBUG"], DEF_COLOR["CL_RESET"]),
-            logging.INFO: _get_fmt(DEF_COLOR["CL_INFO"], DEF_COLOR["CL_RESET"]),
-            logging.WARNING: _get_fmt(DEF_COLOR["CL_WARNING"], DEF_COLOR["CL_RESET"]),
-            logging.ERROR: _get_fmt(DEF_COLOR["CL_ERROR"], DEF_COLOR["CL_RESET"]),
[…]
