+++ pypeec/data/logger.yaml
@@ -23,18 +23,6 @@ INDENTATION: 4 # integer with the number of spaces per indentation
 # display (or) not exception traces
 EXCEPTION_TRACE: true # boolean switch
 
-# options for coloring the log output
-USE_COLOR: false # use (or not) colors for the logs
-
-# options color definition
-DEF_COLOR:
-    CL_DEBUG: '[38;5;247m' # debug level color code
-    CL_INFO: '[38;5;15m' # info level color code
-    CL_WARNING: '[38;5;202m' # warning level color code
-    CL_ERROR: '[38;5;196m' # error level color code
-    CL_CRITICAL: '[38;5;196m' # critical level color code
-    CL_RESET: '[0m' # terminal reset color sequence
-
 # default level for all the handlers
 LEVEL_DEF: "DEBUG"
 
