### 2024-05-28 3081a34
log formatter factorization
