### 2024-05-28 db73528
reintroduce color for the log
