### 2024-05-28 a017995
switch logger to ini and remove YAML dependency (not relevant for PyPEEC)
