+++ pypeec/log.py
@@ -6,7 +6,7 @@ Module for handling the logging.
     - Log exceptions.
 
 The log config is defined by the following files.
-    - First the default configuration is loaded ("pypeec/data/logger.yaml").
+    - First the default configuration is loaded ("pypeec/data/logger.ini").
     - Afterward, a custom file can be loaded with an environment variable ("PYTHONLOGGER").
 
 Warning
@@ -25,7 +25,7 @@ import sys
 import datetime
 import threading
 import logging
-import yaml
+import configparser
 import importlib.resources
 
 # global timestamp (constant over the complete run)
@@ -35,14 +35,25 @@ GLOBAL_TIMESTAMP = datetime.datetime.today()
 GLOBAL_LEVEL = 0
 
 
+def _decode_escape(value):
+    """
+    Decode the escape sequence in a string.
+    """
+
+    value = bytes(value, "utf-8")
+    value = value.decode("unicode_escape")
+
+    return value
+
+
 def _get_level(name):
     """
     Get the corresponding logging level.
     """
 
     # check for exact match
-    if name in LEVEL_NAME:
-        return LEVEL_NAME[name]
+    if name in MODULE_LEVEL:
+        return MODULE_LEVEL[name]
 
     # check for parent level
     split = name.rsplit(".", 1)
@@ -87,7 +98,7 @@ def _check_string(name, data):
         raise AssertionError("%s: cannot be empty" % name)
 
 
-def _check_dict(name, data, key_list):
+def _check_dict(name, data):
     """
     Check a dict.
     """
@@ -97,71 +108,38 @@ def _check_dict(name, data, key_list):
         raise AssertionError("%s: should be a dict" % name)
 
     # check keys
-    for tag in key_list:
-        if tag not in data:
-            raise AssertionError("%s: dict is incomplete: %s" % (name, tag))
+    for tag, value in data.items():
+        _check_string(name, tag)
+        _check_string(name, value)
 
 
-def _check_config(data):
+def _check_config():
     """
     Check the integrity of the config file.
     """
 
-    # check dict
-    key_list = [
-        "FORMAT",
-        "TIMING",
-        "INDENTATION",
-        "EXCEPTION_TRACE",
-        "USE_COLOR",
-        "LEVEL_COLOR",
-        "RESET_COLOR",
-        "LEVEL_DEFAULT",
-        "LEVEL_NAME",
-    ]
-    _check_dict("data", data, key_list)
+    # check format data
+    _check_string("FORMAT", FORMAT)
+    _check_string("LEVEL_DEFAULT", LEVEL_DEFAULT)
+    _check_boolean("EXCEPTION_TRACE", EXCEPTION_TRACE)
+    _check_integer("INDENTATION", INDENTATION)
 
-    # check data
-    _check_string("FORMAT", data["FORMAT"])
-    _check_string("LEVEL_DEFAULT", data["LEVEL_DEFAULT"])
-    _check_integer("INDENTATION", data["INDENTATION"])
-    _check_boolean("EXCEPTION_TRACE", data["EXCEPTION_TRACE"])
-    _check_string("RESET_COLOR", data["RESET_COLOR"])
-    _check_boolean("USE_COLOR", data["USE_COLOR"])
+    # check color data
+    _check_string("COLOR_RESET", COLOR_RESET)
+    _check_string("COLOR_DEFAULT", COLOR_DEFAULT)
+    _check_boolean("USE_COLOR", COLOR_USE)
 
-    # check sub dict
-    key_list = [
-        "DEBUG",
-        "INFO",
[…]
