+++ pypeec/data/logger.yaml
@@ -1,50 +1,0 @@
-# PyPEEC default logger configuration file.
-#
-# Thomas Guillod - Dartmouth College
-# Mozilla Public License Version 2.0
-
-# logging module logger format with the following custom attributes
-#   - name: full name of the logger (formatted as a string)
-#   - tag: short name of the logger (formatted as a string)
-#   - timestamp: current timestamp (formatted as a string)
-#   - duration: elapsed time (formatted as a string)
-#   - process_id: process identification number
-#   - thread_id: thread identification number
-FORMAT: "%(duration)s : %(tag)-8s : %(levelname)-8s : %(message)s" # logger format
-
-
-# format instruction for the timestamps and durations
-TIMING:
-    TIMESTAMP_FMT: "%m/%d/%Y : %H:%M:%S" # strftime formatting instructions for the timestamp
-    DURATION_FMT: "%H:%M:%S.%f" # strftime formatting instructions for the duration
-    TIMESTAMP_TRC: 0 # truncate trailing characters for the timestamp string
-    DURATION_TRC: 4 # truncate trailing characters for the duration string
-
-# indentation level for the blocks
-INDENTATION: 4 # integer with the number of spaces per indentation
-
-# options for coloring the log output
-USE_COLOR: true # use (or not) colors for the logs
-
-# color definition for the different levels
-LEVEL_COLOR:
-    DEBUG: "\x1b[38;5;247m" # debug level color code
-    INFO: "\x1b[38;5;15m" # info level color code
-    WARNING: "\x1b[38;5;202m" # warning level color code
-    ERROR: "\x1b[38;5;196m" # error level color code
-    CRITICAL: "\x1b[38;5;196m" # critical level color code
-
-# terminal reset color sequence
-RESET_COLOR: "\x1b[0m"
-
-# display (or) not full exception traces
-EXCEPTION_TRACE: true
-
-# default level for all the handlers
-LEVEL_DEFAULT: "DEBUG"
-
-# custom level for specific handlers
-LEVEL_NAME:
-    "pypeec": "DEBUG" # level for the pypeec package
-    "pypeec.main": "DEBUG" # level for pypeec.main module
-    "pypeec.script": "DEBUG" # level for pypeec.script module
