+++ pypeec/data/logger.ini
@@ -1,0 +1,82 @@
+# PyPEEC default logger configuration file.
+#
+# Thomas Guillod - Dartmouth College
+# Mozilla Public License Version 2.0
+
+######################################################################################################
+# global options
+######################################################################################################
+[GLOBAL]
+
+# logging module logger format with the following custom attributes
+#   - name: full name of the logger (formatted as a string)
+#   - tag: short name of the logger (formatted as a string)
+#   - timestamp: current timestamp (formatted as a string)
+#   - duration: elapsed time (formatted as a string)
+#   - process_id: process identification number
+#   - thread_id: thread identification number
+FORMAT = %(duration)s : %(tag)-8s : %(levelname)-8s : %(message)s
+
+# integer with the number of spaces per indentation
+INDENTATION = 4
+
+# default level for all the handlers
+LEVEL_DEFAULT = DEBUG
+
+# display (or) not full exception traces
+EXCEPTION_TRACE = true
+
+# use (or not) colors for the logs
+COLOR_USE = false
+
+# terminal reset color sequence
+COLOR_RESET = \x1b[0m
+
+# default terminal color
+COLOR_DEFAULT = \x1b[38;5;15m
+
+# strftime formatting instructions for the timestamp
+TIMESTAMP_FMT = %m/%d/%Y|%H:%M:%S
+
+# strftime formatting instructions for the duration
+DURATION_FMT = %H:%M:%S.%f
+
+# truncate trailing characters for the timestamp string
+TIMESTAMP_TRC = 0
+
+# truncate trailing characters for the duration string
+DURATION_TRC = 4
+
+######################################################################################################
+# color definition for the different levels
+######################################################################################################
+[COLOR_LEVEL]
+
+# debug level color code
+DEBUG = \x1b[38;5;247m
+
+# info level color code
+INFO = \x1b[38;5;15m
+
+# warning level color code
+WARNING = \x1b[38;5;202m
+
+# error level color code
+ERROR = \x1b[38;5;196m
+
+# critical level color code
+CRITICAL = \x1b[38;5;196m
+
+######################################################################################################
+# custom level for specific handlers
+######################################################################################################
+[MODULE_LEVEL]
+
+# level for the pypeec package
+pypeec = DEBUG
+
+#level for pypeec.main module
+pypeec.main = DEBUG
+
+# level for pypeec.script module
+pypeec.script = DEBUG
