+++ pypeec/data/logger.ini
@@ -1,4 +1,5 @@
 # PyPEEC default logger configuration file.
+# This file is written in for the Python ConfigParser.
 #
 # Thomas Guillod - Dartmouth College
 # Mozilla Public License Version 2.0
@@ -26,15 +27,6 @@ LEVEL_DEFAULT = DEBUG
 # display (or) not full exception traces
 EXCEPTION_TRACE = true
 
-# use (or not) colors for the logs
-COLOR_USE = false
-
-# terminal reset color sequence
-COLOR_RESET = \x1b[0m
-
-# default terminal color
-COLOR_DEFAULT = \x1b[38;5;15m
-
 # strftime formatting instructions for the timestamp
 TIMESTAMP_FMT = %m/%d/%Y|%H:%M:%S
 
@@ -47,36 +39,33 @@ TIMESTAMP_TRC = 0
 # truncate trailing characters for the duration string
 DURATION_TRC = 4
 
+# use (or not) colors for the logs
+#   - if true: ANSI escape characters are used for setting the color
+#   - if false: ANSI escape characters are not used
+COLOR_USE = false
+
+# terminal ANSI escape characters for resetting color
+COLOR_RESET = \x1b[0m
+
+# terminal ANSI escape characters for the default color
+COLOR_DEFAULT = \x1b[38;5;15m
+
 ######################################################################################################
-# color definition for the different levels
+# color definition (ANSI escape characters) for the different levels
 ######################################################################################################
 [COLOR_LEVEL]
 
-# debug level color code
 DEBUG = \x1b[38;5;247m
-
-# info level color code
 INFO = \x1b[38;5;15m
-
-# warning level color code
 WARNING = \x1b[38;5;202m
-
-# error level color code
 ERROR = \x1b[38;5;196m
-
-# critical level color code
 CRITICAL = \x1b[38;5;196m
 
 ######################################################################################################
-# custom level for specific handlers
+# custom level for specific handler names
 ######################################################################################################
 [MODULE_LEVEL]
 
-# level for the pypeec package
 pypeec = DEBUG
-
-#level for pypeec.main module
 pypeec.main = DEBUG
-
-# level for pypeec.script module
 pypeec.script = DEBUG
