+++ pypeec/data/logger.ini
@@ -1,5 +1,5 @@
 # PyPEEC default logger configuration file.
-# This file is written in for the Python ConfigParser.
+# This file is written for the Python ConfigParser.
 #
 # Thomas Guillod - Dartmouth College
 # Mozilla Public License Version 2.0
