+++ pypeec/data/logger.ini
@@ -1,21 +1,22 @@
 # PyPEEC default logger configuration file.
-# This file is written for the Python ConfigParser.
+# This file is written as a Python ConfigParser INI file.
 #
 # Thomas Guillod - Dartmouth College
 # Mozilla Public License Version 2.0
 
 ######################################################################################################
-# global options
+# Global logger options
 ######################################################################################################
 [GLOBAL]
+######################################################################################################
 
 # logging module logger format with the following custom attributes
-#   - name: full name of the logger (formatted as a string)
-#   - tag: short name of the logger (formatted as a string)
-#   - timestamp: current timestamp (formatted as a string)
-#   - duration: elapsed time (formatted as a string)
-#   - process_id: process identification number
-#   - thread_id: thread identification number
+#   - name: full name of the logger emitting the message (string)
+#   - tag: non-unique tag assigned to the logger (string)
+#   - timestamp: timestamp with the current time (string)
+#   - duration: elapsed time since log creation (string)
+#   - process_id: process identification number (integer)
+#   - thread_id: thread identification number (integer)
 FORMAT = %(duration)s : %(tag)-8s : %(levelname)-8s : %(message)s
 
 # integer with the number of spaces per indentation
@@ -27,10 +28,10 @@ LEVEL_DEFAULT = DEBUG
 # display (or) not full exception traces
 EXCEPTION_TRACE = true
 
-# strftime formatting instructions for the timestamp
+# formatting instructions for the timestamp (datetime/strftime format)
 TIMESTAMP_FMT = %m/%d/%Y|%H:%M:%S
 
-# strftime formatting instructions for the duration
+# formatting instructions for the duration (datetime/strftime format)
 DURATION_FMT = %H:%M:%S.%f
 
 # truncate trailing characters for the timestamp string
@@ -51,9 +52,10 @@ COLOR_RESET = \x1b[0m
 COLOR_DEFAULT = \x1b[38;5;15m
 
 ######################################################################################################
-# color definition (ANSI escape characters) for the different levels
+# Color definition for the different levels (ANSI escape characters)
 ######################################################################################################
 [COLOR_LEVEL]
+######################################################################################################
 
 DEBUG = \x1b[38;5;247m
 INFO = \x1b[38;5;15m
@@ -62,9 +64,10 @@ ERROR = \x1b[38;5;196m
 CRITICAL = \x1b[38;5;196m
 
 ######################################################################################################
-# custom level for specific handler names
+# Custom logging level for specific handler names
 ######################################################################################################
 [MODULE_LEVEL]
+######################################################################################################
 
 pypeec = DEBUG
 pypeec.main = DEBUG
