### 2024-05-28 58ff6bc
bug in default log config (shadowing)
