+++ pypeec/data/logger.ini
@@ -57,6 +57,9 @@ COLOR_DEFAULT = \x1b[38;5;15m
 [COLOR_LEVEL]
 ######################################################################################################
 
+# keys-values dictionary
+#   - the keys are the logging levels
+#   - the values are the color codes
 DEBUG = \x1b[38;5;247m
 INFO = \x1b[38;5;15m
 WARNING = \x1b[38;5;202m
@@ -69,6 +72,8 @@ CRITICAL = \x1b[38;5;196m
 [MODULE_LEVEL]
 ######################################################################################################
 
-pypeec = DEBUG
-pypeec.main = DEBUG
-pypeec.script = DEBUG
+# keys-values dictionary
+#   - the keys are the module names
+#   - the values are the logging levels
+pypeec.log = CRITICAL
+pypeec.io = CRITICAL
