### 2024-05-30 2c979c2
keep the exception context
