### 2024-05-30 8aa2d9e
remove exception context for reraise
