### 2024-06-01 a6955ed
log module (not pypeec specific)
