+++ pypeec/data/logger.ini
@@ -75,5 +75,5 @@ CRITICAL = \x1b[38;5;196m
 # keys-values dictionary
 #   - the keys are the module names
 #   - the values are the logging levels
-pypeec.log = CRITICAL
-pypeec.io = CRITICAL
+pypeec.module_error = ERROR
+pypeec.module_info = INFO
