### 2024-06-01 9d4ca7f
better block logging
