+++ pypeec/run/viewer.py
@@ -117,7 +117,7 @@ def run(
     (grid, voxel, point, reference) = _get_grid_voxel(data_geom)
 
     # make the plots
-    with log.BlockTimer(LOGGER, "generate plots"):
+    with log.BlockIndent(LOGGER, "generate plots"):
         for i, (tag_plot, data_viewer_tmp) in enumerate(data_viewer.items()):
             LOGGER.info("plotting %d / %d / %s" % (i+1, len(data_viewer), tag_plot))
             _get_plot(tag_plot, data_viewer_tmp, grid, voxel, point, reference, gui_obj)
