### 2024-06-01 6a0a2b7
better exception logging
