+++ pypeec/run/viewer.py
@@ -119,7 +119,7 @@ def run(
     # make the plots
     with log.BlockIndent(LOGGER, "generate plots"):
         for i, (tag_plot, data_viewer_tmp) in enumerate(data_viewer.items()):
-            LOGGER.info("plotting %d / %d / %s" % (i+1, len(data_viewer), tag_plot))
+            LOGGER.info("plot / %d / %d / %s" % (i+1, len(data_viewer), tag_plot))
             _get_plot(tag_plot, data_viewer_tmp, grid, voxel, point, reference, gui_obj)
 
     # enter the event loop (should be at the end, blocking call)
