### 2024-06-02 c942782
duration bug if greater the 24h
