+++ pypeec/data/logger.ini
@@ -31,14 +31,12 @@ EXCEPTION_TRACE = true
 # formatting instructions for the timestamp (datetime/strftime format)
 TIMESTAMP_FMT = %m/%d/%Y|%H:%M:%S
 
-# formatting instructions for the duration (datetime/strftime format)
-DURATION_FMT = %H:%M:%S.%f
-
-# truncate trailing characters for the timestamp string
-TIMESTAMP_TRC = 0
-
-# truncate trailing characters for the duration string
-DURATION_TRC = 4
+# formatting instructions for the duration (string format)
+#   - days: number of elapsed days (integer)
+#   - hours: number of elapsed hours (integer)
+#   - minutes: number of elapsed minutes (integer)
+#   - seconds: number of elapsed seconds (float)
+DURATION_FMT = {hours:02d}:{minutes:02d}:{seconds:05.2f}
 
 # use (or not) colors for the logs
 #   - if true: ANSI escape characters are used for setting the color
