+++ pypeec/data/logger.ini
@@ -25,8 +25,11 @@ INDENTATION = 4
 # default level for all the handlers
 LEVEL_DEFAULT = DEBUG
 
-# display (or) not full exception traces
+# display (or not) full exception traceback
 EXCEPTION_TRACE = true
+
+# use (or not) UTC times for the timestamp
+TIMESTAMP_UTC = false
 
 # formatting instructions for the timestamp (datetime/strftime format)
 TIMESTAMP_FMT = %m/%d/%Y|%H:%M:%S
