### 2024-06-06 otvam/pypeec#7 @thelfer
[JOSS review] Various questions
@otvam First of all, my sincere apologies for the review process which is much slower than what it should. My professional duties did not allow me to make it better.

I opened this issue mostly for discussion around the review.

Here are some questions associated with the review criteria of JOSS:

> Whether the software has already been cited in academic papers.

I did not find any by search `PyPEEC` in google scholar. Do you have any information on that matter ?

> About authorship

Would you precise the role of Charles R. Sullivan in the development of `PyPEEC` or the writing of the paper. According to the insights of the project, you seem to only contributor. Maybe a section describing the contributions of the authors would help in the paper (this seems to have become a common practice in most papers)

> Community guideline (https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines)

I did not section any `CONTRIBUTING.md` file in the sources, nor any dedicated sections in the documentation. Did I missed it ?
If you don't have one, feel free to get inspiration from this one for instance: https://github.com/thelfer/MFrontGenericInterfaceSupport/blob/master/CONTRIBUTING.md

> Are there automated tests or manual steps described so that the functionality of the software can be verified?

Would you comment on this ?
