### 2024-06-06 #6644 @thelfer
Issues related to the review:

- https://github.com/otvam/pypeec/issues/5 (installation, closed)
- https://github.com/otvam/pypeec/issues/7
