### 2024-06-06 24ce2eb
author list and warning for MKL and CUDA
