### 2024-06-07 otvam/pypeec#7 @otvam
No problem for the delay!

> Whether the software has already been cited in academic papers.

Not yet. PyPEEC has only been released last September so it is still a bit early. PyPEEC is currently used in two U.S. National Science Foundation sponsored projects and these projects are expected to produce papers (probably next year). Additionally, I have already received emails from three universities and a company asking questions about PyPEEC usage, indicating that the tool is actually used.

> About authorship

For PyPEEC itself, I've added the list of contributions in https://github.com/otvam/pypeec/blob/main/docs/content/about.rst. The contributions of Charles Sullivan are not into the programming itself but into the tool architecture and the "which/what/how" for the electromagnetic solver.

For the paper, the contributions are as following:
   - Myself: "Statement of Need", "Capabilities and Workflow", and "Solver Performance" (doing the benchmark)
   - Charles Sullivan: "Statement of Need" and "Solver Performance" (defining the benchmark)

I could include that in the paper. I am not sure what is the JOSS policy about that.

> Community guideline

The contributing file is located there: https://github.com/otvam/pypeec/blob/main/.github/CONTRIBUTING.md. I do agree that this location is strange but this is the GitHub preferred location. The same information are also in the documentation: https://github.com/otvam/pypeec/blob/main/docs/content/technical.rst

> Are there automated tests or manual steps described so that the functionality of the software can be verified?

PyPEEC provides 19 different examples and a tutorial (all located in the `examples` folder). These examples are covering all the relevant functionalities of PyPEEC (different frequencies, materials, sources, distributed parameters, an-isotropic parameters, geometry input formats, etc.).

The automated tests are running all the examples and checking for eventual regressions (integration and end-to-end tests). The tests are running locally and with GitHub actions on different platforms (Linux, MacOS, Windows) and with different installation methods (requirements.txt, pyproject.toml, Conda).
