### 2024-06-10 otvam/pypeec#7 @thelfer
> I could include that in the paper. I am not sure what is the JOSS policy about that.

We may ask the editor about this :) @likeajumprope, is there any guidance about adding a description of the respective contributions of the authors. This seems to have became common practice in other journals.
