### 2024-06-10 3268a15
documentation for the numerical options and the tests
