+++ examples/config/cfg_data_matplotlib.yaml
@@ -6,4 +6,3 @@
 "style": "default"         # string with a Matplotlib style
 "legend": "best"           # location of the legend box
 "font": 12                 # reference font size of the figure
-
