### 2024-06-10 86d778b
remove is_truncated (complex and not useful) / factorize the field solution
