### 2024-06-10 8542904
plotter for arbitrary field variables
