### 2024-06-10 58415fb
simplify variable datastruct
