### 2024-06-10 fd68b31
simplify variable datastruct
