### 2024-06-10 37a6291
better checks for field variable existence
