### 2024-06-11 a4c47d9
rename abs to norm
