### 2024-06-11 d62f75b
phasor plot for scalars
