+++ pypeec/lib_visualization/manage_pyvista.py
@@ -129,39 +129,10 @@ def _get_clip_mesh(pl, obj, arg, plot_clip):
         )
 
 
-def _get_phase_vector(obj, var, phase):
+def _get_scale_norm(obj, scale):
     """
-    Get a vector for a specified phase shift.
-    The vector is constructed from the phasor.
-    """
-
-    # get var
-    data_re = obj[var + "_re"]
-    data_im = obj[var + "_im"]
-
-    # construct the vector
-    time = data_re+1j*data_im
-
-    # phase shift the vector
-    time *= np.exp(1j*phase)
-
-    # get the vector in the time domain
-    time = np.real(time)
-
-    # get the norm
-    norm = lna.norm(time, axis=1)
-
-    # assign the data
-    obj[var + "_time"] = time
-    obj[var + "_norm"] = norm
-
-    return obj
-
-
-def _get_filter_vector(obj, var, arrow_threshold):
-    """
-    Filter the voxel structure with a vector field.
-    This function is used to remove arrows with extremely low lengths.
+    Scale a variable between a lower and upper bound.
+    The operation is done with respect to the norm variable.
     """
 
     # if the voxel structure is empty, nothing to do
@@ -169,7 +140,30 @@ def _get_filter_vector(obj, var, arrow_threshold):
         return obj
 
     # get var
-    data = obj[var]
+    data = obj["norm"]
+
+    # add scaling
+    data = scale*data
+
+    # assign data
+    obj["norm"] = data
+
+    return obj
+
+
+def _get_filter_arrow(obj, arrow_threshold):
+    """
+    Filter the voxel structure with a variable.
+    This function is used to remove arrows with extremely low lengths.
+    The operation is done with respect to the norm variable.
+    """
+
+    # if the voxel structure is empty, nothing to do
+    if obj.n_cells == 0:
+        return obj
+
+    # get var
+    data = obj["norm"]
 
     # threshold for arrow removal
     if np.any(np.isfinite(data)):
@@ -186,30 +180,10 @@ def _get_filter_vector(obj, var, arrow_threshold):
     return obj
 
 
-def _get_scale_scalar(obj, var, scale):
+def _get_filter_norm(obj, filter_lim):
     """
-    Scale a scalar variable between a lower and upper bound.
-    """
-
-    # if the voxel structure is empty, nothing to do
-    if obj.n_cells == 0:
-        return obj
-
-    # get var
-    data = obj[var]
-
-    # add scaling
-    data = scale*data
-
-    # assign data
-    obj[var] = data
-
-    return obj
-
-
-def _get_filter_scalar(obj, var, filter_lim):
-    """
-    Filter the voxel structure with provided limits with respect to a scalar variable.
+    Filter the voxel structure with provided limits.
+    The operation is done with respect to the norm variable.
     """
 
     # if the voxel structure is empty, nothing to do
@@ -224,7 +198,7 @@ def _get_filter_scalar(obj, var, filter_lim):
         (f_min, f_max) = filter_lim
 
     # get var
-    data = obj[var]
+    data = obj["norm"]
 
     # indices to be kept
     idx = np.logical_and(data >= f_min, data <= f_max)
@@ -235,9 +209,10 @@ def _get_filter_scalar(obj, var, filter_lim):
     return obj
 
 
-def _get_clamp_scalar(obj, var, clamp_lim):
[…]
