+++ examples/config/plotter.yaml
@@ -17,8 +17,8 @@
 #       - color_current_source: "color of the current source domains"
 #       - color_voltage_source: color of the voltage source domains
 #
-#   - "scalar_voxel": plot a scalar variable on the voxel structure
-#   - "scalar_point": plot a scalar variable on the point cloud
+#   - "norm_voxel": plot a scalar variable on the voxel structure
+#   - "norm_point": plot a scalar variable on the point cloud
 #       - title: title located at the corner of the plot
 #       - legend: legend of variable being shown with the color axis
 #       - var: name of the scalar variable to be plotted (color scale)
@@ -71,62 +71,31 @@
         "color_voltage_source": "royalblue"
 "V_c_norm":
     "framework": "pyvista"
-    "layout": "scalar_voxel"
+    "layout": "norm_voxel"
     "title": "El. Potential"
     "data_window": !include cfg_data_window.yaml
     "data_options": !include cfg_data_pyvista.yaml
     "data_plot":
         "title": "El. Potential"
         "legend": "El. Potential [V]"
-        "var": "V_c_norm"
+        "var": "V_c"
         "scale": 1.0
         "log": false
         "color_lim": null
         "clamp_lim": null
         "filter_lim": null
         "point_size": 10.0
-"V_m_norm":
+"V_c_phasor":
     "framework": "pyvista"
-    "layout": "scalar_voxel"
-    "title": "Mag. Potential"
+    "layout": "phasor_voxel"
+    "title": "El. Potential"
     "data_window": !include cfg_data_window.yaml
     "data_options": !include cfg_data_pyvista.yaml
     "data_plot":
-        "title": "Mag. Potential"
-        "legend": "Mag. Potential [A]"
-        "var": "V_m_norm"
-        "scale": 1.0
-        "log": false
-        "color_lim": null
-        "clamp_lim": null
-        "filter_lim": null
-        "point_size": 10.0
-"S_c_norm":
-    "framework": "pyvista"
-    "layout": "scalar_voxel"
-    "title": "El. Source"
-    "data_window": !include cfg_data_window.yaml
-    "data_options": !include cfg_data_pyvista.yaml
-    "data_plot":
-        "title": "El. Source"
-        "legend": "El. Source [A/mm3]"
-        "var": "S_c_norm"
-        "scale": 1.0e-9
-        "log": false
-        "color_lim": null
-        "clamp_lim": null
-        "filter_lim": null
-        "point_size": 10.0
-"Q_m_norm":
-    "framework": "pyvista"
-    "layout": "scalar_voxel"
-    "title": "Mag. Source"
-    "data_window": !include cfg_data_window.yaml
-    "data_options": !include cfg_data_pyvista.yaml
-    "data_plot":
-        "title": "Mag. Source"
-        "legend": "Mag. Source [mT/mm]"
-        "var": "Q_m_norm"
+        "title": "El. Potential"
+        "legend": "El. Potential [V]"
+        "var": "V_c"
+        "phase": 0.0
         "scale": 1.0
         "log": false
         "color_lim": null
@@ -135,14 +104,14 @@
         "point_size": 10.0
 "P_c_norm":
     "framework": "pyvista"
-    "layout": "scalar_voxel"
+    "layout": "norm_voxel"
     "title": "El. Losses"
     "data_window": !include cfg_data_window.yaml
     "data_options": !include cfg_data_pyvista.yaml
     "data_plot":
         "title": "El. Losses"
         "legend": "El. Losses [W/cm3]"
-        "var": "P_c_norm"
+        "var": "P_c"
         "scale": 1.0e-6
         "log": false
[…]
