### 2024-06-11 c973bc7
plotter.yaml documentation
