+++ examples/config/plotter.yaml
@@ -17,25 +17,38 @@
 #       - color_current_source: "color of the current source domains"
 #       - color_voltage_source: color of the voltage source domains
 #
-#   - "norm_voxel": plot a scalar variable on the voxel structure
-#   - "norm_point": plot a scalar variable on the point cloud
+#   - "norm_voxel": plot a scalar norm variable on the voxel structure
+#   - "norm_point": plot a scalar norm variable on the point cloud
 #       - title: title located at the corner of the plot
 #       - legend: legend of variable being shown with the color axis
-#       - var: name of the scalar variable to be plotted (color scale)
-#       - scale: scaling of the variable (scaling done just before plotting)
+#       - var: name of the scalar norm variable to be plotted
+#       - scale: scaling of the variable (scaling is done first)
 #       - log: use (or not) a log scale for the color axis
 #       - color_lim: limit range for the color axis (null for complete range)
 #       - clamp_lim: clamping range for the color axis (null for complete range)
 #       - filter_lim: hide voxels/points with values outside this range (null for complete range)
 #       - point_size: size of the marker used for plotting on the point cloud
 #
-#   - "arrow_voxel": plot a vector variable on the voxel structure (arrow plot)
-#   - "arrow_voxel": plot a vector variable on the point cloud (arrow plot)
+#   - "phasor_voxel": plot a scalar phasor variable on the voxel structure
+#   - "phasor_point": plot a scalar phasor variable on the point cloud
 #       - title: title located at the corner of the plot
 #       - legend: legend of variable being shown with the color axis
-#       - var_scalar: name of the scalar variable to be plotted (arrow color)
-#       - var_vector: name of the vector variable to be plotted (arrow direction)
-#       - scale: scaling of the scalar variable (scaling done just before plotting)
+#       - var: name of the scalar phasor variable to be plotted
+#       - phase: phase to be considered for the phasor (in radian)
+#       - scale: scaling of the variable (scaling is done first)
+#       - log: use (or not) a log scale for the color axis
+#       - color_lim: limit range for the color axis (null for complete range)
+#       - clamp_lim: clamping range for the color axis (null for complete range)
+#       - filter_lim: hide voxels/points with values outside this range (null for complete range)
+#       - point_size: size of the marker used for plotting on the point cloud
+#
+#   - "arrow_voxel": plot a vector phasor variable on the voxel structure (arrow plot)
+#   - "arrow_voxel": plot a vector phasor variable on the point cloud (arrow plot)
+#       - title: title located at the corner of the plot
+#       - legend: legend of variable being shown with the color axis
+#       - var: name of the vector phasor variable to be plotted
+#       - phase: phase to be considered for the phasor (in radian)
+#       - scale: scaling of the variable (scaling is done first)
 #       - log: use (or not) a log scale for the color axis
 #       - color_lim: limit range for the color axis (null for complete range)
 #       - clamp_lim: clamping range for the color axis (null for complete range)
@@ -46,7 +59,8 @@
 #   - "convergence": plot the convergence of the iterative matrix solver
 #       - color_active: color of the plot for the active power
 #       - color_reactive: color of the plot for the reactive power
-#       - marker: marker shape
+#       - marker: plot marker size
+#       - width: line line width
 #
 #   - "convergence": plot the solver residuum (histogram)
 #       - n_bins: number of bins
@@ -166,7 +180,7 @@
         "color_lim": null
         "clamp_lim": null
         "filter_lim": null
-        "arrow_threshold": 1.0e-3
+        "arrow_threshold": 0.01
[…]
