### 2024-06-11 56a8fb9
make copy of input before solving (useful for notebooks)
