+++ tests/test_stl.py
@@ -7,7 +7,6 @@ __copyright__ = "Thomas Guillod - Dartmouth College"
 __license__ = "Mozilla Public License Version 2.0"
 
 from tests.code import test_workflow
-from pypeec import main
 
 
 # duplicate of the test class
