### 2024-06-12 otvam/pypeec#7 @otvam
To run the tests and coverage analysis locally, the shell scripts "scripts/run_tests.sh" and "scripts/run_coverage.sh" are used. These scripts are also used for the CI with GitHub actions.

The commit 3268a15a6d352d9328e0d95efd694365de826c78 adds all these information to the documentation.
