### 2024-06-12 b92ab31
pardiso and no cond check
