### 2024-06-12 17c3cb1
dirty global var to profile FFT
