### 2024-06-12 bf3848d
comments of the performances
