### 2024-06-13 #6644 @thelfer
@likeajumprope I mostly finished my review. Sorry again for the very long delay, I am more than one month late....

I must remind you that  I do not have any experience of this scientific field, so my review was mostly on the project itself (installation guide, development rules) and the paper. Another reviewer would surely help.

@otvam kindly answered to all my request in the issues previously mentioned.

There is one last remaining question about the description of the contributions of the authors of the paper to the project, As the editor, I let you give us guidance about this.

For the rest, I think that I have done my part (up to what my limited knowledge of the field allows):

- the paper is IHMO good for publication,
- the project now seems to match JOSS criteria, in particular concerning " contributors friendliness"
