### 2024-07-03 0a97215
pin packages in requirements.txt
