+++ requirements.txt
@@ -13,18 +13,18 @@
 PyYAML~=6.0.1
 
 # numerical libraries
-numpy~=1.24.1
-scipy~=1.9.3
-joblib~=1.3.0
+numpy~=1.26.4
+scipy~=1.13.1
+joblib~=1.4.2
 
 # geometry manipulation
-vtk~=9.2.6
-pyvista~=0.40.0
-shapely~=2.0.1
-rasterio~=1.3.8
-Pillow~=10.0.0
+vtk~=9.3.1
+pyvista~=0.43.10
+shapely~=2.0.4
+rasterio~=1.3.10
+Pillow~=10.4.0
 
 # plotting libraries
-PyQt5~=5.15.7
-pyvistaqt~=0.11.0
-matplotlib~=3.7.1
+PyQt5~=5.15.10
+pyvistaqt~=0.11.1
+matplotlib~=3.9.0
