+++ .github/workflows/run_platform.yml
@@ -53,6 +53,6 @@ jobs:
           ${{ matrix.method.cmd }}
       - name: Run the tests
         shell: bash -el {0}
-        run:
+        run: |
           chmod +x scripts/run_tests.sh
           ./scripts/run_tests.sh
