### 2024-07-03 275b399
limit scipy and numpy versions
