### 2024-07-03 39f8183
limit scipy and numpy versions
