+++ pyproject.toml
@@ -52,8 +52,8 @@ classifiers = [
 dependencies = [
     "PyYAML >= 6.0",
     "joblib >= 1.3",
-    "numpy >= 1.24 < 1.27",
-    "scipy >= 1.9 < 1.14",
+    "numpy >= 1.24, < 1.27",
+    "scipy >= 1.9, < 1.14",
     "matplotlib >= 3.7",
     "shapely >= 2.0",
     "rasterio >= 1.3",
