### 2024-07-03 2968b5d
change shell parameters
