+++ .github/workflows/run_platform.yml
@@ -52,7 +52,7 @@ jobs:
           python -m pip install --upgrade pip
           ${{ matrix.method.cmd }}
       - name: Run the tests
-        shell: bash -el {0}
+        shell: bash
         run: |
           chmod +x scripts/run_tests.sh
           ./scripts/run_tests.sh
