+++ .github/workflows/run_standard.yml
@@ -13,6 +13,7 @@
 name: Run Standard Tests
 
 on:
+
   push:
   pull_request:
   workflow_dispatch:
