+++ .github/workflows/run_platform.yml
@@ -31,13 +31,12 @@ jobs:
       fail-fast: false
       matrix:
         method:
-          - name: "requirements"
+          - name: "req"
             cmd: "pip install -r requirements.txt"
-          - name: "package"
+          - name: "pkg"
             cmd: "pip install ."
         system: ["ubuntu-latest",  "macos-latest", "windows-latest"]
         version: ["3.9", "3.10", "3.11", "3.12"]
-
     name: ${{ matrix.method.name }} / ${{ matrix.system }} / ${{ matrix.version }}
     runs-on: ${{ matrix.system }}
     timeout-minutes: 20
