### 2024-07-03 3e06b2f
typo in install
