+++ .github/workflows/run_pypi.yml
@@ -18,7 +18,7 @@
 # Thomas Guillod - Dartmouth College
 # Mozilla Public License Version 2.0
 
-name: Run Platform Tests
+name: Run PyPi Tests
 
 on:
   workflow_dispatch:
@@ -26,25 +26,35 @@ on:
     - cron: "0 2 * * 1"
 
 jobs:
-  platform:
+  build:
     strategy:
       fail-fast: false
       matrix:
-        system: ["ubuntu-latest",  "macos-latest", "windows-latest"]
-        version: ["3.9", "3.10", "3.11", "3.12"]
-    name: ${{ matrix.system }} / ${{ matrix.version }}
-    runs-on: ${{ matrix.system }}
+        method:
+          - name: "requirements"
+            cmd:  |
+              python -m pip install --upgrade pip
+              pip install -r requirements.txt
+          - name: "local"
+            cmd:  |
+              python -m pip install --upgrade pip
+              pip install .
+          - name: "remote"
+            cmd:  |
+              python -m pip install --upgrade pip
+              pip install pypeec
+              python -m pip uninstall --yes pypeec
+    name: ${{ matrix.method.name }}
+    runs-on: "ubuntu-latest"
     timeout-minutes: 20
     steps:
       - uses: actions/checkout@v4
       - name: Set up Python
         uses: actions/setup-python@v5
         with:
-          python-version: ${{ matrix.version }}
+          python-version: "3.10"
       - name: Install dependencies
-        run: |
-          python -m pip install --upgrade pip
-          pip install -r requirements.txt
+        run: ${{ matrix.method.cmd }}
       - name: Run the tests
         shell: bash
         run: |
