+++ .github/workflows/run_standard.yml
@@ -19,7 +19,7 @@ on:
   workflow_dispatch:
 
 jobs:
-  build:
+  standard:
     runs-on: "ubuntu-latest"
     timeout-minutes: 20
     steps:
