+++ .github/workflows/run_pypi.yml
@@ -21,9 +21,9 @@
 name: Run PyPi Tests
 
 on:
+  push:
+  pull_request:
   workflow_dispatch:
-  schedule:
-    - cron: "0 2 * * 1"
 
 jobs:
   build:
