+++ .github/workflows/run_standard.yml
@@ -13,7 +13,6 @@
 name: Run Standard Tests
 
 on:
-
   push:
   pull_request:
   workflow_dispatch:
