+++ .github/workflows/run_platform.yml
@@ -26,7 +26,7 @@ on:
     - cron: "0 2 * * 1"
 
 jobs:
-  platform:
+  default:
     strategy:
       fail-fast: false
       matrix:
@@ -42,6 +42,7 @@ jobs:
         with:
           python-version: ${{ matrix.version }}
       - name: Install dependencies
+        shell: bash
         run: |
           python -m pip install --upgrade pip
           pip install -r requirements.txt
