+++ .github/workflows/run_standard.yml
@@ -18,7 +18,7 @@ on:
   workflow_dispatch:
 
 jobs:
-  standard:
+  default:
     runs-on: "ubuntu-latest"
     timeout-minutes: 20
     steps:
@@ -28,6 +28,7 @@ jobs:
         with:
           python-version: "3.10"
       - name: Install dependencies
+        shell: bash
         run: |
           python -m pip install --upgrade pip
           pip install -r requirements.txt
