### 2024-07-04 a5be9a5
simplify install and actions
