+++ .github/workflows/run_platform.yml
@@ -21,6 +21,7 @@
 name: Run Platform Tests
 
 on:
+  push:
   workflow_dispatch:
   schedule:
     - cron: "0 2 * * 1"
