+++ .github/workflows/run_platform.yml
@@ -43,12 +43,12 @@ jobs:
         with:
           python-version: ${{ matrix.version }}
       - name: Install dependencies
-        shell: bash
+        shell: bash -el {0}
         run: |
           python -m pip install --upgrade pip
           pip install -r requirements.txt
       - name: Run the tests
-        shell: bash
+        shell: bash -el {0}
         run: |
           chmod +x scripts/run_tests.sh
           ./scripts/run_tests.sh
