+++ .github/workflows/run_standard.yml
@@ -28,12 +28,12 @@ jobs:
         with:
           python-version: "3.10"
       - name: Install dependencies
-        shell: bash
+        shell: bash -el {0}
         run: |
           python -m pip install --upgrade pip
           pip install -r requirements.txt
       - name: Run the tests
-        shell: bash
+        shell: bash -el {0}
         run: |
           chmod +x scripts/run_tests.sh
           ./scripts/run_tests.sh
