+++ .github/workflows/run_pypi.yml
@@ -21,7 +21,6 @@
 name: Run PyPi Tests
 
 on:
-  push:
   workflow_dispatch:
   schedule:
     - cron: "0 2 * * 1"
@@ -50,10 +49,10 @@ jobs:
         with:
           python-version: "3.10"
       - name: Install dependencies
-        shell: bash
+        shell: bash -el {0}
         run: ${{ matrix.method.cmd }}
       - name: Run the tests
-        shell: bash
+        shell: bash -el {0}
         run: |
           chmod +x scripts/run_tests.sh
           ./scripts/run_tests.sh
