+++ .github/workflows/run_repository.yml
@@ -13,28 +13,19 @@
 # Thomas Guillod - Dartmouth College
 # Mozilla Public License Version 2.0
 
-name: Run Conda Tests
+name: Run Repository Tests
 
 on:
+  push:
   workflow_dispatch:
   schedule:
     - cron: "0 2 * * 1"
 
 jobs:
   default:
-    strategy:
-      fail-fast: false
-      matrix:
-        method:
-          - name: "conda_base"
-            file:  "conda_base.yaml"
-          - name: "conda_dev"
-            file:  "conda_dev.yaml"
-    name: ${{ matrix.method.name }}
     runs-on: "ubuntu-latest"
     timeout-minutes: 20
     steps:
-      - uses: actions/checkout@v4
       - name: Set up Conda
         uses: conda-incubator/setup-miniconda@v3
         with:
@@ -42,10 +33,15 @@ jobs:
           python-version: "3.10"
           auto-activate-base: false
           activate-environment: pypeec
-          environment-file: ${{ matrix.method.file }}
           use-mamba: true
+      - name: Install dependencies
+        shell: bash -el {0}
+        run: |
+          conda install --yes conda-forge::pypeec
+          pypeec examples examples
+          pypeec documentation documentation
       - name: Run the tests
         shell: bash -el {0}
         run: |
-          chmod +x scripts/run_tests.sh
-          ./scripts/run_tests.sh
+          pypeec --version
+          ls -lsa
