### 2024-07-04 7dec44e
run tutorial with conda and pypi
