+++ .github/workflows/run_standard.yml
@@ -30,8 +30,8 @@ jobs:
       - name: Install dependencies
         shell: bash -el {0}
         run: |
-          python -m pip install --upgrade pip
-          pip install -r requirements.txt
+          pip install --upgrade pip
+          pip install --requirement requirements.txt
       - name: Run the tests
         shell: bash -el {0}
         run: |
