+++ .github/workflows/run_repository.yml
@@ -34,10 +34,19 @@ env:
         --problem tutorial/problem.yaml \
         --tolerance config/tolerance.yaml \
         --solution tutorial/solution.json.gz
-
+  viewer: |
+    pypeec viewer \
+        --voxel tutorial/voxel.json.gz \
+        --viewer config/viewer.yaml \
+        --plot_mode none
+  plotter: |
+    pypeec plotter \
+        --solution tutorial/solution.json.gz \
+        --plotter config/plotter.yaml \
+        --plot_mode none
 
 jobs:
-  default:
+  conda:
     runs-on: "ubuntu-latest"
     timeout-minutes: 20
     steps:
@@ -53,9 +62,31 @@ jobs:
         shell: bash -el {0}
         run: |
           conda install --yes conda-forge::pypeec
+          conda info
       - name: Run the tests
         shell: bash -el {0}
         run: |
           ${{ env.extract }}
           ${{ env.mesher }}
-          ${{ env.solver }}
+          ${{ env.viewer }}
+          ${{ env.plotter }}
+  pypi:
+    runs-on: "ubuntu-latest"
+    timeout-minutes: 20
+    steps:
+      - name: Set up Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: "3.10"
+      - name: Install dependencies
+        shell: bash -el {0}
+        run: |
+          pip install --upgrade pip
+          pip install pypeec
+      - name: Run the tests
+        shell: bash -el {0}
+        run: |
+          ${{ env.extract }}
+          ${{ env.mesher }}
+          ${{ env.viewer }}
+          ${{ env.plotter }}
