+++ .github/workflows/run_pypi.yml
@@ -1,19 +1,10 @@
-# Run the "Platform tests" with the following conditions.
+# Run the "Pypi tests" with the following conditions.
 #   - dispatch
 #   - weekly
 #
-# The test are run on:
-#   - Linux / Ubuntu
-#   - Apple / macOS
-#   - Microsoft / Windows
-#   - Python 3.9
-#   - Python 3.10
-#   - Python 3.11
-#   - Python 3.12
-#
 # Two different kinds of tests are made:
 #   - using the "requirements.txt" file
-#   - using the package dependencies
+#   - using the "pyproject.toml" file
 #
 # Thomas Guillod - Dartmouth College
 # Mozilla Public License Version 2.0
