+++ .github/workflows/run_standard.yml
@@ -3,10 +3,6 @@
 #   - pull request
 #   - dispatch
 #
-# The test are run on:
-#   - Linux / Ubuntu
-#   - Python 3.10
-#
 # Thomas Guillod - Dartmouth College
 # Mozilla Public License Version 2.0
 
