+++ .github/workflows/run_pypi.yml
@@ -1,49 +1,0 @@
-# Run the "Pypi tests" with the following conditions.
-#   - dispatch
-#   - weekly
-#
-# Two different kinds of tests are made:
-#   - using the "requirements.txt" file
-#   - using the "pyproject.toml" file
-#
-# Thomas Guillod - Dartmouth College
-# Mozilla Public License Version 2.0
-
-name: Run PyPi Tests
-
-on:
-  workflow_dispatch:
-  schedule:
-    - cron: "0 2 * * 1"
-
-jobs:
-  default:
-    strategy:
-      fail-fast: false
-      matrix:
-        method:
-          - name: "requirements"
-            cmd:  |
-              pip install --upgrade pip
-              pip install --requirement requirements.txt
-          - name: "package"
-            cmd:  |
-              pip install --upgrade pip
-              pip install .
-    name: ${{ matrix.method.name }}
-    runs-on: "ubuntu-latest"
-    timeout-minutes: 20
-    steps:
-      - uses: actions/checkout@v4
-      - name: Set up Python
-        uses: actions/setup-python@v5
-        with:
-          python-version: "3.10"
-      - name: Install dependencies
-        shell: bash -el {0}
-        run: ${{ matrix.method.cmd }}
-      - name: Run the tests
-        shell: bash -el {0}
-        run: |
-          chmod +x scripts/run_tests.sh
-          ./scripts/run_tests.sh
