+++ .github/workflows/run_platform.yml
@@ -19,6 +19,7 @@
 name: Run Platform Tests
 
 on:
+  push:
   workflow_dispatch:
   schedule:
     - cron: "0 2 * * 1"
