+++ .github/workflows/run_conda.yml
@@ -1,47 +1,0 @@
-# Run the "Conda tests" with the following conditions.
-#   - dispatch
-#   - weekly
-#
-# Two different kinds of tests are made:
-#   - using the "conda_base.yaml" file
-#   - using the "conda_dev.yaml" file
-#
-# Thomas Guillod - Dartmouth College
-# Mozilla Public License Version 2.0
-
-name: Run Conda Tests
-
-on:
-  workflow_dispatch:
-  schedule:
-    - cron: "0 2 * * 1"
-
-jobs:
-  default:
-    strategy:
-      fail-fast: false
-      matrix:
-        method:
-          - name: "conda_base"
-            file:  "conda_base.yaml"
-          - name: "conda_dev"
-            file:  "conda_dev.yaml"
-    name: ${{ matrix.method.name }}
-    runs-on: "ubuntu-latest"
-    timeout-minutes: 20
-    steps:
-      - uses: actions/checkout@v4
-      - name: Set up Conda
-        uses: conda-incubator/setup-miniconda@v3
-        with:
-          miniforge-version: latest
-          python-version: "3.10"
-          auto-activate-base: false
-          activate-environment: pypeec
-          environment-file: ${{ matrix.method.file }}
-          use-mamba: true
-      - name: Run the tests
-        shell: bash -el {0}
-        run: |
-          chmod +x scripts/run_tests.sh
-          ./scripts/run_tests.sh
