+++ .github/workflows/run_platform.yml
@@ -48,5 +48,4 @@ jobs:
       - name: Run the tests
         shell: bash -el {0}
         run: |
-          chmod +x scripts/run_tests.sh
           ./scripts/run_tests.sh
