+++ .github/workflows/run_repository.yml
@@ -2,7 +2,7 @@
 #   - dispatch
 #   - weekly
 #
-# Two different kinds of tests are made:
+# The tests are run on the following repositories:
 #   - installing the package from the PyPi repository
 #   - installing the package from the Conda repository
 #
@@ -77,6 +77,7 @@ jobs:
         uses: actions/setup-python@v5
         with:
           python-version: "3.10"
+          cache: 'pip'
       - name: Install dependencies
         shell: bash -el {0}
         run: |
