+++ .github/workflows/run_hook.yml
@@ -23,7 +23,6 @@ jobs:
         uses: actions/setup-python@v5
         with:
           python-version: "3.10"
-          cache: 'pip'
       - name: Install dependencies
         shell: bash -el {0}
         run: |
