+++ .github/workflows/run_platform.yml
@@ -40,7 +40,6 @@ jobs:
         uses: actions/setup-python@v5
         with:
           python-version: ${{ matrix.version }}
-          cache: 'pip'
       - name: Install dependencies
         shell: bash -el {0}
         run: |
