+++ .github/workflows/run_repository.yml
@@ -77,7 +77,6 @@ jobs:
         uses: actions/setup-python@v5
         with:
           python-version: "3.10"
-          cache: 'pip'
       - name: Install dependencies
         shell: bash -el {0}
         run: |
