+++ .github/workflows/run_repository.yml
@@ -12,7 +12,6 @@
 name: Run Repository Tests
 
 on:
-  push:
   workflow_dispatch:
   schedule:
     - cron: "0 2 * * 1"
