+++ .github/workflows/run_platform.yml
@@ -2,7 +2,7 @@
 #   - dispatch
 #   - weekly
 #
-# The tests are run on the following platforms:
+# The tests are run on the following systems:
 #   - Linux / Ubuntu
 #   - Apple / macOS
 #   - Microsoft / Windows
