### 2024-07-04 f5781be
pyvista plot env vars
