+++ tests/code/test_pypeec.py
@@ -17,9 +17,6 @@ from pypeec import io
 # crash on warnings
 warnings.filterwarnings("error")
 
-# ignore deprecation warnings
-warnings.filterwarnings("ignore", category=DeprecationWarning)
-
 # disable logging to prevent clutter during test evaluation
 logging.disable(logging.INFO)
 
