### 2024-07-04 08b318c
env var for pyvista
