+++ .github/workflows/run_package.yml
@@ -18,11 +18,6 @@ on:
   schedule:
     - cron: "0 2 * * 1"
 
-env:
-  ALLOW_PLOTTING: "true"
-  PYTHONIOENCODING: "utf8"
-  PYTHONUNBUFFERED: "1"
-
 jobs:
   # ===================== run the tests with conda
   conda:
