+++ tests/code/test_pypeec.py
@@ -14,11 +14,11 @@ from pypeec import main
 from pypeec import script
 from pypeec import io
 
-# crash on warnings
-warnings.filterwarnings("error")
-
 # ignore deprecation warnings
-warnings.filterwarnings("ignore", category=DeprecationWarning)
+#warnings.filterwarnings("ignore", category=DeprecationWarning)
+
+# crash on warnings
+#warnings.filterwarnings("error")
 
 # disable logging to prevent clutter during test evaluation
 logging.disable(logging.INFO)
