+++ .github/workflows/run_package.yml
@@ -55,12 +55,12 @@ jobs:
         method:
           - name: "requirements"
             cmd: |
-              pip install --upgrade pip
-              pip install --requirement requirements.txt
+              python -m pip install --upgrade pip
+              python -m pip install --requirement requirements.txt
           - name: "pyproject"
             cmd: |
-              pip install --upgrade pip
-              pip install .
+              python -m pip install --upgrade pip
+              python -m pip install .
     name: ${{ matrix.method.name }}
     runs-on: "ubuntu-latest"
     timeout-minutes: 20
