+++ .github/workflows/run_repository.yml
@@ -78,8 +78,8 @@ jobs:
       - name: Install dependencies
         shell: bash -el {0}
         run: |
-          pip install --upgrade pip
-          pip install pypeec
+          python -m pip install --upgrade pip
+          python -m pip install pypeec
       - name: Run the tests
         shell: bash -el {0}
         run: |
