### 2024-07-09 c617a7c
new loop verification example
