+++ examples/examples_shape/parallel/geometry.yaml
@@ -1,0 +1,85 @@
+# Definition of a "file_geometry" data.
+#
+# This file is used by the mesher.
+# Definition of the voxel structure.
+#
+# Thomas Guillod - Dartmouth College
+# Mozilla Public License Version 2.0
+
+"mesh_type": "shape"
+"data_voxelize":
+    "param":
+        "dx": 50.0e-6
+        "dy": 50.0e-6
+        "dz": 500.0e-6
+        "cz": 0.0e-6
+        "simplify": 1.0e-6
+        "construct": null
+        "xy_min": null
+        "xy_max": null
+    "layer_stack":
+        - {"n_layer": 1, "tag_layer": "src"}
+        - {"n_layer": 50, "tag_layer": "wire"}
+        - {"n_layer": 1, "tag_layer": "sink"}
+    "geometry_shape":
+        "wire":
+            -
+                "shape_layer": ["wire"]
+                "shape_operation": "add"
+                "shape_type": "pad"
+                "shape_data":
+                    "diameter": 0.5e-3
+                    "coord": [[0.0e-3, +1.0e-3]]
+            -
+                "shape_layer": ["wire"]
+                "shape_operation": "add"
+                "shape_type": "pad"
+                "shape_data":
+                    "diameter": 0.5e-3
+                    "coord": [[0.0e-3, -1.0e-3]]
+        "src_1":
+            -
+                "shape_layer": ["src"]
+                "shape_operation": "add"
+                "shape_type": "pad"
+                "shape_data":
+                    "diameter": 0.5e-3
+                    "coord": [[0.0e-3, +1.0e-3]]
+        "src_2":
+            -
+                "shape_layer": ["src"]
+                "shape_operation": "add"
+                "shape_type": "pad"
+                "shape_data":
+                    "diameter": 0.5e-3
+                    "coord": [[0.0e-3, -1.0e-3]]
+        "sink_1":
+            -
+                "shape_layer": ["sink"]
+                "shape_operation": "add"
+                "shape_type": "pad"
+                "shape_data":
+                    "diameter": 0.5e-3
+                    "coord": [[0.0e-3, +1.0e-3]]
+        "sink_2":
+            -
+                "shape_layer": ["sink"]
+                "shape_operation": "add"
+                "shape_type": "pad"
+                "shape_data":
+                    "diameter": 0.5e-3
+                    "coord": [[0.0e-3, -1.0e-3]]
+"resampling":
+    "use_reduce": false
+    "use_resample": false
+    "resampling_factor": [1, 1, 1]
+"pts_cloud": []
+"check_cloud": true
+"check_conflict": true
+"check_connection": true
+"domain_conflict":
+    - {"domain_resolve": ["wire"], "domain_keep": ["src_1", "src_2"]}
+    - {"domain_resolve": ["wire"], "domain_keep": ["sink_1", "sink_2"]}
+"domain_connection":
+    "connect_1": {"domain_group": [["wire"], ["src_1"], ["sink_1"]], "connected": true}
+    "connect_2": {"domain_group": [["wire"], ["src_2"], ["sink_2"]], "connected": true}
