### 2024-07-09 ddd9cf4
new loop verification example documentation
