+++ examples/run_plotter.sh
@@ -15,7 +15,7 @@ FILE_SOLUTION="$PATH_ROOT/$FOLDER_EXAMPLE/solution.json.gz"
 FILE_PLOTTER="$PATH_ROOT/$FOLDER_CONFIG/plotter.yaml"
 
 # plot tag (from plotter.yaml)
-TAG_PLOT="V_c_norm J_c_norm H_norm residuum"
+TAG_PLOT="V_c_norm J_c_norm H_p_norm residuum"
 
 # run
 pypeec plotter \
