### 2024-07-23 9af173d
rename solver.py variables
