### 2024-07-23 5fe4129
use scipy mu_0
