### 2024-08-10 4d3cadd
better thread control for FFTW
