+++ examples/config/tolerance.yaml
@@ -42,8 +42,8 @@ sweep_pool: null
     #   - "CuPy" is extremely fast but require GPUs compatible with the CUDA toolkit.
     "library": "SciPy"
 
-    "scipy_worker": null               # number of workers for SciPy (null for number of CPUs)
-    "fftw_thread": null                # number of threads for FFTW (null for number of CPUs)
+    "scipy_worker": -1                 # number of workers for SciPy (0 for disabling, -1 for number of cores)
+    "fftw_thread": -1                  # number of threads for FFTW (0 for disabling, -1 for number of cores)
     "fftw_timeout": 100.0              # cache timeout in seconds for FFTW (null for no caching)
     "fftw_byte_align": 16              # size for byte alignment FFTW (null for default)
 
