### 2024-08-10 c1c8151
better thread control for PARDISO
