### 2024-08-10 8aa3500
better thread control for joblib
