+++ examples/config/tolerance.yaml
@@ -11,8 +11,9 @@
 
 # parallel pool for sweeps
 #   - number of parallel processes for computing the solver sweeps
-#   - if null, the parallel processing is disabled
-sweep_pool: null
+#   - if zero, the parallel processing is disabled
+#   - if negative, the number of cores if used
+sweep_pool: 0
 
 # control where numerical approximations are used for the Green and coupling functions
 #   - if the normalized voxel distance is smaller than the threshold, analytical solutions are used
