+++ bench/run_bench.py
@@ -44,5 +44,7 @@ LOGGER.info("======== end")
 now_2 = time.time()
 LOGGER.info("======== time = %f" % (now_2 - now_1))
 
+from joblib import Parallel, delayed, parallel_config
+parallel_config
 ###########################################################################################
 ###########################################################################################
