### 2024-08-10 8b6aa49
finner inner thread control
