+++ examples/examples_stl/inductor_toroid/geometry.yaml
@@ -1,0 +1,33 @@
+# Definition of a "file_geometry" data.
+#
+# This file is used by the mesher.
+# Definition of the voxel structure.
+#
+# Thomas Guillod - Dartmouth College
+# Mozilla Public License Version 2.0
+
+"mesh_type": "stl"
+"data_voxelize":
+    "param":
+        "d": [0.5e-3, 0.5e-3, 1.0e-3]
+        "xyz_min": null
+        "xyz_max": null
+    "domain_stl":
+        "src": [{"scale": 1.0, "offset": [0.0, 0.0, 0.0], "filename": !path "stl/src.stl"}]
+        "sink": [{"scale": 1.0, "offset": [0.0, 0.0, 0.0], "filename": !path "stl/sink.stl"}]
+        "wire": [{"scale": 1.0, "offset": [0.0, 0.0, 0.0], "filename": !path "stl/wire.stl"}]
+        "core": [{"scale": 1.0, "offset": [0.0, 0.0, 0.0], "filename": !path "stl/core.stl"}]
+"resampling":
+    "use_reduce": false
+    "use_resample": false
+    "resampling_factor": [1, 1, 1]
+"pts_cloud": []
+"check_cloud": true
+"check_conflict": true
+"check_connection": true
+"domain_conflict":
+    - {"domain_resolve": ["wire"], "domain_keep": ["src"]}
+    - {"domain_resolve": ["wire"], "domain_keep": ["sink"]}
+"domain_connection":
+    "wire": {"domain_group": [["src"], ["sink"], ["wire"]], "connected": true}
+    "core": {"domain_group": [["wire"], ["core"]], "connected": false}
