+++ examples/examples_config.py
@@ -33,6 +33,7 @@ FOLDER_CONFIG = "config"
 #   examples_stl
 #       examples_stl/inductor_air
 #       examples_stl/inductor_core
+#       examples_stl/inductor_toroid
 #       examples_stl/transformer
 #   examples_png
 #       examples_png/inductor_spiral
