+++ examples/examples_config.sh
@@ -28,6 +28,7 @@ FOLDER_CONFIG="config"
 #   examples_stl
 #       examples_stl/inductor_air
 #       examples_stl/inductor_core
+#       examples_stl/inductor_toroid
 #       examples_stl/transformer
 #   examples_png
 #       examples_png/inductor_spiral
