+++ examples/examples_stl/inductor_toroid/geometry.yaml
@@ -13,9 +13,9 @@
         "xyz_min": null
         "xyz_max": null
     "domain_stl":
-        "src": [{"scale": 1.0, "offset": [0.0, 0.0, 0.0], "filename": !path "stl/src.stl"}]
-        "sink": [{"scale": 1.0, "offset": [0.0, 0.0, 0.0], "filename": !path "stl/sink.stl"}]
         "wire": [{"scale": 1.0, "offset": [0.0, 0.0, 0.0], "filename": !path "stl/wire.stl"}]
+        "src": [ { "scale": 1.0, "offset": [ 0.0, 0.0, 0.0 ], "filename": !path "stl/src.stl" } ]
+        "sink": [ { "scale": 1.0, "offset": [ 0.0, 0.0, 0.0 ], "filename": !path "stl/sink.stl" } ]
         "core": [{"scale": 1.0, "offset": [0.0, 0.0, 0.0], "filename": !path "stl/core.stl"}]
 "resampling":
     "use_reduce": false
