### 2024-08-15 otvam/pypeec#7 @thelfer
@otvam Thanks for closing the issue, I shall have done it myself.
