### 2024-08-21 5d50cdb
flag for doing random conflict resolution
