+++ examples/tutorial/geometry.yaml
@@ -110,6 +110,7 @@
 "check_cloud": true
 "check_conflict": true
 "check_connection": true
+"random_resolution": false
 "domain_conflict":
     - {"domain_resolve": ["trace"], "domain_keep": ["src"]}
     - {"domain_resolve": ["trace"], "domain_keep": ["sink"]}
