+++ examples/examples_voxel/distributed/geometry.yaml
@@ -24,6 +24,7 @@
 "check_cloud": true
 "check_conflict": true
 "check_connection": true
+"random_resolution": false
 "domain_conflict": []
 "domain_connection":
     "connect": {"domain_group": [["cond"], ["src"], ["sink"]], "connected": true}
