+++ examples/examples_png/inductor_gap/geometry.yaml
@@ -33,6 +33,7 @@
 "check_cloud": true
 "check_conflict": true
 "check_connection": true
+"random_resolution": false
 "domain_conflict": []
 "domain_connection":
     "cond": {"domain_group": [["coil"], ["src"], ["sink"]], "connected": true}
