+++ examples/examples_png/inductor_spiral/geometry.yaml
@@ -28,6 +28,7 @@
 "check_cloud": true
 "check_conflict": true
 "check_connection": true
+"random_resolution": false
 "domain_conflict": []
 "domain_connection":
     "connect": {"domain_group": [["coil"], ["src"], ["sink"]], "connected": true}
