+++ examples/examples_png/iron_core/geometry.yaml
@@ -34,6 +34,7 @@
 "check_cloud": true
 "check_conflict": true
 "check_connection": true
+"random_resolution": false
 "domain_conflict": []
 "domain_connection":
     "cond": {"domain_group": [["wire"], ["src"], ["sink"]], "connected": true}
