### 2024-09-23 #6644 @likeajumprope
> I just wanted to ask the status of the review process and if there is anything that I do to assist (like trying to find reviewers in the power electronics community)?
>
> Best, Thomas

If you could provide some suggestions for reveiwers that would be fantastic!  I have not have any luck so far
