+++ paper.bib
@@ -136,6 +136,15 @@ title={{A Fast Tool for the Parametric Analysis of Human Body Exposed to LF Elec
   url={https://doi.org/10.1109/ICEMIC.2003.1287800}
 }
 
+@book{peec_book,
+  title={Circuit Oriented Electromagnetic Modeling Using the PEEC Techniques},
+  author={Ruehli, A. and Antonini, G. and Jiang, L.},
+  isbn={9781118436646},
+  series={IEEE Press},
+  year={2017},
+  publisher={Wiley},
+}
+
 @misc{peec_code,
   author={Torchio, R. and Lucchini, L.},
   title={{FFT-PEEC}},
