### 2024-09-23 86d14ba
add pypi and conda
