+++ paper.bib
@@ -137,7 +137,7 @@ title={{A Fast Tool for the Parametric Analysis of Human Body Exposed to LF Elec
 }
 
 @book{peec_book,
-  title={Circuit Oriented Electromagnetic Modeling Using the PEEC Techniques},
+  title={{Circuit Oriented Electromagnetic Modeling Using the PEEC Techniques}},
   author={Ruehli, A. and Antonini, G. and Jiang, L.},
   isbn={9781118436646},
   series={IEEE Press},
