### 2024-09-23 #6644 @otvam
No problem! And I agree that it would be good for the second reviewer to have an electromagnetic and/or power electronic background.

I will ask in the power electronic community in someone would be interested.
