+++ conda_dev.yaml
@@ -26,14 +26,14 @@ dependencies:
   - setuptools_scm=7.1.0
   - sphinx_rtd_theme=1.2.2
   - sphinx=6.2.1
-  - pkg-config==0.29.2
+  - pkg-config=0.29.2
   # YAML parsing library
   - pyyaml=6.0
   # numerical libraries
   - numpy=1.24.1
   - scipy=1.9.3
   - joblib=1.3.0
-  # geometry manipulation
+  # geometry libraries
   - vtk=9.2.6
   - pyvista=0.40.0
   - shapely=2.0.1
@@ -43,18 +43,18 @@ dependencies:
   - pyqt=5.15.7
   - pyvistaqt=0.11.0
   - matplotlib=3.7.1
-  # optional Jupyter libs
+  # optional packages for Jupyter
   - jupyterlab=4.0.4
   - ipywidgets=8.1.0
   - trame=3.1.0
-  # optional HPC libs
+  # optional optimized numerical libraries
   - pyamg=5.0.1
   - cupy=12.1.0
   - scikit-umfpack=0.3.3
-  # optional HPC toolkits
-  - mkl-devel==2023.2.0
-  - cudatoolkit==11.8.0
-  # optional HPC libs from PyPi
+  # optional optimized numerical toolkits
+  - mkl-devel=2023.2.0
+  - cudatoolkit=11.8.0
+  # optional optimized numerical libraries
   - pip:
     - pyFFTW==0.13.1
     - pydiso==0.0.4
