+++ conda_base.yaml
@@ -22,7 +22,7 @@ dependencies:
   - numpy=1.24.1
   - scipy=1.9.3
   - joblib=1.3.0
-  # geometry manipulation
+  # geometry libraries
   - vtk=9.2.6
   - pyvista=0.40.0
   - shapely=2.0.1
