### 2024-09-30 #6644 @otvam
I would have two suggestions for reviewers: @imperialiluc and @svenweihe

There are both researcher in power electronics at ETH Zurich, Switzerland. Power electronics being the main target audience for the proposed software package.
