### 2024-10-02 c44ea11
naive singleton for the PQ app (require X server)
