### 2024-10-02 2eedb4a
singleton class for the Qt app
