### 2024-10-03 ae3af97
connection checks for empty domains
