### 2024-10-03 efafc09
refactor connection checks
