### 2024-10-03 067ec2d
preparation for adjacent voxel checks
