### 2024-10-03 800daf5
reafactor connected components
