### 2024-10-03 41a9082
check adjacent domains
