+++ pypeec/lib_mesher/voxel_integrity.py
@@ -23,26 +23,29 @@ def _get_all_indices(domain_def):
 
     # get the indices and colors
     for tag in domain_def:
-        idx_tmp = domain_def[tag]
-        idx = np.append(idx, idx_tmp)
+        idx = np.append(idx, domain_def[tag])
 
     return idx
 
 
-def _get_tag_indices(domain_def, domain_list):
+def _get_group_indices(domain_def, domain_group):
     """
     Get the indices of the non-empty voxels for the given domains.
     """
 
-    # init
-    idx = np.empty(0, dtype=np.int_)
+    group_def = []
+    for domain_group_tmp in domain_group:
+        # init
+        idx_tmp = np.empty(0, dtype=np.int_)
 
-    # get the indices and colors
-    for tag in domain_list:
-        idx_tmp = domain_def[tag]
-        idx = np.append(idx, idx_tmp)
+        # get the indices and colors
+        for tag in domain_group_tmp:
+            idx_tmp = np.append(idx_tmp, domain_def[tag])
 
-    return idx
+        if len(idx_tmp) > 0:
+            group_def.append(idx_tmp)
+
+    return group_def
 
 
 def _get_connection_matrix(n):
@@ -65,37 +68,46 @@ def _get_connection_matrix(n):
     idx = idx_x+idx_y*nx+idx_z*nx*ny
 
     # create the sparse matrix
-    voxel_connection = sps.csc_matrix((nv, nv), dtype=np.int_)
+    voxel_matrix = sps.csc_matrix((nv, nv), dtype=np.int_)
+
+    # self connections
+    idx_col = np.arange(nv)
+    idx_row = np.arange(nv)
+    data = np.ones(nv, dtype=np.int_)
+    voxel_matrix += sps.csc_matrix((data, (idx_row, idx_col)), shape=(nv, nv), dtype=np.int_)
 
     # connections along x direction
-    idx_col = idx[0:-1, :, :].flatten()
-    idx_row = idx[1:, :, :].flatten()
+    idx_col = idx[:-1, :, :].flatten()
+    idx_row = idx[+1:, :, :].flatten()
     data = np.ones((nx-1)*ny*nz, dtype=np.int_)
-    voxel_connection += sps.csc_matrix((data, (idx_row, idx_col)), shape=(nv, nv), dtype=np.int_)
+    voxel_matrix += sps.csc_matrix((data, (idx_row, idx_col)), shape=(nv, nv), dtype=np.int_)
+    voxel_matrix += sps.csc_matrix((data, (idx_col, idx_row)), shape=(nv, nv), dtype=np.int_)
 
     # connections along y direction
-    idx_col = idx[:, 0:-1, :].flatten()
-    idx_row = idx[:, 1:, :].flatten()
+    idx_col = idx[:, :-1, :].flatten()
+    idx_row = idx[:, +1:, :].flatten()
     data = np.ones(nx*(ny-1)*nz, dtype=np.int_)
-    voxel_connection += sps.csc_matrix((data, (idx_row, idx_col)), shape=(nv, nv), dtype=np.int_)
+    voxel_matrix += sps.csc_matrix((data, (idx_row, idx_col)), shape=(nv, nv), dtype=np.int_)
+    voxel_matrix += sps.csc_matrix((data, (idx_col, idx_row)), shape=(nv, nv), dtype=np.int_)
 
     # connections along z direction
-    idx_col = idx[:, :, 0:-1].flatten()
-    idx_row = idx[:, :, 1:].flatten()
+    idx_col = idx[:, :, :-1].flatten()
+    idx_row = idx[:, :, +1:].flatten()
[…]
