+++ examples/examples_voxel/anisotropic/geometry.yaml
@@ -21,10 +21,12 @@
     "use_resample": false
     "resampling_factor": [1, 1, 1]
 "pts_cloud": []
-"check_cloud": true
-"check_conflict": true
-"check_connection": true
+"resolve_cloud": true
+"resolve_conflict": true
+"check_integrity": true
 "random_resolution": false
 "domain_conflict": []
 "domain_connection":
+    "connect": {"domain_group": [["cond"], ["src"], ["sink"]], "connected": true}
+"domain_adjacent":
     "connect": {"domain_group": [["cond"], ["src"], ["sink"]], "connected": true}
