### 2024-10-03 f33d6d8
lift example with the new mesh checks
