+++ examples/examples_png/gerber/geometry.yaml
@@ -25,12 +25,14 @@
     "use_resample": true
     "resampling_factor": [1, 1, 1]
 "pts_cloud": []
-"check_cloud": true
-"check_conflict": true
-"check_connection": true
+"resolve_cloud": true
+"resolve_conflict": true
+"check_integrity": true
 "random_resolution": false
 "domain_conflict":
     - {"domain_resolve": ["coil"], "domain_keep": ["src"]}
     - {"domain_resolve": ["coil"], "domain_keep": ["sink"]}
-"domain_connection":
-    "connect": {"domain_group": [["coil"], ["src"], ["sink"]], "connected": true}
+"domain_connected":
+    "conductor": {"domain_group": [["coil"], ["src"], ["sink"]], "connected": true}
+"domain_adjacent":
+    "terminal": {"domain_group": [["src"], ["sink"]], "connected": false}
