+++ examples/examples_stl/inductor_toroid/geometry.yaml
@@ -13,7 +13,7 @@
         "xyz_min": null
         "xyz_max": null
     "domain_stl":
-        "wire": [{"scale": 1.0, "offset": [0.0, 0.0, 0.0], "filename": !path "stl/wire.stl"}]
+        "coil": [{"scale": 1.0, "offset": [0.0, 0.0, 0.0], "filename": !path "stl/coil.stl"}]
         "src": [ { "scale": 1.0, "offset": [ 0.0, 0.0, 0.0 ], "filename": !path "stl/src.stl" } ]
         "sink": [ { "scale": 1.0, "offset": [ 0.0, 0.0, 0.0 ], "filename": !path "stl/sink.stl" } ]
         "core": [{"scale": 1.0, "offset": [0.0, 0.0, 0.0], "filename": !path "stl/core.stl"}]
@@ -22,13 +22,15 @@
     "use_resample": false
     "resampling_factor": [1, 1, 1]
 "pts_cloud": []
-"check_cloud": true
-"check_conflict": true
-"check_connection": true
+"resolve_cloud": true
+"resolve_conflict": true
+"check_integrity": true
 "random_resolution": false
 "domain_conflict":
-    - {"domain_resolve": ["wire"], "domain_keep": ["src"]}
-    - {"domain_resolve": ["wire"], "domain_keep": ["sink"]}
-"domain_connection":
-    "wire": {"domain_group": [["src"], ["sink"], ["wire"]], "connected": true}
-    "core": {"domain_group": [["wire"], ["core"]], "connected": false}
+    - {"domain_resolve": ["coil"], "domain_keep": ["src"]}
+    - {"domain_resolve": ["coil"], "domain_keep": ["sink"]}
+"domain_connected":
+    "conductor": {"domain_group": [["coil"], ["src"], ["sink"]], "connected": true}
+    "core": {"domain_group": [["core"], ["coil", "src", "sink"]], "connected": false}
+"domain_adjacent":
+    "terminal": {"domain_group": [["src"], ["sink"]], "connected": false}
