+++ examples/examples_shape/wire/geometry.yaml
@@ -51,12 +51,14 @@
     "use_resample": false
     "resampling_factor": [1, 1, 1]
 "pts_cloud": []
-"check_cloud": true
-"check_conflict": true
-"check_connection": true
+"resolve_cloud": true
+"resolve_conflict": true
+"check_integrity": true
 "random_resolution": false
 "domain_conflict":
     - {"domain_resolve": ["wire"], "domain_keep": ["src"]}
     - {"domain_resolve": ["wire"], "domain_keep": ["sink"]}
-"domain_connection":
-    "connect": {"domain_group": [["wire"], ["src"], ["sink"]], "connected": true}
+"domain_connected":
+    "conductor": {"domain_group": [["wire"], ["src"], ["sink"]], "connected": true}
+"domain_adjacent":
+    "terminal": {"domain_group": [["src"], ["sink"]], "connected": false}
