+++ examples/examples_shape/trace/geometry.yaml
@@ -80,14 +80,17 @@
     "use_resample": false
     "resampling_factor": [1, 1, 1]
 "pts_cloud": []
-"check_cloud": true
-"check_conflict": true
-"check_connection": true
+"resolve_cloud": true
+"resolve_conflict": true
+"check_integrity": true
 "random_resolution": false
 "domain_conflict":
     - {"domain_resolve": ["copper"], "domain_keep": ["src_front"]}
     - {"domain_resolve": ["copper"], "domain_keep": ["src_back"]}
     - {"domain_resolve": ["copper"], "domain_keep": ["sink"]}
-"domain_connection":
-    "top": {"domain_group": [["copper"], ["src_front"], ["sink"]], "connected": true}
-    "bottom": {"domain_group": [["copper"], ["src_back"], ["sink"]], "connected": true}
+"domain_connected":
+    "conductor_front": {"domain_group": [["src_front"], ["sink"]], "connected": true}
+    "conductor_back": {"domain_group": [["src_back"], ["sink"]], "connected": true}
+"domain_adjacent":
+    "terminal_front": {"domain_group": [["src_front"], ["sink"]], "connected": false}
+    "terminal_back": {"domain_group": [["src_back"], ["sink"]], "connected": false}
