+++ examples/examples_png/shield/geometry.yaml
@@ -76,11 +76,13 @@
     - [+05.00e-3, +10.00e-3, +08.00e-3]
     - [+10.00e-3, +10.00e-3, -08.00e-3]
     - [+10.00e-3, +10.00e-3, +08.00e-3]
-"check_cloud": true
-"check_conflict": true
-"check_connection": true
+"resolve_cloud": true
+"resolve_conflict": true
+"check_integrity": true
 "random_resolution": false
 "domain_conflict": []
-"domain_connection":
-    "cond": {"domain_group": [["coil"], ["src"], ["sink"]], "connected": true}
-    "core": {"domain_group": [["coil"], ["core"]], "connected": false}
+"domain_connected":
+    "conductor": {"domain_group": [["coil"], ["src"], ["sink"]], "connected": true}
+    "core": {"domain_group": [["coil", "src", "sink"], ["core"]], "connected": false}
+"domain_adjacent":
+    "terminal": {"domain_group": [["src"], ["sink"]], "connected": false}
