### 2024-10-03 df111eb
make example naming consistent
