+++ examples/examples_voxel/logo/geometry.yaml
@@ -13,7 +13,7 @@
         "d": [1.0, 1.0, 1.0]
         "c": [0.0, 0.0, 0.0]
     "domain_index":
-        "cond": [9, 0, 1, 2, 11, 20, 23, 26, 25, 24, 15, 6, 7]
+        "wire": [9, 0, 1, 2, 11, 20, 23, 26, 25, 24, 15, 6, 7]
         "src": [18]
         "sink": [8]
 "resampling":
@@ -27,6 +27,6 @@
 "random_resolution": false
 "domain_conflict": []
 "domain_connected":
-    "conductor": {"domain_group": [["cond"], ["src"], ["sink"]], "connected": true}
+    "conductor": {"domain_group": [["wire"], ["src"], ["sink"]], "connected": true}
 "domain_adjacent":
     "terminal": {"domain_group": [["src"], ["sink"]], "connected": false}
