+++ examples/examples_shape/wire/problem.yaml
@@ -7,7 +7,7 @@
 # Mozilla Public License Version 2.0
 
 "material_def":
-    "wire":
+    "copper":
         "domain_list": ["src", "sink", "wire"]
         "material_type": "electric"
         "orientation_type": "isotropic"
@@ -28,14 +28,14 @@
     "sim_dc":
         "freq": 0.0
         "material_val":
-            "wire": {"rho_re": 1.75e-8, "rho_im": 0.0}
+            "copper": {"rho_re": 1.75e-8, "rho_im": 0.0}
         "source_val":
             "src": {"V_re": +1.0, "V_im": 0.0, "Z_re": 1.0e-3, "Z_im": 0.0}
             "sink": {"V_re": -1.0, "V_im": 0.0, "Z_re": 1.0e-3, "Z_im": 0.0}
     "sim_ac":
         "freq": 100.0e+3
         "material_val":
-            "wire": {"rho_re": 1.75e-8, "rho_im": 0.0}
+            "copper": {"rho_re": 1.75e-8, "rho_im": 0.0}
         "source_val":
             "src": {"V_re": +1.0, "V_im": 0.0, "Z_re": 1.0e-3, "Z_im": 0.0}
             "sink": {"V_re": -1.0, "V_im": 0.0, "Z_re": 1.0e-3, "Z_im": 0.0}
