### 2024-10-03 4733770
pcb hole example
