+++ tests/test_shape.py
@@ -18,7 +18,7 @@ class TestShape(test_workflow.TestWorkflow):
 name_list = [
     "examples_shape/busbar",
     "examples_shape/wire",
-    "examples_shape/trace",
+    "examples_shape/hole",
     "examples_shape/coplanar",
     "examples_shape/parallel",
 ]
