+++ examples/examples_voxel/core/geometry.yaml
@@ -13,9 +13,11 @@
         "d": [10.0e-3, 10.0e-3, 10.0e-3]
         "c": [0.0, 0.0, 0.0]
     "domain_index":
-        "core": [50, 51, 52, 53, 54, 55, 59, 60, 64, 65, 69, 70, 71, 72, 73, 74, 75,
-                 76, 77, 78, 79, 80, 84, 85, 89, 90, 94, 95, 96, 97, 98, 99, 100, 101,
-                 102, 103, 104, 105, 109, 110, 114, 115, 119, 120, 121, 122, 123, 124]
+        "core": [
+            50, 51, 52, 53, 54, 55, 59, 60, 64, 65, 69, 70, 71, 72, 73, 74, 75,
+            76, 77, 78, 79, 80, 84, 85, 89, 90, 94, 95, 96, 97, 98, 99, 100, 101,
+            102, 103, 104, 105, 109, 110, 114, 115, 119, 120, 121, 122, 123, 124,
+        ]
         "wire": [37, 62, 87, 112, 137]
         "src": [12]
         "sink": [162]
