### 2024-10-04 40a73bb
complete refactoring of the mesher data format
