+++ pypeec/lib_check/check_data_geometry.py
@@ -13,7 +13,26 @@ from pypeec.lib_check import check_data_geometry_shape
 from pypeec.lib_check import check_data_geometry_voxel
 
 
-def _check_resampling(resampling):
+def _check_data_point(data_point):
+    """
+    Check the validity of cloud point parameters.
+    """
+
+    # check type
+    key_list = [
+        "check_cloud",
+        "full_cloud",
+        "pts_cloud",
+    ]
+    datachecker.check_dict("data_point", data_point, key_list=key_list)
+
+    # check data
+    datachecker.check_boolean("check_cloud", data_point["check_cloud"])
+    datachecker.check_boolean("full_cloud", data_point["full_cloud"])
+    datachecker.check_float_pts("pts_cloud", data_point["pts_cloud"], size=3, can_be_empty=True)
+
+
+def _check_data_resampling(data_resampling):
     """
     Check the validity of resampling parameters.
     """
@@ -24,31 +43,27 @@ def _check_resampling(resampling):
         "use_resample",
         "resampling_factor",
     ]
-    datachecker.check_dict("resampling", resampling, key_list=key_list)
+    datachecker.check_dict("data_resampling", data_resampling, key_list=key_list)
 
     # check data
-    datachecker.check_boolean("use_reduce", resampling["use_reduce"])
-    datachecker.check_boolean("use_resample", resampling["use_resample"])
-    datachecker.check_integer_array("resampling_factor", resampling["resampling_factor"], size=3, is_positive=True, can_be_zero=False)
+    datachecker.check_boolean("use_reduce", data_resampling["use_reduce"])
+    datachecker.check_boolean("use_resample", data_resampling["use_resample"])
+    datachecker.check_integer_array("resampling_factor", data_resampling["resampling_factor"], size=3, is_positive=True, can_be_zero=False)
 
 
-def _check_domain_conflict(domain_list, domain_conflict):
+def _check_conflict_rules(domain_list, conflict_rules):
     """
-    Check the validity of the rules to solve conflict between domains (STL mesher).
+    Check the conflict resolution rules.
     """
 
-    # check type
-    datachecker.check_list("domain_conflict", domain_conflict, can_be_empty=True)
-
-    # check value
-    for domain_conflict_tmp in domain_conflict:
+    for conflict_rules_tmp in conflict_rules:
         # check type
         key_list = ["domain_keep", "domain_resolve"]
-        datachecker.check_dict("domain_conflict", domain_conflict_tmp, key_list=key_list)
+        datachecker.check_dict("conflict_rules", conflict_rules_tmp, key_list=key_list)
 
         # extract the data
-        domain_keep = domain_conflict_tmp["domain_keep"]
-        domain_resolve = domain_conflict_tmp["domain_resolve"]
+        domain_keep = conflict_rules_tmp["domain_keep"]
+        domain_resolve = conflict_rules_tmp["domain_resolve"]
 
         # check type
         datachecker.check_list("domain_keep", domain_keep, sub_type=str, can_be_empty=False)
@@ -61,10 +76,29 @@ def _check_domain_conflict(domain_list, domain_conflict):
             datachecker.check_choice("domain_resolve", tag, domain_list)
 
 
+def _check_data_conflict(domain_list, data_conflict):
+    """
+    Check the validity of the rules to solve conflict between domains.
+    """
+
+    # check type
+    key_list = [
+        "resolve_rules",
+        "resolve_random",
+        "conflict_rules",
+    ]
+    datachecker.check_dict("data_conflict", data_conflict, key_list=key_list)
+
+    # check type
+    datachecker.check_boolean("resolve_rules", data_conflict["resolve_rules"])
[…]
