+++ examples/examples_voxel/logo/geometry.yaml
@@ -16,17 +16,20 @@
         "wire": [9, 0, 1, 2, 11, 20, 23, 26, 25, 24, 15, 6, 7]
         "src": [18]
         "sink": [8]
-"resampling":
+"data_point":
+    "check_cloud": true
+    "full_cloud": true
+    "pts_cloud": []
+"data_resampling":
     "use_reduce": false
     "use_resample": false
     "resampling_factor": [1, 1, 1]
-"pts_cloud": []
-"resolve_cloud": true
-"resolve_conflict": false
-"check_integrity": true
-"random_resolution": false
-"domain_conflict": []
-"domain_connected":
-    "conductor": {"domain_group": [["wire"], ["src"], ["sink"]], "connected": true}
-"domain_adjacent":
-    "terminal": {"domain_group": [["src"], ["sink"]], "connected": false}
+"data_conflict":
+    "resolve_rules": true
+    "resolve_random": false
+    "conflict_rules": []
+"data_integrity":
+    "domain_connected":
+        "conductor": {"domain_group": [["wire"], ["src"], ["sink"]], "connected": true}
+    "domain_adjacent":
+        "terminal": {"domain_group": [["src"], ["sink"]], "connected": false}
