+++ examples/examples_shape/parallel/geometry.yaml
@@ -69,22 +69,25 @@
                 "shape_data":
                     "diameter": 0.5e-3
                     "coord": [[0.0e-3, -1.0e-3]]
-"resampling":
+"data_point":
+    "check_cloud": true
+    "full_cloud": true
+    "pts_cloud": []
+"data_resampling":
     "use_reduce": false
     "use_resample": false
     "resampling_factor": [1, 1, 1]
-"pts_cloud": []
-"resolve_cloud": true
-"resolve_conflict": true
-"check_integrity": true
-"random_resolution": false
-"domain_conflict":
-    - {"domain_resolve": ["wire"], "domain_keep": ["pri_src", "sec_src"]}
-    - {"domain_resolve": ["wire"], "domain_keep": ["pri_sink", "sec_sink"]}
-"domain_connected":
-    "pri_conductor": {"domain_group": [["pri_src"], ["pri_sink"]], "connected": true}
-    "sec_conductor": {"domain_group": [["sec_src"], ["sec_sink"]], "connected": true}
-    "insulation": {"domain_group": [["pri_src", "pri_sink"], ["sec_src", "sec_sink"]], "connected": false}
-"domain_adjacent":
-    "pri_terminal": {"domain_group": [["pri_src"], ["pri_sink"]], "connected": false}
-    "sec_terminal": {"domain_group": [["sec_src"], ["sec_sink"]], "connected": false}
+"data_conflict":
+    "resolve_rules": true
+    "resolve_random": false
+    "conflict_rules":
+        - {"domain_resolve": ["wire"], "domain_keep": ["pri_src", "sec_src"]}
+        - {"domain_resolve": ["wire"], "domain_keep": ["pri_sink", "sec_sink"]}
+"data_integrity":
+    "domain_connected":
+        "pri_conductor": {"domain_group": [["pri_src"], ["pri_sink"]], "connected": true}
+        "sec_conductor": {"domain_group": [["sec_src"], ["sec_sink"]], "connected": true}
+        "insulation": {"domain_group": [["pri_src", "pri_sink"], ["sec_src", "sec_sink"]], "connected": false}
+    "domain_adjacent":
+        "pri_terminal": {"domain_group": [["pri_src"], ["pri_sink"]], "connected": false}
+        "sec_terminal": {"domain_group": [["sec_src"], ["sec_sink"]], "connected": false}
