+++ examples/examples_shape/hole/geometry.yaml
@@ -52,19 +52,22 @@
                 "shape_data":
                     "diameter": 1.0e-3
                     "coord": [[0.0e-3, 6.0e-3]]
-"resampling":
+"data_point":
+    "check_cloud": true
+    "full_cloud": true
+    "pts_cloud": []
+"data_resampling":
     "use_reduce": false
     "use_resample": false
     "resampling_factor": [1, 1, 1]
-"pts_cloud": []
-"resolve_cloud": true
-"resolve_conflict": true
-"check_integrity": true
-"random_resolution": false
-"domain_conflict":
-    - {"domain_resolve": ["trace"], "domain_keep": ["src"]}
-    - {"domain_resolve": ["trace"], "domain_keep": ["sink"]}
-"domain_connected":
-    "conductor": {"domain_group": [["src"], ["sink"], ["trace"]], "connected": true}
-"domain_adjacent":
-    "terminal_front": {"domain_group": [["src"], ["sink"]], "connected": false}
+"data_conflict":
+    "resolve_rules": true
+    "resolve_random": false
+    "conflict_rules":
+        - {"domain_resolve": ["trace"], "domain_keep": ["src"]}
+        - {"domain_resolve": ["trace"], "domain_keep": ["sink"]}
+"data_integrity":
+    "domain_connected":
+        "conductor": {"domain_group": [["src"], ["sink"], ["trace"]], "connected": true}
+    "domain_adjacent":
+        "terminal_front": {"domain_group": [["src"], ["sink"]], "connected": false}
