### 2024-10-04 b61dd27
documentation of the new mesher format
