+++ examples/examples_stl/inductor_toroid/problem.yaml
@@ -28,6 +28,7 @@
         "var_type": "lumped"
 "material_val": &material_val
     "copper": {"rho_re": 1.75e-8, "rho_im": 0.0}
+    "core": {"chi_re": 500.0, "chi_im": 0.0}
 "source_val": &source_val
     "src": {"V_re": +1.0, "V_im": 0.0, "Z_re": 1.0e-3, "Z_im": 0.0}
     "sink": {"V_re": -1.0, "V_im": 0.0, "Z_re": 1.0e-3, "Z_im": 0.0}
