### 2024-10-07 a3f21de
io module exception messages
