### 2024-10-09 2ddf00e
encode numpy scalar for JSON
