+++ pypeec/io.py
@@ -1,565 +1,0 @@
-"""
-Module for serialization and deserialization.
-    - load JSON/YAML input files
-    - load and write JSON/Pickle data files
-
-For YAML files, the following custom extensions are used:
-    - "!path" - parse relative paths (with respect to the YAML file)
-    - "!include" - include other YAML files (recursion possible)
-    - "!env" - include YAML string from environment variables
-    - "!merge_dict" - merge a list of dicts
-    - "!merge_list" - merge a list of lists
-
-For JSON files, the following custom extensions are used:
-    - "__complex__" - allows the serialization of complex numbers
-    - "__numpy__" - allows the serialization of NumPy arrays
-
-The JSON/YAML files with the custom extensions are still valid JSON/YAML files.
-
-The JSON files can be serialized/deserialized as/from:
-    - text files
-    - gzip files
-
-Warning
--------
-    - Pickling data is not secure.
-    - Only load pickle files that you trust.
-"""
-
-__author__ = "Thomas Guillod"
-__copyright__ = "Thomas Guillod - Dartmouth College"
-__license__ = "Mozilla Public License Version 2.0"
-
-import os.path
-import json
-import pickle
-import gzip
-import yaml
-import numpy as np
-
-
-class _YamlLoader(yaml.Loader):
-    """
-    This Python class offers extension to the YAML format.
-        - include YAML file in YAML file
-        - include relative filesystem paths
-        - merge list of dicts
-        - merge list of lists
-    """
-
-    def __init__(self, stream):
-        """
-        Constructor.
-        Custom YAML loader subclassing the default loader.
-        """
-
-        # get the path of the YAML file for relative paths
-        self.path_root = os.path.abspath(stream.name)
-        self.path_root = os.path.dirname(self.path_root)
-
-        # flag indicating if any merge commands are used
-        self.has_merge = False
-
-        # call the constructor of the parent
-        super().__init__(stream)
-
-        # handling of YAML files inclusion
-        def fct_handle_include(self, node):
-            res = _YamlLoader._yaml_handling(self, node, self._extract_yaml)
-            return res
-
-        # handling of relative paths
-        def fct_handle_path(self, node):
-            res = _YamlLoader._yaml_handling(self, node, self._extract_path)
-            return res
-
-        # handling of environment variables inclusion
-        def fct_handle_env(self, node):
-            res = _YamlLoader._yaml_handling(self, node, self._extract_env)
-            return res
-
-        # handling merge of a list of dicts
-        def fct_handle_merge_dict(self, node):
-            self.has_merge = True
-            res = _MergeObj(self.construct_sequence(node), "dict")
-            return res
-
-        # handling merge of a list of lists
-        def fct_handle_merge_list(self, node):
-            self.has_merge = True
-            res = _MergeObj(self.construct_sequence(node), "list")
-            return res
-
-        # add the extension to the YAML format
-        _YamlLoader.add_constructor("!include", fct_handle_include)
-        _YamlLoader.add_constructor("!path", fct_handle_path)
-        _YamlLoader.add_constructor("!env", fct_handle_env)
-        _YamlLoader.add_constructor("!merge_dict", fct_handle_merge_dict)
-        _YamlLoader.add_constructor("!merge_list", fct_handle_merge_list)
-
-    def _yaml_handling(self, node, fct):
-        """
-        Apply a function to a YAML node for list, dict, scalar.
-        """
-
-        if isinstance(node, yaml.ScalarNode):
-            return fct(self.construct_scalar(node))
-        elif isinstance(node, yaml.SequenceNode):
-            result = []
-            for arg in self.construct_sequence(node):
-                result.append(fct(arg))
-            return result
[…]
