+++ pypeec/lib_mesher/voxel_resampling.py
@@ -7,11 +7,11 @@ __author__ = "Thomas Guillod"
 __copyright__ = "Thomas Guillod - Dartmouth College"
 __license__ = "Mozilla Public License Version 2.0"
 
+import scilogger
 import numpy as np
-from pypeec import log
 
 # get a logger
-LOGGER = log.get_logger(__name__, "pypeec")
+LOGGER = scilogger.get_logger(__name__, "pypeec")
 
 
 def _get_idx_resample_tensor(resampling_factor, idx_n, idx_r, idx):
