+++ pypeec/run/solver.py
@@ -12,6 +12,7 @@ __copyright__ = "Thomas Guillod - Dartmouth College"
 __license__ = "Mozilla Public License Version 2.0"
 
 import copy
+import scilogger
 from pypeec.lib_matrix import matrix_factorization
 from pypeec.lib_matrix import multiply_fft
 from pypeec.lib_matrix import fourier_transform
@@ -29,11 +30,10 @@ from pypeec.lib_check import check_data_problem
 from pypeec.lib_check import check_data_tolerance
 from pypeec.lib_check import check_data_solver
 from pypeec.lib_check import check_data_options
-from pypeec import log
 
 
 # get a logger
-LOGGER = log.get_logger(__name__, "pypeec")
+LOGGER = scilogger.get_logger(__name__, "pypeec")
 
 
 def _run_solver_options(data_solver):
@@ -74,7 +74,7 @@ def _run_solver_init(data_solver):
     _run_solver_options(data_solver)
 
     # get the voxel geometry and the incidence matrix
-    with log.BlockTimer(LOGGER, "voxel_geometry"):
+    with LOGGER.BlockTimer("voxel_geometry"):
         # get the coordinate of the voxels
         pts_vox = voxel_geometry.get_voxel_coordinate(n, d, c)
 
@@ -82,7 +82,7 @@ def _run_solver_init(data_solver):
         A_vox = voxel_geometry.get_incidence_matrix(n)
 
     # get the Green functions
-    with log.BlockTimer(LOGGER, "system_tensor"):
+    with LOGGER.BlockTimer("system_tensor"):
         # Green function self-coefficient
         G_self = system_tensor.get_green_self(d)
 
@@ -93,7 +93,7 @@ def _run_solver_init(data_solver):
         K_tsr = system_tensor.get_coupling_tensor(n, d, integral_simplify, has_coupling)
 
     # parse the problem geometry (materials and sources)
-    with log.BlockTimer(LOGGER, "problem_geometry"):
+    with LOGGER.BlockTimer("problem_geometry"):
         # parse the materials
         (idx_vc, idx_vm) = problem_geometry.get_material_indices(material_idx)
         material_pos = problem_geometry.get_material_pos(material_idx, idx_vc, idx_vm)
@@ -114,7 +114,7 @@ def _run_solver_init(data_solver):
         problem_status = problem_geometry.get_status(n, idx_vc, idx_vm, idx_fc, idx_fm, idx_src_c, idx_src_v)
 
     # get the system operators
-    with log.BlockTimer(LOGGER, "system_matrix"):
+    with LOGGER.BlockTimer("system_matrix"):
         # get the inductance tensor (preconditioner and full problem)
         (L_c, L_op_c) = system_matrix.get_inductance_matrix(n, d, idx_fc, G_self, G_mutual, has_electric, mult_type)
 
@@ -216,7 +216,7 @@ def _run_solver_sweep(data_solver, data_internal, data_param, sol_init):
     _run_solver_options(data_solver)
 
     # get the material and source values
-    with log.BlockTimer(LOGGER, "problem_value"):
+    with LOGGER.BlockTimer("problem_value"):
         # parse the material parameters
         (rho_vc, rho_vm) = problem_value.get_material_vector(material_val, material_idx)
 
@@ -230,7 +230,7 @@ def _run_solver_sweep(data_solver, data_internal, data_param, sol_init):
         R_m = problem_value.get_resistance_vector(n, d, A_net_m, idx_fm, rho_vm)
 
     # assemble the equation system
-    with log.BlockTimer(LOGGER, "equation_system"):
+    with LOGGER.BlockTimer("equation_system"):
         # compute the right-hand vector with the sources
[…]
