+++ examples/run_solver.py
@@ -8,7 +8,7 @@ __license__ = "Mozilla Public License Version 2.0"
 
 import sys
 import os.path
-from pypeec import main
+import pypeec
 import examples_config
 
 # get config
@@ -26,7 +26,7 @@ if __name__ == "__main__":
 
     # run
     try:
-        main.run_solver_file(file_voxel, file_problem, file_tolerance, file_solution)
+        pypeec.run_solver_file(file_voxel, file_problem, file_tolerance, file_solution)
     except Exception:
         sys.exit(1)
     else:
