+++ examples/run_plotter.py
@@ -8,7 +8,7 @@ __license__ = "Mozilla Public License Version 2.0"
 
 import sys
 import os.path
-from pypeec import main
+import pypeec
 import examples_config
 
 # get config
@@ -32,7 +32,7 @@ if __name__ == "__main__":
 
     # run
     try:
-        main.run_plotter_file(
+        pypeec.run_plotter_file(
             file_solution, file_plotter,
             tag_plot=tag_plot,
             plot_mode="qt",
