+++ pyproject.toml
@@ -77,7 +77,7 @@ file = "README.md"
 content-type = "text/markdown"
 
 [project.scripts]
-pypeec = "pypeec.script:run_script"
+pypeec = "pypeec:run_script"
 
 [tool.setuptools]
 packages = [
