+++ pypeec/data/logger.ini
@@ -1,80 +1,0 @@
-# PyPEEC default logger configuration file.
-# This file is written as a Python ConfigParser INI file.
-#
-# Thomas Guillod - Dartmouth College
-# Mozilla Public License Version 2.0
-
-######################################################################################################
-# Global logger options
-######################################################################################################
-[GLOBAL]
-######################################################################################################
-
-# logging module logger format with the following custom attributes
-#   - name: full name of the logger emitting the message (string)
-#   - tag: non-unique tag assigned to the logger (string)
-#   - timestamp: timestamp with the current time (string)
-#   - duration: elapsed time since log creation (string)
-#   - process_id: process identification number (integer)
-#   - thread_id: thread identification number (integer)
-FORMAT = %(duration)s : %(tag)-8s : %(levelname)-8s : %(message)s
-
-# integer with the number of spaces per indentation
-INDENTATION = 4
-
-# default level for all the handlers
-LEVEL_DEFAULT = DEBUG
-
-# display (or not) full exception traceback
-EXCEPTION_TRACE = true
-
-# use (or not) UTC times for the timestamp
-TIMESTAMP_UTC = false
-
-# formatting instructions for the timestamp (datetime/strftime format)
-TIMESTAMP_FMT = %m/%d/%Y|%H:%M:%S
-
-# formatting instructions for the duration (string format)
-#   - days: number of elapsed days (integer)
-#   - hours: number of elapsed hours (integer)
-#   - minutes: number of elapsed minutes (integer)
-#   - seconds: number of elapsed seconds (float)
-DURATION_FMT = {hours:02d}:{minutes:02d}:{seconds:05.2f}
-
-# use (or not) colors for the logs
-#   - if true: ANSI escape characters are used for setting the color
-#   - if false: ANSI escape characters are not used
-COLOR_USE = false
-
-# terminal ANSI escape characters for resetting color
-COLOR_RESET = \x1b[0m
-
-# terminal ANSI escape characters for the default color
-COLOR_DEFAULT = \x1b[38;5;15m
-
-######################################################################################################
-# Color definition for the different levels (ANSI escape characters)
-######################################################################################################
-[COLOR_LEVEL]
-######################################################################################################
-
-# keys-values dictionary
-#   - the keys are the logging levels
-#   - the values are the color codes
-DEBUG = \x1b[38;5;247m
-INFO = \x1b[38;5;15m
-WARNING = \x1b[38;5;202m
-ERROR = \x1b[38;5;196m
-CRITICAL = \x1b[38;5;196m
-
-######################################################################################################
-# Custom logging level for specific handler names
-######################################################################################################
-[MODULE_LEVEL]
-######################################################################################################
-
-# keys-values dictionary
-#   - the keys are the module names
-#   - the values are the logging levels
-pypeec.module_error = ERROR
-pypeec.module_info = INFO
