+++ docs/api/pypeec.log.rst
@@ -1,5 +1,0 @@
-﻿pypeec.log
-==========
-
-.. automodule:: pypeec.log
-    :members:
