+++ docs/api/pypeec.io.rst
@@ -1,5 +1,0 @@
-﻿pypeec.io
-=========
-
-.. automodule:: pypeec.io
-    :members:
