+++ docs/api/pypeec.script.rst
@@ -1,5 +1,0 @@
-﻿pypeec.script
-=============
-
-.. automodule:: pypeec.script
-    :members:
