+++ docs/api/pypeec.main.rst
@@ -1,5 +1,0 @@
-﻿pypeec.main
-===========
-
-.. automodule:: pypeec.main
-    :members:
