+++ docs/content/workflow.rst
@@ -7,8 +7,8 @@ PyPEEC Workflow
 .. figure:: ../images/workflow.png
 
    Description of the PyPEEC workflow (mesher, viewer, solver, and plotter).
-   The input and output files of the different tools are shown.
-   Additionally, a global configuration file is used for the logger.
+   The input files (geometry, problem, tolerance, plotter, and viewer) are either in JSON or YAML formats.
+   The output files (voxel and solution) are either in JSON or Pickle formats.
 
 PyPEEC Tools
 ------------
@@ -86,21 +86,13 @@ Input/Output File Description
 
 The following input files (JSON or YAML format) are used:
 
-* ``file_geometry`` (JSON or YAML, input file): description of the geometry
-* ``file_problem`` (JSON or YAML, input file): description of the magnetic problem
-* ``file_tolerance`` (JSON or YAML, input file): description of the solver numerical options
-* ``file_viewer`` (JSON or YAML, input file): options for the plots generated by the viewer
-* ``file_plotter`` (JSON or YAML, input file): options for the plots generated by the plotter
-* ``file_config`` (JSON or YAML, config file): global configuration file for PyPEEC
+* ``file_geometry`` (JSON or YAML, input/configuration file): description of the geometry
+* ``file_problem`` (JSON or YAML, input/configuration file): description of the magnetic problem
+* ``file_tolerance`` (JSON or YAML, input/configuration file): description of the solver numerical options
+* ``file_viewer`` (JSON or YAML, input/configuration file): options for the plots generated by the viewer
+* ``file_plotter`` (JSON or YAML, input/configuration file): options for the plots generated by the plotter
 
 The following files (JSON or Pickle format) are generated:
 
-* ``file_voxel`` (JSON or Pickle, output file): file with the definition of the voxel structure
-* ``file_solution`` (JSON or Pickle, output file): file with the solution of the magnetic problem
-
-The following file extensions are used:
-
-* ``yml`` and ``yaml`` for YAML files
-* ``js`` and ``json`` for JSON files
-* ``gz`` and ``gzip`` for GZIP/JSON files
-* ``pck`` and ``pkl`` for Pickle files
+* ``file_voxel`` (JSON or Pickle, output/data file): file with the definition of the voxel structure
+* ``file_solution`` (JSON or Pickle, output/data file): file with the solution of the magnetic problem
