### 2024-10-15 06b6c18
rst typo in documentation
