+++ docs/content/technical.rst
@@ -127,6 +127,7 @@ Please include a clear and concise description of what the bug is.
 Ideally, provide a minimal working example for the bug.
 
 Additionally, please report the following parameters:
+
 * The version of the PyPEEC you are using.
 * The platform/hardware you are using.
 * The version of Python and of the relevant dependencies.
