+++ conda_dev.yaml
@@ -27,8 +27,9 @@ dependencies:
   - sphinx_rtd_theme=1.2.2
   - sphinx=6.2.1
   - pkg-config=0.29.2
-  # YAML parsing library
-  - pyyaml=6.0
+  # logging and serialization
+  - scilogger=1.1.0
+  - scisave=1.1.0
   # numerical libraries
   - numpy=1.24.1
   - scipy=1.9.3
