+++ requirements.txt
@@ -11,7 +11,7 @@
 
 # logging and serialization
 scilogger~=1.1.0
-scisave~=1.1.0
+scisave~=1.2.0
 
 # numerical libraries
 numpy~=1.26.4
