+++ pyproject.toml
@@ -52,7 +52,7 @@ classifiers = [
 
 dependencies = [
     "scilogger >= 1.1",
-    "scisave >= 1.1",
+    "scisave >= 1.2",
     "joblib >= 1.3",
     "numpy >= 1.24, < 1.27",
     "scipy >= 1.9, < 1.14",
