+++ conda_base.yaml
@@ -18,7 +18,7 @@ dependencies:
   - python=3.10
   # logging and serialization
   - scilogger=1.1.0
-  - scisave=1.1.0
+  - scisave=1.2.0
   # numerical libraries
   - numpy=1.24.1
   - scipy=1.9.3
