+++ conda_dev.yaml
@@ -29,7 +29,7 @@ dependencies:
   - pkg-config=0.29.2
   # logging and serialization
   - scilogger=1.1.0
-  - scisave=1.1.0
+  - scisave=1.2.0
   # numerical libraries
   - numpy=1.24.1
   - scipy=1.9.3
