### 2024-10-16 fbbbadf
conda-forge as unique channel
