+++ conda_dev.yaml
@@ -12,9 +12,7 @@
 # Mozilla Public License Version 2.0
 
 name: pypeec
-channels:
-  - conda-forge
-  - defaults
+channels: [conda-forge]
 dependencies:
   # base installation
   - python=3.10
