+++ requirements.txt
@@ -12,6 +12,7 @@
 # logging and serialization
 scilogger~=1.1.0
 scisave~=1.2.0
+jsonschema~=4.23
 
 # numerical libraries
 numpy~=1.26.4
