+++ conda_base.yaml
@@ -17,6 +17,7 @@ dependencies:
   # logging and serialization
   - scilogger=1.1.0
   - scisave=1.2.0
+  - jsonschema=4.23.0
   # numerical libraries
   - numpy=1.24.1
   - scipy=1.9.3
