+++ pyproject.toml
@@ -53,6 +53,7 @@ classifiers = [
 dependencies = [
     "scilogger >= 1.1",
     "scisave >= 1.2",
+    "jsonschema >= 4.23",
     "joblib >= 1.3",
     "numpy >= 1.24, < 1.27",
     "scipy >= 1.9, < 1.14",
