+++ conda_dev.yaml
@@ -28,6 +28,7 @@ dependencies:
   # logging and serialization
   - scilogger=1.1.0
   - scisave=1.2.0
+  - jsonschema=4.23.0
   # numerical libraries
   - numpy=1.24.1
   - scipy=1.9.3
