### 2024-10-17 096b3ea
jsonschema for data_geometry
