+++ pypeec/data/schema_geometry_stl.yaml
@@ -1,0 +1,56 @@
+"type": "object"
+"required":
+    - "param"
+    - "domain_stl"
+"properties":
+    "param":
+        "type": "object"
+        "required":
+            - "d"
+            - "xyz_min"
+            - "xyz_max"
+        "properties":
+            "d":
+                "type": "array"
+                "minItems": 3
+                "maxItems": 3
+                "items":
+                    "type": "number"
+                    "minimum": 0
+            "xyz_min":
+                "type": ["null", "array"]
+                "minItems": 3
+                "maxItems": 3
+                "items":
+                    "type": "number"
+            "xyz_max":
+                "type": ["null", "array"]
+                "minItems": 3
+                "maxItems": 3
+                "items":
+                    "type": "number"
+    "domain_stl":
+        "type": "object"
+        "minProperties": 1
+        "additionalProperties":
+            "type": "array"
+            "minItems": 1
+            "items":
+                "type": "object"
+                "required":
+                    - "scale"
+                    - "offset"
+                    - "filename"
+                "properties":
+                    "scale":
+                        "type": "number"
+                        "minimum": 0
+                    "offset":
+                        "type": "array"
+                        "minItems": 3
+                        "maxItems": 3
+                        "items":
+                            "type": "number"
+                    "filename":
+                        "type": "string"
+                        "minLength": 1
