### 2024-10-17 bb457ab
checks for domain name
