### 2024-10-17 e6ac3bb
check solver and mesher data
