+++ pypeec/lib_mesher/mesher_shape.py
@@ -321,8 +321,7 @@ def _get_layer_stack(layer_stack, dz, cz):
     stack_pos *= dz
 
     # add offset
-    if cz is not None:
-        stack_pos += cz
+    stack_pos += cz
 
     return stack_pos, stack_idx, stack_tag
 
