### 2024-10-17 1513f2f
clean schema
