### 2024-10-17 d969554
better solver sweep
