+++ examples/tutorial/problem.yaml
@@ -32,15 +32,16 @@
 "source_val": &source_val
     "src": {"V_re": +1.0, "V_im": 0.0, "Z_re": 1.0e-3, "Z_im": 0.0}
     "sink": {"V_re": -1.0, "V_im": 0.0, "Z_re": 1.0e-3, "Z_im": 0.0}
-"sweep_config":
-    "sim_dc": null
-    "sim_ac": "sim_dc"
-"sweep_param":
+"sweep_solver":
     "sim_dc":
-        "freq": 0.0
-        "material_val": *material_val
-        "source_val": *source_val
+        "init": null
+        "param":
+            "freq": 0.0
+            "material_val": *material_val
+            "source_val": *source_val
     "sim_ac":
-        "freq": 1.0e+6
-        "material_val": *material_val
-        "source_val": *source_val
+        "init": "sim_dc"
+        "param":
+            "freq": 1.0e+6
+            "material_val": *material_val
+            "source_val": *source_val
