### 2024-10-17 9cd1102
solver data structure
