+++ pypeec/lib_check/check_data_solver.py
@@ -1,277 +1,0 @@
-"""
-Combine the voxel and problem data into a new solver data structure.
-"""
-
-__author__ = "Thomas Guillod"
-__copyright__ = "Thomas Guillod - Dartmouth College"
-__license__ = "Mozilla Public License Version 2.0"
-
-import numpy as np
-from pypeec.lib_check import datachecker
-
-
-def _check_indices(idx_c, idx_m, idx_s):
-    """
-    Check that the material and source indices are valid.
-    The indices should be unique and compatible with the voxel size.
-    The source indices should be included in the electric indices.
-    Check that there is at least one source per connected electrical.
-    """
-
-    # check the indices
-    datachecker.check_index_array("index electric", idx_c, can_be_empty=False)
-    datachecker.check_index_array("index magnetic", idx_m, can_be_empty=True)
-    datachecker.check_index_array("index source", idx_s, can_be_empty=False)
-
-    # check that the terminal indices are electric indices
-    cond = np.all(np.in1d(idx_s, idx_c))
-    datachecker.check_assert("index", cond, "source indices should overlap with electric indices")
-
-
-def _check_source_graph(idx_c, idx_s, graph_def):
-    """
-    Check that there is at least one source per connected component.
-    A connected components without a source would lead to a singular problem.
-    """
-
-    for idx_graph in graph_def:
-        has_electric = len(np.intersect1d(idx_graph, idx_c)) > 0
-        has_source = len(np.intersect1d(idx_graph, idx_s)) > 0
-
-        cond = (not has_electric) or has_source
-        datachecker.check_assert("index", cond, "electric components should include at least one source")
-
-
-def _get_field(val_dict, idx, var_type, orientation_type):
-    """
-    Cast and check the material and source vectors.
-    If the variable is a scalar, cast to an array.
-    If the variable is an array, check the length.
-    """
-
-    for tag, val in val_dict.items():
-        # cast
-        if var_type == "lumped" and orientation_type is None:
-            val = np.full(len(idx), val, dtype=np.float_)
-        elif var_type == "lumped" and orientation_type == "isotropic":
-            val = np.full((len(idx), 3), val, dtype=np.float_)
-        elif var_type == "lumped" and orientation_type == "anisotropic":
-            val = np.array(val, dtype=np.float_)
-            val = np.tile(val, (len(idx), 1))
-        elif var_type == "distributed" and orientation_type is None:
-            val = np.array(val, dtype=np.float_)
-        elif var_type == "distributed" and orientation_type == "isotropic":
-            val = np.array(val, dtype=np.float_)
-            val = np.tile(val, (3, 1)).transpose()
-        elif var_type == "distributed" and orientation_type == "anisotropic":
-            val = np.array(val, dtype=np.float_)
-
-        # check
-        cond = len(val) == len(idx)
-        datachecker.check_assert(tag, cond, "vector length does not match the number of voxels")
-
-        # update
-        val_dict[tag] = val
-
-    return val_dict
-
-
-def _get_domain_indices(domain_list, domain_def):
-    """
-    Get indices from a list of domain names.
-    """
-
-    idx_all = np.empty(0, dtype=np.int_)
-    for tag in domain_list:
-        idx_tmp = np.array(domain_def[tag], dtype=np.int_)
-        idx_all = np.append(idx_all, idx_tmp)
-
-    return idx_all
-
-
-def _get_material_idx(material_def, domain_def):
-    """
-    Get the indices of the material.
-    Create a new dict with the indices in place of the domain names.
-    Split the electric and magnetic materials.
-    """
-
-    # init
-    material_idx = {}
-    idx_c = np.empty(0, dtype=np.int_)
[…]
