+++ pypeec/lib_solver/problem_geometry.py
@@ -14,30 +14,153 @@ import numpy as np
 LOGGER = scilogger.get_logger(__name__, "pypeec")
 
 
-def get_material_indices(material_idx):
+def _get_domain_indices(domain_list, domain_def):
     """
-    Get the indices of the voxels for the different material types.
+    Get indices from a list of domain names.
     """
 
-    # array for the indices
+    idx_all = np.empty(0, dtype=np.int_)
+    for tag in domain_list:
+        idx_tmp = np.array(domain_def[tag], dtype=np.int_)
+        idx_all = np.append(idx_all, idx_tmp)
+
+    return idx_all
+
+
+def _check_indices(idx_vc, idx_src_c, idx_src_v):
+    """
+    Check that the material and source indices are valid.
+    The indices should be unique and compatible with the voxel size.
+    The source indices should be included in the electric indices.
+    Check that there is at least one source per connected electrical.
+    """
+
+    # check the indices
+    if len(idx_vc) == 0:
+        raise RuntimeError("electric indices cannot be empty")
+
+    if (len(idx_src_c) == 0) and (len(idx_src_v) == 0):
+        raise RuntimeError("source indices cannot be empty")
+
+    # check that the terminal indices are electric indices
+    if not np.all(np.in1d(idx_src_c, idx_vc)):
+        raise RuntimeError("current source indices should overlap with electric indices")
+    if not np.all(np.in1d(idx_src_v, idx_vc)):
+        raise RuntimeError("voltage source indices should overlap with electric indices")
+
+
+def _check_source_graph(idx_vc, idx_src_c, idx_src_v, graph_def):
+    """
+    Check that there is at least one source per connected component.
+    A connected components without a source would lead to a singular problem.
+    """
+
+    for idx_graph in graph_def:
+        has_electric = len(np.intersect1d(idx_graph, idx_vc)) > 0
+        has_current_source = len(np.intersect1d(idx_graph, idx_src_c)) > 0
+        has_voltage_source = len(np.intersect1d(idx_graph, idx_src_v)) > 0
+
+        if has_electric:
+            if not (has_current_source or has_voltage_source):
+                raise RuntimeError("electric components should include at least one source")
+
+
+def get_problem_type(idx_vc, idx_vm, idx_src_c, idx_src_v, graph_def):
+
+    # check the existence of magnetic domains
+    has_electric = len(idx_vc) > 0
+    has_magnetic = len(idx_vm) > 0
+    has_coupling = has_electric and has_magnetic
+
+    # check voxel indices
+    _check_indices(idx_vc, idx_src_c, idx_src_v)
+    _check_source_graph(idx_vc, idx_src_c, idx_src_v, graph_def)
+
+    # check the existence of magnetic domains
+    has_electric = len(idx_vc) > 0
+    has_magnetic = len(idx_vm) > 0
+
+    return  has_electric, has_magnetic
+
+
+def get_material_idx(material_def, domain_def):
+    """
+    Get the indices of the material.
+    Create a new dict with the indices in place of the domain names.
+    Split the electric and magnetic materials.
+    """
+
+    # init
+    material_idx = {}
     idx_vc = np.empty(0, dtype=np.int_)
     idx_vm = np.empty(0, dtype=np.int_)
 
-    # populate the arrays
-    for tag, material_idx_tmp in material_idx.items():
+    for tag, material_def_tmp in material_def.items():
         # extract the data
-        material_type = material_idx_tmp["material_type"]
-        idx = material_idx_tmp["idx"]
+        var_type = material_def_tmp["var_type"]
[…]
