### 2024-10-17 aa333b3
solver data structure
