### 2024-10-17 bd619de
solver data structure
