+++ examples/examples_shape/coplanar/problem.yaml
@@ -36,10 +36,10 @@
     "pri_sink": {"V_re": 0.0, "V_im": 0.0, "Z_re": 1.0e-3, "Z_im": 0.0}
     "sec_src": {"V_re": -1.0, "V_im": 0.0, "Z_re": 1.0e-3, "Z_im": 0.0}
     "sec_sink": {"V_re": 0.0, "V_im": 0.0, "Z_re": 1.0e-3, "Z_im": 0.0}
-"sweep_config":
-    "sim_default": null
-"sweep_param":
+"sweep_solver":
     "sim_default":
-        "freq": 1.0e+6
-        "material_val": *material_val
-        "source_val": *source_val
+        "init": null
+        "param":
+            "freq": 1.0e+6
+            "material_val": *material_val
+            "source_val": *source_val
