+++ examples/examples_stl/inductor_air/problem.yaml
@@ -26,10 +26,10 @@
 "source_val": &source_val
     "src": {"V_re": +1.0, "V_im": 0.0, "Z_re": 1.0e-3, "Z_im": 0.0}
     "sink": {"V_re": -1.0, "V_im": 0.0, "Z_re": 1.0e-3, "Z_im": 0.0}
-"sweep_config":
-    "sim_default": null
-"sweep_param":
+"sweep_solver":
     "sim_default":
-        "freq": 0.0
-        "material_val": *material_val
-        "source_val": *source_val
+        "init": null
+        "param":
+            "freq": 0.0
+            "material_val": *material_val
+            "source_val": *source_val
