+++ examples/examples_voxel/transformer/problem.yaml
@@ -37,15 +37,16 @@
     "pri_src": {"I_re": 1.0, "I_im": 0.0, "Y_re": 0.0, "Y_im": 0.0}
     "pri_sink": {"V_re": 0.0, "V_im": 0.0, "Z_re": 0.0, "Z_im": 0.0}
     "sec_short": {"V_re": 0.0, "V_im": 0.0, "Z_re": 0.0, "Z_im": 0.0}
-"sweep_config":
-    "sim_dc": null
-    "sim_ac": "sim_dc"
-"sweep_param":
+"sweep_solver":
     "sim_dc":
-        "freq": 0.0
-        "material_val": *material_val
-        "source_val": *source_val
+        "init": null
+        "param":
+            "freq": 0.0
+            "material_val": *material_val
+            "source_val": *source_val
     "sim_ac":
-        "freq": 10.0e+6
-        "material_val": *material_val
-        "source_val": *source_val
+        "init": "sim_dc"
+        "param":
+            "freq": 10.0e+6
+            "material_val": *material_val
+            "source_val": *source_val
