+++ examples/examples_voxel/core/problem.yaml
@@ -32,7 +32,7 @@
 "source_val": &source_val
     "src": {"I_re": 1.0, "I_im": 0.0, "Y_re": 0.0, "Y_im": 0.0}
     "sink": {"V_re": 0.0, "V_im": 0.0, "Z_re": 0.0, "Z_im": 0.0}
-"sweep_param":
+"sweep_solver":
     "sim_dc":
         "init": null
         "data_param":
@@ -40,7 +40,7 @@
             "material_val": *material_val
             "source_val": *source_val
     "sim_ac":
-        "init": "sim_ac"
+        "init": "sim_dc"
         "data_param":
             "freq": 1.0e+3
             "material_val": *material_val

[…]
