+++ pypeec/lib_check/check_data_format.py
@@ -14,12 +14,21 @@ import scisave
 # preload the schemas
 with importlib.resources.path("pypeec.data", "schema_geometry.yaml") as file:
     SCHEMA_GEOMETRY = scisave.load_config(file)
+with importlib.resources.path("pypeec.data", "schema_problem.yaml") as file:
+    SCHEMA_PROBLEM = scisave.load_config(file)
 
 
 def check_data_geometry(data_geometry):
     """
-    Check the mesher data type and extract the data.
+    Check the mesher geometry data.
     """
 
     jsonschema.validate(instance=data_geometry, schema=SCHEMA_GEOMETRY)
 
+
+def check_data_problem(data_geometry):
+    """
+    Check the solver problem data.
+    """
+
+    jsonschema.validate(instance=data_geometry, schema=SCHEMA_PROBLEM)
