+++ pypeec/lib_check/check_data_problem.py
@@ -1,194 +1,0 @@
-"""
-Module for checking the solver problem data.
-"""
-
-__author__ = "Thomas Guillod"
-__copyright__ = "Thomas Guillod - Dartmouth College"
-__license__ = "Mozilla Public License Version 2.0"
-
-from pypeec.lib_check import datachecker
-
-
-def _check_field(val_dict, key_list, var_type, orientation_type):
-    """
-    Check a value dict for a variable (lumped / distributed / isotropic / anisotropic).
-    """
-
-    # check type
-    datachecker.check_dict("val", val_dict, key_list=key_list)
-
-    # check data
-    for tag, val in val_dict.items():
-        if var_type == "lumped" and orientation_type is None:
-            datachecker.check_float(tag, val)
-        elif var_type == "lumped" and orientation_type == "isotropic":
-            datachecker.check_float(tag, val)
-        elif var_type == "lumped" and orientation_type == "anisotropic":
-            datachecker.check_float_array(tag, val, size=3)
-        elif var_type == "distributed" and orientation_type is None:
-            datachecker.check_float_array(tag, val)
-        elif var_type == "distributed" and orientation_type == "isotropic":
-            datachecker.check_float_array(tag, val)
-        elif var_type == "distributed" and orientation_type == "anisotropic":
-            datachecker.check_float_pts(tag, val, size=3)
-
-
-def _check_material_def(material_def):
-    """
-    Check that the material definition is valid.
-    """
-
-    # check type
-    datachecker.check_dict("material_def", material_def, can_be_empty=False)
-
-    # check value
-    for material_def_tmp in material_def.values():
-        # check type
-        key_list = ["material_type", "orientation_type", "var_type", "domain_list"]
-        datachecker.check_dict("material_def", material_def_tmp, key_list=key_list)
-
-        # extract the data
-        var_type = material_def_tmp["var_type"]
-        material_type = material_def_tmp["material_type"]
-        orientation_type = material_def_tmp["orientation_type"]
-        domain_list = material_def_tmp["domain_list"]
-
-        # check data
-        var_type_list = ["lumped", "distributed"]
-        orientation_type_list = ["isotropic", "anisotropic"]
-        material_type_list = ["electric", "magnetic", "electromagnetic"]
-        datachecker.check_choice("var_type", var_type, var_type_list)
-        datachecker.check_choice("material_type", material_type, material_type_list)
-        datachecker.check_choice("orientation_type", orientation_type, orientation_type_list)
-        datachecker.check_list("domain_list", domain_list, can_be_empty=False, sub_type=str)
-
-
-def _check_source_def(source_def):
-    """
-    Check that the source definition is valid.
-    """
-
-    # check type
-    datachecker.check_dict("source_def", source_def, can_be_empty=False)
-
-    # check value
-    for source_def_tmp in source_def.values():
-        # check type
-        key_list = ["source_type", "var_type", "domain_list"]
-        datachecker.check_dict("source_def", source_def_tmp, key_list=key_list)
-
-        # extract the data
-        var_type = source_def_tmp["var_type"]
-        source_type = source_def_tmp["source_type"]
-        domain_list = source_def_tmp["domain_list"]
-
-        # check data
-        source_type_list = ["current", "voltage"]
-        var_type_list = ["lumped", "distributed"]
-        datachecker.check_choice("var_type", var_type, var_type_list)
-        datachecker.check_choice("source_type", source_type, source_type_list)
-        datachecker.check_list("domain_list", domain_list, can_be_empty=False, sub_type=str)
-
-
-def _check_sweep_param(sweep_param, material_def, source_def):
-    """
[…]
