### 2024-10-17 eb653b2
tolerance data json schema
