+++ pypeec/lib_check/check_data_tolerance.py
@@ -1,243 +1,0 @@
-"""
-Module for checking the solver tolerance data.
-"""
-
-__author__ = "Thomas Guillod"
-__copyright__ = "Thomas Guillod - Dartmouth College"
-__license__ = "Mozilla Public License Version 2.0"
-
-from pypeec.lib_check import datachecker
-
-
-def _check_parallel_sweep(parallel_sweep):
-    """
-    Check the parallel sweep options.
-    """
-
-    # check type
-    key_list = [
-        "n_jobs",
-        "n_threads",
-    ]
-    datachecker.check_dict("parallel_sweep", parallel_sweep, key_list=key_list)
-
-    # extract field
-    n_jobs = parallel_sweep["n_jobs"]
-    n_threads = parallel_sweep["n_threads"]
-
-    # check the data
-    datachecker.check_integer("n_jobs", n_jobs, is_positive=False, can_be_zero=True, can_be_none=False)
-    datachecker.check_integer("n_threads", n_threads, is_positive=True, can_be_zero=False, can_be_none=True)
-
-
-def _check_fft_options(fft_options):
-    """
-    Check the FFT algorithm options.
-    """
-
-    # check type
-    key_list = [
-        "matrix_split",
-        "library",
-        "scipy_worker",
-        "fftw_thread",
-        "fftw_timeout",
-        "fftw_byte_align",
-    ]
-    datachecker.check_dict("fft_options", fft_options, key_list=key_list)
-
-    # extract field
-    matrix_split = fft_options["matrix_split"]
-    library = fft_options["library"]
-    scipy_worker = fft_options["scipy_worker"]
-    fftw_thread = fft_options["fftw_thread"]
-    fftw_timeout = fft_options["fftw_timeout"]
-    fftw_byte_align = fft_options["fftw_byte_align"]
-
-    # check the data
-    datachecker.check_boolean("matrix_split", matrix_split)
-    datachecker.check_choice("library", library, ["SciPy", "FFTW", "CuPy"])
-    datachecker.check_integer("scipy_worker", scipy_worker)
-    datachecker.check_integer("fftw_thread", fftw_thread)
-    datachecker.check_integer("fftw_byte_align", fftw_byte_align, is_positive=True, can_be_zero=False, can_be_none=True)
-    datachecker.check_float("fftw_timeout", fftw_timeout, is_positive=True, can_be_zero=True, can_be_none=True)
-
-
-def _check_factorization_options(factorization_options):
-    """
-    Check the matrix factorization options.
-    """
-
-    # check type
-    key_list = ["library", "pyamg_options", "pardiso_options"]
-    datachecker.check_dict("factorization_options", factorization_options, key_list=key_list)
-
-    # extract field
-    library = factorization_options["library"]
-    pyamg_options = factorization_options["pyamg_options"]
-    pardiso_options = factorization_options["pardiso_options"]
-
-    # check the data
-    datachecker.check_choice("library", library, ["SuperLU", "UMFPACK", "PARDISO", "PyAMG"])
-
-    # check the data
-    key_list = ["tol", "solver", "krylov"]
-    datachecker.check_dict("pyamg_options", pyamg_options, key_list=key_list)
-    datachecker.check_choice("solver", pyamg_options["solver"], ["adapt", "root"])
-    datachecker.check_choice("krylov", pyamg_options["krylov"], ["gmres", "fgmres"], can_be_none=True)
-    datachecker.check_float("tol", pyamg_options["tol"], is_positive=True, can_be_zero=False)
-
-    # check the data
-    key_list = ["thread_pardiso", "thread_mkl"]
-    datachecker.check_dict("pardiso_options", pardiso_options, key_list=key_list)
[…]
