+++ pypeec/data/schema_tolerance.yaml
@@ -1,0 +1,235 @@
+#############################################################################
+"iter_options": &iter_options
+    "type": "object"
+    "required":
+        - "solver"
+        - "n_inner"
+        - "n_outer"
+        - "rel_tol"
+        - "abs_tol"
+    "properties":
+        "solver":
+            "type": "string"
+            "enum": [ "gmres", "gcrot" ]
+        "n_inner":
+            "type": "integer"
+            "minimum": 0
+        "n_outer":
+            "type": "integer"
+            "minimum": 0
+        "rel_tol":
+            "type": "number"
+            "minimum": 0
+        "abs_tol":
+            "type": "number"
+            "minimum": 0
+
+#############################################################################
+"segregated_options": &segregated_options
+    "type": "object"
+    "required":
+        - "relax_electric"
+        - "relax_magnetic"
+        - "n_min"
+        - "n_max"
+        - "rel_tol"
+        - "abs_tol"
+    "properties":
+        "relax_electric":
+            "type": "number"
+            "minimum": 0
+        "relax_magnetic":
+            "type": "number"
+            "minimum": 0
+        "n_min":
+            "type": "integer"
+            "minimum": 0
+        "n_max":
+            "type": "integer"
+            "minimum": 0
+        "rel_tol":
+            "type": "number"
+            "minimum": 0
+        "abs_tol":
+            "type": "number"
+            "minimum": 0
+        "iter_electric_options": *iter_options
+        "iter_magnetic_options": *iter_options
+
+#############################################################################
+"tolerance_options": &tolerance_options
+    "type": "object"
+    "required":
+        - "ignore_status"
+        - "ignore_res"
+        - "rel_tol"
+        - "abs_tol"
+    "properties":
+        "ignore_status":
+            "type": "boolean"
+        "ignore_res":
+            "type": "boolean"
+        "rel_tol":
+            "type": "number"
+            "minimum": 0
+        "abs_tol":
+            "type": "number"
+            "minimum": 0
+
+#############################################################################
+"power_options": &power_options
+    "type": "object"
+    "required":
+        - "stop"
+        - "n_min"
+        - "n_cmp"
+        - "rel_tol"
+        - "abs_tol"
+    "properties":
+        "stop":
+            "type": "boolean"
+        "n_min":
+            "type": "integer"
+            "minimum": 0
+        "n_cmp":
+            "type": "integer"
+            "minimum": 0
+        "rel_tol":
+            "type": "number"
+            "minimum": 0
+        "abs_tol":
+            "type": "number"
+            "minimum": 0
+
+#############################################################################
+"type": "object"
+"required":
+    - "parallel_sweep"
+    - "integral_simplify"
+    - "mult_type"
+    - "fft_options"
+    - "factorization_options"
+    - "solver_options"
+    - "condition_options"
+"properties":
+    "parallel_sweep":
+        "type": "object"
+        "required":
+            - "n_jobs"
+            - "n_threads"
+        "properties":
+            "n_jobs":
+                "type": "integer"
+            "n_threads":
+                "type": [ "null", "integer" ]
+    "integral_simplify":
+        "type": "number"
+        "minimum": 0
+    "mult_type":
+        "type": "string"
+        "enum": [ "fft", "direct" ]
+    "fft_options":
+        "type": "object"
+        "required":
+            - "matrix_split"
+            - "library"
+            - "scipy_worker"
+            - "fftw_thread"
+            - "fftw_timeout"
+            - "fftw_byte_align"
+        "properties":
+            "matrix_split":
+                "type": "boolean"
+            "library":
+                "type": "string"
+                "enum": [ "SciPy", "FFTW", "CuPy" ]
+            "scipy_worker":
[…]
