### 2024-10-17 b3982f4
tolerance data json schema
