### 2024-10-17 7c503dd
electromagnetic color
