### 2024-10-17 3dbd155
viewer and plotter json schema
