+++ pypeec/lib_check/datachecker.py
@@ -1,319 +1,0 @@
-"""
-Module for checking the data type and content.
-"""
-
-__author__ = "Thomas Guillod"
-__copyright__ = "Thomas Guillod - Dartmouth College"
-__license__ = "Mozilla Public License Version 2.0"
-
-import os.path
-import numpy as np
-
-
-def check_dict(name, data, key_list=None, sub_type=None, can_be_empty=True, can_be_none=False):
-    """
-    Check a dict.
-    """
-
-    # check if none
-    if can_be_none and data is None:
-        return
-
-    # check type
-    if not isinstance(data, dict):
-        raise AssertionError("%s: should be a dict" % name)
-    if (not can_be_empty) and (len(data) == 0):
-        raise AssertionError("%s: should not be empty" % name)
-
-    # check value
-    for tag, value in data.items():
-        if not isinstance(tag, str):
-            raise AssertionError("%s: all the keys should be strings" % name)
-        if len(tag) == 0:
-            raise AssertionError("%s: keys cannot be empty" % name)
-        if (sub_type is not None) and not isinstance(value, sub_type):
-            raise AssertionError("%s: invalid type" % name)
-    if key_list is not None:
-        for tag in key_list:
-            if tag not in data:
-                raise AssertionError("%s: dict is incomplete: %s" % (name, tag))
-
-
-def check_list(name, data, sub_type=None, can_be_empty=True, can_be_none=False):
-    """
-    Check a list.
-    """
-
-    # check if none
-    if can_be_none and data is None:
-        return
-
-    # check type
-    if not isinstance(data, list):
-        raise AssertionError("%s: should be a list" % name)
-    if (not can_be_empty) and (len(data) == 0):
-        raise AssertionError("%s: should not be empty" % name)
-
-    # check value
-    for value in data:
-        if (sub_type is not None) and not isinstance(value, sub_type):
-            raise AssertionError("%s: invalid type" % name)
-
-
-def check_float_pts(name, data, size=None, is_positive=False, can_be_zero=True, can_be_empty=True, can_be_none=False):
-    """
-    Check a 2D float array.
-    """
-
-    # check if none
-    if can_be_none and data is None:
-        return
-
-    # cast to array
-    data = np.array(data)
-
-    # check type and value
-    if (not can_be_empty) and (len(data) == 0):
-        raise AssertionError("%s: cannot be empty" % name)
-    for data_tmp in data:
-        check_float_array(name, data_tmp, size, is_positive, can_be_zero, can_be_empty)
-
-
-def check_integer_pts(name, data, size=None, is_positive=False, can_be_zero=True, can_be_empty=True, can_be_none=False):
-    """
-    Check a 2D integer array.
-    """
-
-    # check if none
-    if can_be_none and data is None:
-        return
-
-    # cast to array
-    data = np.array(data)
-
-    # check type and value
-    if (not can_be_empty) and (len(data) == 0):
-        raise AssertionError("%s: cannot be empty" % name)
-    for data_tmp in data:
-        check_integer_array(name, data_tmp, size, is_positive, can_be_zero, can_be_empty)
-
-
-def check_float_array(name, data, size=None, is_positive=False, can_be_zero=True, can_be_empty=True, can_be_none=False):
-    """
-    Check a float array.
-    """
-
-    # check if none
-    if can_be_none and data is None:
-        return
-
-    # cast to array
-    data = np.array(data)
-
-    # check type and value
-    if len(data.shape) != 1:
-        raise AssertionError("%s: invalid shape" % name)
-    if size is not None:
-        if not (len(data) == size):
-            raise AssertionError("%s: invalid array size" % name)
-    if (not can_be_empty) and (len(data) == 0):
-        raise AssertionError("%s: cannot be empty" % name)
[…]
