+++ pypeec/lib_check/check_data_visualization.py
@@ -1,352 +1,0 @@
-"""
-Module for checking the viewer and plotter data.
-"""
-
-__author__ = "Thomas Guillod"
-__copyright__ = "Thomas Guillod - Dartmouth College"
-__license__ = "Mozilla Public License Version 2.0"
-
-from pypeec.lib_check import datachecker
-
-
-def _check_data_window(data_window):
-    """
-    Check the plot window options (for the viewer and plotter).
-    Check the window size and menu configuration.
-    """
-
-    # check type
-    key_list = ["show_menu", "image_size", "window_size", "notebook_size"]
-    datachecker.check_dict("data_window", data_window, key_list=key_list)
-
-    # extract the data
-    show_menu = data_window["show_menu"]
-    image_size = data_window["image_size"]
-    window_size = data_window["window_size"]
-    notebook_size = data_window["notebook_size"]
-
-    # check data
-    datachecker.check_boolean("show_menu", show_menu)
-    datachecker.check_integer_array("image_size", image_size, size=2, is_positive=True, can_be_zero=False, can_be_none=True)
-    datachecker.check_integer_array("window_size", window_size, size=2, is_positive=True, can_be_zero=False, can_be_none=True)
-    datachecker.check_integer_array("notebook_size", notebook_size, size=2, is_positive=True, can_be_zero=False, can_be_none=True)
-
-
-def _check_data_options_matplotlib(data_options):
-    """
-    Check the plot options (for Matplotlib).
-    """
-
-    # check type
-    key_list = ["style", "legend", "font"]
-    datachecker.check_dict("data_options", data_options, key_list=key_list)
-
-    # extract the data
-    style = data_options["style"]
-    legend = data_options["legend"]
-    font = data_options["font"]
-
-    # check data
-    datachecker.check_string("style", style, can_be_empty=False)
-    datachecker.check_string("legend", legend, can_be_empty=False)
-    datachecker.check_integer("font", font, is_positive=True, can_be_zero=False)
-
-
-def _check_data_options_pyvista(data_options):
-    """
-    Check the plot options (for PyVista).
-    """
-
-    # check type
-    key_list = ["plot_clip", "plot_view", "plot_theme"]
-    datachecker.check_dict("data_options", data_options, key_list=key_list)
-
-    # extract the data
-    plot_clip = data_options["plot_clip"]
-    plot_view = data_options["plot_view"]
-    plot_theme = data_options["plot_theme"]
-
-    _check_plot_clip(plot_clip)
-    _check_plot_view(plot_view)
-    _check_plot_theme(plot_theme)
-
-
-def _check_plot_view(plot_view):
-    """
-    Check the validity of the plot view (for PyVista).
-    """
-
-    # check type
-    key_list = [
-        "grid_plot", "grid_thickness", "grid_color", "grid_opacity",
-        "geom_plot", "geom_thickness", "geom_color", "geom_opacity",
-        "point_plot", "point_size", "point_color", "point_opacity",
-        "camera_roll", "camera_azimuth", "camera_elevation",
-    ]
-    datachecker.check_dict("data_window", plot_view, key_list=key_list)
-
-    # check grid data
-    datachecker.check_boolean("grid_plot", plot_view["grid_plot"])
-    datachecker.check_string("grid_color", plot_view["grid_color"], can_be_empty=False)
-    datachecker.check_float("grid_thickness", plot_view["grid_thickness"], is_positive=True, can_be_zero=True)
-    datachecker.check_float("grid_opacity", plot_view["grid_opacity"], is_positive=True, can_be_zero=True)
-
-    # check geom data
-    datachecker.check_boolean("geom_plot", plot_view["geom_plot"])
-    datachecker.check_string("geom_color", plot_view["geom_color"], can_be_empty=False)
[…]
