+++ pypeec/run/plotter.py
@@ -28,7 +28,7 @@ from pypeec.lib_visualization import parse_voxel
 from pypeec.lib_visualization import manage_pyvista
 from pypeec.lib_visualization import manage_matplotlib
 from pypeec.lib_visualization import manage_plotgui
-from pypeec.lib_check import check_data_visualization
+from pypeec.lib_check import check_data_format
 from pypeec.lib_check import check_data_options
 
 # get a logger
@@ -156,7 +156,7 @@ def run(
 
     # check the input data
     LOGGER.info("check the input data")
-    check_data_visualization.check_data_plotter(data_plotter)
+    check_data_format.check_data_plotter(data_plotter)
     check_data_options.check_plot_options(plot_mode, folder, name)
     check_data_options.check_tag_list(data_sweep, tag_sweep)
     check_data_options.check_tag_list(data_plotter, tag_plot)
