### 2024-10-17 508caa7
clean the json schema (indent)
