### 2024-10-17 4f99db7
current and voltage lumped source scaling
