### 2024-10-17 0f592bb
cz can be None for shape mesher
