+++ docs/format/file_geometry_shape.yaml
@@ -7,7 +7,7 @@
     #   - dx: voxel dimension for the x direction
     #   - dy: voxel dimension for the y direction
     #   - dz: voxel dimension for the z direction
-    #   - cz: coordinates of the voxel structure center for the z direction
+    #   - cz: coordinates of the voxel structure center for the z direction  (null for centering)
     #   - simplify: tolerance for simplifying the shapes (null for disabling simplification)
     #   - construct: grid size for constructing the shapes (null for default grid size)
     #   - xy_min: array with the lower corner coordinates of the voxel structure (x and y directions)
