### 2024-10-17 fd78304
JSON strict validator with numpy support
