+++ pypeec/lib_check/check_data_options.py
@@ -1,5 +1,8 @@
 """
-Module for the checking the data formats.
+Module for the checking the data format:
+    - check the user provided options
+    - check the data generated by the mesher
+    - check the data generated by the solver
 """
 
 __author__ = "Thomas Guillod"
