+++ pypeec/data/schema_tolerance.yaml
@@ -258,8 +258,8 @@
                     - "n_iter_max"
                 "properties":
                     "t_accuracy":
-                        "type": "number"
+                        "type": "integer"
                         "minimum": 0
                     "n_iter_max":
-                        "type": "number"
+                        "type": "integer"
                         "minimum": 0
