### 2024-10-17 6edc8b1
JSON strict validator with numpy support
