+++ .github/workflows/run_macos.yml
@@ -1,0 +1,47 @@
+# Run the "Platform tests" with the following conditions.
+#   - dispatch
+#   - weekly
+#
+# The tests are run on the following systems:
+#   - Linux / Ubuntu
+#   - Apple / macOS
+#   - Microsoft / Windows
+#
+# The tests are run on the following interpreter versions:
+#   - Python 3.9
+#   - Python 3.10
+#   - Python 3.11
+#   - Python 3.12
+#
+# Thomas Guillod - Dartmouth College
+# Mozilla Public License Version 2.0
+
+name: MacOS
+
+on:
+  push:
+  pull_request:
+  workflow_dispatch:
+
+jobs:
+  # ===================== run the tests with different systems
+  system:
+    strategy:
+      fail-fast: false
+      matrix:
+        system:
+          - "macos-latest"
+    name: ${{ matrix.system }}
+    runs-on: ${{ matrix.system }}
+    timeout-minutes: 20
+    steps:
+      - uses: actions/checkout@v4
+      - name: Set up Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: "3.10"
+      - name: Install dependencies
+        shell: bash -el {0}
+        run: |
+          python -m pip install --upgrade pip
+          pip install --requirement requirements.txt
