### 2024-10-18 07c1bae
delete macos test debug
