### 2024-10-18 64704cb
remove the useless has_electric since it is always true
