+++ pypeec/lib_solver/problem_geometry.py
@@ -80,11 +80,10 @@ def get_problem_type(idx_vc, idx_vm, idx_src_c, idx_src_v, graph_def):
     _check_indices(idx_vc, idx_src_c, idx_src_v)
     _check_source_graph(idx_vc, idx_src_c, idx_src_v, graph_def)
 
-    # check the existence of magnetic domains
-    has_electric = len(idx_vc) > 0
+    # detect the existence of magnetic domains
     has_magnetic = len(idx_vm) > 0
 
-    return has_electric, has_magnetic
+    return has_magnetic
 
 
 def get_material_idx(material_def, domain_def):
