### 2024-10-18 cc36a87
simplify empty matrix operators
