### 2024-10-20 0709bc5
fix vtk version (9.3 is problematic with conda)
