+++ pyproject.toml
@@ -57,11 +57,11 @@ dependencies = [
     "joblib >= 1.3",
     "numpy >= 1.24, < 1.27",
     "scipy >= 1.9, < 1.14",
+    "vtk >= 9.2, < 9.3",
     "matplotlib >= 3.7",
     "shapely >= 2.0",
     "rasterio >= 1.3",
     "Pillow >= 10.0",
-    "vtk >= 9.2",
     "pyvista >= 0.40",
     "pyvistaqt >= 0.11",
     "PyQt5 >= 5.15",
