+++ conda_base.yaml
@@ -15,8 +15,8 @@ dependencies:
   # base installation
   - python=3.10
   # logging and serialization
-  - scilogger=1.1.0
-  - scisave=1.2.0
+  - scilogger=1.2.0
+  - scisave=1.3.0
   - jsonschema=4.23.0
   # numerical libraries
   - numpy=1.24.1
