+++ conda_dev.yaml
@@ -26,8 +26,8 @@ dependencies:
   - sphinx=6.2.1
   - pkg-config=0.29.2
   # logging and serialization
-  - scilogger=1.1.0
-  - scisave=1.2.0
+  - scilogger=1.2.0
+  - scisave=1.3.0
   - jsonschema=4.23.0
   # numerical libraries
   - numpy=1.24.1
