+++ pyproject.toml
@@ -51,8 +51,8 @@ classifiers = [
 ]
 
 dependencies = [
-    "scilogger >= 1.1",
-    "scisave >= 1.2",
+    "scilogger >= 1.2",
+    "scisave >= 1.3",
     "jsonschema >= 4.23",
     "joblib >= 1.3",
     "numpy >= 1.24, < 1.27",
