### 2024-10-27 8f27fab
plotter and viewer documentation
