### 2024-10-28 79d83fa
scipy for requirements.txt
