+++ .github/workflows/run_package.yml
@@ -38,7 +38,7 @@ jobs:
         uses: conda-incubator/setup-miniconda@v3
         with:
           miniforge-version: latest
-          python-version: "3.10"
+          python-version: "3.11"
           auto-activate-base: false
           activate-environment: pypeec
           environment-file: ${{ matrix.method.file }}
@@ -69,7 +69,7 @@ jobs:
       - name: Set up Python
         uses: actions/setup-python@v5
         with:
-          python-version: "3.10"
+          python-version: "3.11"
       - name: Install dependencies
         shell: bash -el {0}
         run: |
