+++ .github/workflows/run_repository.yml
@@ -52,7 +52,7 @@ jobs:
         uses: conda-incubator/setup-miniconda@v3
         with:
           miniforge-version: latest
-          python-version: "3.10"
+          python-version: "3.11"
           auto-activate-base: false
           activate-environment: pypeec
           use-mamba: true
