+++ conda_base.yaml
@@ -13,7 +13,7 @@ name: pypeec
 channels: [conda-forge]
 dependencies:
   # base installation
-  - python=3.10
+  - python=3.11
   # logging and serialization
   - scilogger=1.2.0
   - scisave=1.3.0
