### 2024-10-28 f0b5e3b
lift to python 3.11
