### 2024-10-28 7dffddf
test osx scipy
