+++ .github/workflows/run_platform.yml
@@ -1,81 +1,0 @@
-# Run the "Platform tests" with the following conditions.
-#   - dispatch
-#   - weekly
-#
-# The tests are run on the following systems:
-#   - Linux / Ubuntu
-#   - Apple / macOS
-#   - Microsoft / Windows
-#
-# The tests are run on the following interpreter versions:
-#   - Python 3.9
-#   - Python 3.10
-#   - Python 3.11
-#   - Python 3.12
-#
-# Thomas Guillod - Dartmouth College
-# Mozilla Public License Version 2.0
-
-name: Run Platform Tests
-
-on:
-  workflow_dispatch:
-  schedule:
-    - cron: "0 2 * * 1"
-
-jobs:
-  # ===================== run the tests with different systems
-  system:
-    strategy:
-      fail-fast: false
-      matrix:
-        system:
-          - "ubuntu-latest"
-          - "macos-latest"
-          - "windows-latest"
-    name: ${{ matrix.system }}
-    runs-on: ${{ matrix.system }}
-    timeout-minutes: 20
-    steps:
-      - uses: actions/checkout@v4
-      - name: Set up Python
-        uses: actions/setup-python@v5
-        with:
-          python-version: "3.11"
-      - name: Install dependencies
-        shell: bash -el {0}
-        run: |
-          python -m pip install --upgrade pip
-          pip install --requirement requirements.txt
-      - name: Run the tests
-        shell: bash -el {0}
-        run: |
-          ./scripts/run_tests.sh
-  # ===================== run the tests with different interpreters
-  interpreter:
-    strategy:
-      fail-fast: false
-      matrix:
-        version:
-          - "3.9"
-          - "3.10"
-          - "3.11"
-          - "3.12"
-    name: ${{ matrix.version }}
-    runs-on: "ubuntu-latest"
-    timeout-minutes: 20
-    steps:
-      - uses: actions/checkout@v4
-      - name: Set up Python
-        uses: actions/setup-python@v5
-        with:
-          python-version: ${{ matrix.version }}
-      - name: Install dependencies
-        shell: bash -el {0}
-        run: |
-          python -m pip install --upgrade pip
-          python -m pip install --requirement requirements.txt
-      - name: Run the tests
-        shell: bash -el {0}
-        run: |
-          ./scripts/run_tests.sh
