### 2024-10-28 a514cd3
test osx scipy
