+++ requirements.txt
@@ -9,24 +9,8 @@
 # Thomas Guillod - Dartmouth College
 # Mozilla Public License Version 2.0
 
-# logging and serialization
-scilogger~=1.2.0
-scisave~=1.3.0
-jsonschema~=4.23
-
 # numerical libraries
 numpy~=1.26.4
 scipy~=1.13.1
 joblib~=1.4.2
 
-# geometry manipulation
-vtk~=9.3.1
-pyvista~=0.43.10
-shapely~=2.0.4
-rasterio~=1.3.10
-Pillow~=10.4.0
-
-# plotting libraries
-PyQt5~=5.15.10
-pyvistaqt~=0.11.1
-matplotlib~=3.9.0
