### 2024-10-28 70f25db
test osx scipy
