+++ requirements.txt
@@ -10,7 +10,4 @@
 # Mozilla Public License Version 2.0
 
 # numerical libraries
-numpy~=1.26.4
 scipy~=1.13.1
-joblib~=1.4.2
-
