### 2024-10-28 6999b35
test osx scipy
