### 2024-10-28 858e554
test osx scipy
