+++ .github/workflows/run_hook.yml
@@ -26,5 +26,4 @@ jobs:
       - name: Install dependencies
         shell: bash -el {0}
         run: |
-          python -m pip install --upgrade pip
           python -m pip install --requirement requirements.txt
