### 2024-10-28 67a0094
test osx scipy
