### 2024-10-28 ce1cad7
changes for numpy 2 and for scipy 1.12
