+++ pypeec/lib_matrix/multiply_direct.py
@@ -19,9 +19,9 @@ def _get_voxel_indices(nx, ny, nz):
     """
 
     # get the indices array
-    idx_x = np.arange(nx, dtype=np.int_)
-    idx_y = np.arange(ny, dtype=np.int_)
-    idx_z = np.arange(nz, dtype=np.int_)
+    idx_x = np.arange(nx, dtype=np.int64)
+    idx_y = np.arange(ny, dtype=np.int64)
+    idx_z = np.arange(nz, dtype=np.int64)
     [idx_x, idx_y, idx_z] = np.meshgrid(idx_x, idx_y, idx_z, indexing="ij")
 
     # flatten the indices into vectors
@@ -42,13 +42,13 @@ def _get_dense_zero(idx_out, idx_in, mat, idx_row, idx_col):
     nv = nx*ny*nz
 
     # get the matrix size
-    idx_row = np.in1d(np.arange(idx_row*nv, (idx_row+1)*nv, dtype=np.int_), idx_out)
-    idx_col = np.in1d(np.arange(idx_col*nv, (idx_col+1)*nv, dtype=np.int_), idx_in)
+    idx_row = np.in1d(np.arange(idx_row*nv, (idx_row+1)*nv, dtype=np.int64), idx_out)
+    idx_col = np.in1d(np.arange(idx_col*nv, (idx_col+1)*nv, dtype=np.int64), idx_in)
     n_row = np.count_nonzero(idx_row)
     n_col = np.count_nonzero(idx_col)
 
     # create an empty matrix
-    mat_dense = np.zeros((n_row, n_col), dtype=np.float_)
+    mat_dense = np.zeros((n_row, n_col), dtype=np.float64)
 
     return mat_dense
 
@@ -69,8 +69,8 @@ def _get_dense_diag(idx_out, idx_in, mat, idx_row, idx_col, sign_type):
     mat_tmp = mat.flatten(order="F")
 
     # get the indices of the non-empty face for the current dimension
-    idx_row = np.in1d(np.arange(idx_row*nv, (idx_row+1)*nv, dtype=np.int_), idx_out)
-    idx_col = np.in1d(np.arange(idx_col*nv, (idx_col+1)*nv, dtype=np.int_), idx_in)
+    idx_row = np.in1d(np.arange(idx_row*nv, (idx_row+1)*nv, dtype=np.int64), idx_out)
+    idx_col = np.in1d(np.arange(idx_col*nv, (idx_col+1)*nv, dtype=np.int64), idx_in)
     n_row = np.count_nonzero(idx_row)
     n_col = np.count_nonzero(idx_col)
 
@@ -96,7 +96,7 @@ def _get_dense_diag(idx_out, idx_in, mat, idx_row, idx_col, sign_type):
 
     # get the sign matrix
     idx_neg = np.logical_not(idx_pos)
-    sign = np.empty((n_row, n_col), dtype=np.int_)
+    sign = np.empty((n_row, n_col), dtype=np.int64)
     sign[idx_pos] = +1
     sign[idx_neg] = -1
 
@@ -126,7 +126,7 @@ def _get_prepare_sub(name, idx_out, idx_in, mat):
     # get the matrix size
     n_out = len(idx_out)
     n_in = len(idx_in)
-    itemsize = np.dtype(np.float_).itemsize
+    itemsize = np.dtype(np.float64).itemsize
     footprint = (itemsize*n_out*n_in)/(1024**2)
 
     # display the matrix size
@@ -136,10 +136,10 @@ def _get_prepare_sub(name, idx_out, idx_in, mat):
     # get the permutation for sorting
     idx_perm_out = np.argsort(idx_out)
[…]
