### 2024-10-28 f2d9732
scipy for macos
