+++ requirements.txt
@@ -16,7 +16,7 @@ jsonschema~=4.23
 
 # numerical libraries
 numpy~=1.26.4
-scipy~=1.13.1
+scipy~=1.14.1
 joblib~=1.4.2
 
 # geometry manipulation
