+++ INSTALL.md
@@ -14,7 +14,7 @@ The optional HPC libraries are usually easier to install through Conda.
 ```bash
 # Install a Python interpreter
 #   - Website: https://www.python.org/downloads
-#   - Supported versions: 3.9, 3.10, 3.11, and 3.12
+#   - Supported versions: 3.10, 3.11, and 3.12
 #   - Python executable: "python" or "python3"
 
 # Create a Python Virtual Environment
