+++ requirements.txt
@@ -15,9 +15,9 @@ scisave~=1.3.0
 jsonschema~=4.23
 
 # numerical libraries
-numpy~=1.26.4
-scipy~=1.14.1
-joblib~=1.4.2
+numpy
+scipy
+joblib
 
 # geometry manipulation
 vtk
