+++ pypeec/lib_solver/equation_solver.py
@@ -184,7 +184,7 @@ class _OpCounter:
             y = op(x)
             return y
 
-        op_count = sla.LinearOperator((n_dof, n_dof), matvec=fct, dtype=np.complex_)
+        op_count = sla.LinearOperator((n_dof, n_dof), matvec=fct, dtype=np.complex128)
 
         return op_count
 
@@ -198,7 +198,7 @@ class _OpCounter:
             y = op(x)
             return y
 
-        op_count = sla.LinearOperator((n_dof, n_dof), matvec=fct, dtype=np.complex_)
+        op_count = sla.LinearOperator((n_dof, n_dof), matvec=fct, dtype=np.complex128)
 
         return op_count
 
@@ -472,7 +472,7 @@ def get_solver(sol_init, fct_cpl, fct_sys, fct_pcd, rhs, fct_conv, solver_option
 
     # get initial solution
     if sol_init is None:
-        sol_init = np.zeros(n_dof_total, dtype=np.complex_)
+        sol_init = np.zeros(n_dof_total, dtype=np.complex128)
 
     # create operator and iter counter object
     op_obj = _OpCounter()
