+++ pypeec/lib_solver/equation_system.py
@@ -152,12 +152,12 @@ def _get_coupling_electric(sol_m, freq, n_vc, n_fc, n_fm, n_src, K_op_c):
 
     # compute the couplings
     if freq == 0:
-        cpl_fc = np.zeros(n_fc, dtype=np.complex_)
+        cpl_fc = np.zeros(n_fc, dtype=np.complex128)
     else:
         cpl_fc = s*K_op_c(I_fm)
 
-    cpl_vc = np.zeros(n_vc, dtype=np.complex_)
-    cpl_src = np.zeros(n_src, dtype=np.complex_)
+    cpl_vc = np.zeros(n_vc, dtype=np.complex128)
+    cpl_src = np.zeros(n_src, dtype=np.complex128)
 
     # assemble the vectors
     cpl_c = np.concatenate((cpl_fc, cpl_vc, cpl_src))
@@ -179,7 +179,7 @@ def _get_coupling_magnetic(sol_c, n_fc, n_vm, K_op_m):
 
     # compute the couplings
     cpl_fm = -K_op_m(I_fc)
-    cpl_vm = np.zeros(n_vm, dtype=np.complex_)
+    cpl_vm = np.zeros(n_vm, dtype=np.complex128)
 
     # assemble the vectors
     cpl_m = np.concatenate((cpl_fm, cpl_vm))
@@ -220,14 +220,14 @@ def _get_cond_fact_electric(freq, A_net_c, R_c, L_c, A_src):
 
     # expand for the source matrices
     A_add = sps.csc_matrix((n_fc, n_src), dtype=np.int64)
-    A_12_mat = sps.hstack([A_12_mat, A_add], dtype=np.complex_, format="csr")
+    A_12_mat = sps.hstack([A_12_mat, A_add], dtype=np.complex128, format="csr")
 
     # expand for the source matrices
     A_add = sps.csc_matrix((n_src, n_fc), dtype=np.int64)
-    A_21_mat = sps.vstack([A_21_mat, A_add], dtype=np.complex_, format="csc")
+    A_21_mat = sps.vstack([A_21_mat, A_add], dtype=np.complex128, format="csc")
 
     # add the source
-    A_22_mat = sps.bmat([[A_22_mat, A_vc_src], [A_src_vc, A_src_src]], dtype=np.complex_, format="csc")
+    A_22_mat = sps.bmat([[A_22_mat, A_vc_src], [A_src_vc, A_src_src]], dtype=np.complex128, format="csc")
 
     # computing the Schur complement (with respect to the diagonal admittance matrix)
     S_mat = A_22_mat-A_21_mat*Y_mat*A_12_mat
@@ -250,7 +250,7 @@ def _get_cond_fact_magnetic(A_net_m, R_m, P_m):
 
     # get the magnetic admittance (avoid singularity for DC solution)
     Y_m = 1/R_m
-    I_m = np.ones(n_vm, dtype=np.complex_)
+    I_m = np.ones(n_vm, dtype=np.complex128)
 
     # admittance matrix
     Y_mat = sps.diags(Y_m, format="csc")
@@ -318,7 +318,7 @@ def _get_system_multiply_electric(sol, freq, A_net_c, A_src, R_c, L_op_c):
 
     # multiply the inductance matrix
     if freq == 0:
-        rhs_kvl_ind = np.zeros(n_fc, dtype=np.complex_)
+        rhs_kvl_ind = np.zeros(n_fc, dtype=np.complex128)
     else:
         rhs_kvl_ind = s*L_op_c(I_fc)
 
@@ -389,8 +389,8 @@ def get_source_vector(idx_vc, idx_vm, idx_fc, idx_fm, I_src_c, V_src_v):
     n_m = len(idx_vm)+len(idx_fm)
 
[…]

[…]
