### 2024-10-28 bfd9f90
cupy types upgrade
