+++ pypeec/lib_matrix/multiply_fft.py
@@ -32,19 +32,19 @@ def _get_tensor_sign(name, nd_in):
     """
 
     if name == "potential":
-        sign = NPCP.ones((2, 2, 2, nd_in), dtype=NPCP.float_)
+        sign = NPCP.ones((2, 2, 2, nd_in), dtype=NPCP.float64)
     elif name == "inductance":
-        sign = NPCP.ones((2, 2, 2, nd_in), dtype=NPCP.float_)
+        sign = NPCP.ones((2, 2, 2, nd_in), dtype=NPCP.float64)
     elif name == "coupling":
-        sign = NPCP.empty((2, 2, 2, nd_in), dtype=NPCP.float_)
-        sign[0, 0, 0, :] = NPCP.array([+1, +1, +1], dtype=NPCP.float_)
-        sign[1, 0, 0, :] = NPCP.array([-1, +1, +1], dtype=NPCP.float_)
-        sign[0, 1, 0, :] = NPCP.array([+1, -1, +1], dtype=NPCP.float_)
-        sign[0, 0, 1, :] = NPCP.array([+1, +1, -1], dtype=NPCP.float_)
-        sign[1, 1, 0, :] = NPCP.array([-1, -1, +1], dtype=NPCP.float_)
-        sign[1, 0, 1, :] = NPCP.array([-1, +1, -1], dtype=NPCP.float_)
-        sign[0, 1, 1, :] = NPCP.array([+1, -1, -1], dtype=NPCP.float_)
-        sign[1, 1, 1, :] = NPCP.array([-1, -1, -1], dtype=NPCP.float_)
+        sign = NPCP.empty((2, 2, 2, nd_in), dtype=NPCP.float64)
+        sign[0, 0, 0, :] = NPCP.array([+1, +1, +1], dtype=NPCP.float64)
+        sign[1, 0, 0, :] = NPCP.array([-1, +1, +1], dtype=NPCP.float64)
+        sign[0, 1, 0, :] = NPCP.array([+1, -1, +1], dtype=NPCP.float64)
+        sign[0, 0, 1, :] = NPCP.array([+1, +1, -1], dtype=NPCP.float64)
+        sign[1, 1, 0, :] = NPCP.array([-1, -1, +1], dtype=NPCP.float64)
+        sign[1, 0, 1, :] = NPCP.array([-1, +1, -1], dtype=NPCP.float64)
+        sign[0, 1, 1, :] = NPCP.array([+1, -1, -1], dtype=NPCP.float64)
+        sign[1, 1, 1, :] = NPCP.array([-1, -1, -1], dtype=NPCP.float64)
     else:
         raise ValueError("invalid matrix type")
 
@@ -64,7 +64,7 @@ def _get_tensor_circulant(mat, sign):
     (nx, ny, nz, nd_in) = mat.shape
 
     # init the circulant tensor
-    mat_fft = NPCP.zeros((2*nx, 2*ny, 2*nz, nd_in), dtype=NPCP.float_)
+    mat_fft = NPCP.zeros((2*nx, 2*ny, 2*nz, nd_in), dtype=NPCP.float64)
 
     # cube none
     mat_fft[0:nx, 0:ny, 0:nz, :] = mat[0:nx, 0:ny, 0:nz, :]*sign[0:1, 0:1, 0:1, :]
@@ -136,7 +136,7 @@ def _get_tensor(idx, vec):
     idx_mat = idx["idx_mat"]
 
     # init the tensor
-    res = NPCP.zeros(shape, dtype=NPCP.complex_)
+    res = NPCP.zeros(shape, dtype=NPCP.complex128)
 
     # assign the tensor (4D or 3D slice)
     if idx_sel is None:
[…]
