+++ .github/workflows/run_platform.yml
@@ -28,11 +28,15 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
+        version:
+          - "3.10"
+          - "3.11"
+          - "3.12"
         system:
           - "ubuntu-latest"
           - "macos-latest"
           - "windows-latest"
-    name: ${{ matrix.system }}
+    name: ${{ matrix.system }} / ${{ matrix.version }}
     runs-on: ${{ matrix.system }}
     timeout-minutes: 20
     steps:
@@ -40,7 +44,7 @@ jobs:
       - name: Set up Python
         uses: actions/setup-python@v5
         with:
-          python-version: "3.11"
+          python-version: ${{ matrix.version }}
       - name: Install dependencies
         shell: bash -el {0}
         run: |
