+++ .github/workflows/run_platform.yml
@@ -33,9 +33,7 @@ jobs:
           - "3.11"
           - "3.12"
         system:
-          - "ubuntu-latest"
           - "macos-latest"
-          - "windows-latest"
     name: ${{ matrix.system }} / ${{ matrix.version }}
     runs-on: ${{ matrix.system }}
     timeout-minutes: 20
