+++ requirements.txt
@@ -16,8 +16,8 @@ jsonschema~=4.23
 
 # numerical libraries
 numpy~=1.26.4
-scipy~=1.14.1
-joblib~=1.4.2
+scipy~=1.13.1
+joblib~=1.3.2
 
 # geometry manipulation
 vtk~=9.3.1
