+++ requirements.txt
@@ -10,7 +10,7 @@
 # Mozilla Public License Version 2.0
 
 # numerical libraries
-numpy~=1.26.4
+numpy~=1.24.1
 scipy~=1.14.1
 joblib~=1.3.2
 
