+++ .github/workflows/run_hook.yml
@@ -19,12 +19,16 @@ jobs:
     timeout-minutes: 20
     steps:
       - uses: actions/checkout@v4
-      - name: Set up Python
-        uses: actions/setup-python@v5
+      - name: Set up Conda
+        uses: conda-incubator/setup-miniconda@v3
         with:
+          miniforge-version: latest
           python-version: "3.11"
-      - name: Install dependencies
+          auto-activate-base: false
+          activate-environment: pypeec
+          environment-file: "conda_base.yaml"
+          use-mamba: true
+      - name: Run the tests
         shell: bash -el {0}
         run: |
-          python -m pip install --upgrade pip
-          python -m pip install --requirement requirements.txt
+          ./scripts/run_tests.sh
