+++ .github/workflows/run_repository.yml
@@ -16,7 +16,6 @@
 #   - Python 3.11
 #   - Python 3.12
 #
-#
 # Thomas Guillod - Dartmouth College
 # Mozilla Public License Version 2.0
 
@@ -67,7 +66,7 @@ jobs:
           - "3.10"
           - "3.11"
           - "3.12"
-    name: ${{ matrix.system }} / ${{ matrix.version }}
+    name: conda / ${{ matrix.system }} / ${{ matrix.version }}
     runs-on: ${{ matrix.system }}
     timeout-minutes: 20
     steps:
@@ -89,7 +88,7 @@ jobs:
           ${{ env.CMD_EXTRACT }}
           ${{ env.CMD_SOLVE }}
           ${{ env.CMD_PLOT }}
-  # ===================== run the tests with pip
+  # ===================== run the tests with pypi
   pip:
     strategy:
       fail-fast: false
@@ -101,7 +100,7 @@ jobs:
           - "3.10"
           - "3.11"
           - "3.12"
-    name: ${{ matrix.system }} / ${{ matrix.version }}
+    name: pypi / ${{ matrix.system }} / ${{ matrix.version }}
     runs-on: ${{ matrix.system }}
     timeout-minutes: 20
     steps:
