+++ conda_dev.yaml
@@ -17,14 +17,14 @@ dependencies:
   # base installation
   - python=3.11
   # dev tools
-  - pip=23.1.2
-  - twine=4.0.2
-  - python-build=0.10.0
-  - coverage=7.4.1
-  - setuptools_scm=7.1.0
+  - pip=24.2
+  - twine=5.1.1
+  - coverage=7.6.4
+  - setuptools_scm=8.1.0
   - sphinx_rtd_theme=3.0.1
   - sphinx=8.1.3
   - pkg-config=0.29.2
+  - python-build=1.2.2
   # logging and serialization
   - scilogger=1.2.0
   - scisave=1.3.0
