+++ .github/workflows/run_repository.yml
@@ -88,7 +88,7 @@ jobs:
           ${{ env.CMD_EXTRACT }}
           ${{ env.CMD_SOLVE }}
           ${{ env.CMD_PLOT }}
-  # ===================== run the tests with pypi
+  # ===================== run the tests with pip
   pip:
     strategy:
       fail-fast: false
@@ -100,7 +100,7 @@ jobs:
           - "3.10"
           - "3.11"
           - "3.12"
-    name: pypi / ${{ matrix.system }} / ${{ matrix.version }}
+    name: pip / ${{ matrix.system }} / ${{ matrix.version }}
     runs-on: ${{ matrix.system }}
     timeout-minutes: 20
     steps:
