+++ .github/workflows/run_package.yml
@@ -25,11 +25,11 @@ jobs:
       fail-fast: false
       matrix:
         method:
-          - name: "conda_base"
+          - name: "base"
             file: "conda_base.yaml"
-          - name: "conda_dev"
+          - name: "dev"
             file: "conda_dev.yaml"
-    name: ${{ matrix.method.name }}
+    name: conda / ${{ matrix.method.name }}
     runs-on: "ubuntu-latest"
     timeout-minutes: 20
     steps:
@@ -53,15 +53,15 @@ jobs:
       fail-fast: false
       matrix:
         method:
-          - name: "requirements"
+          - name: "list"
             cmd: |
               python -m pip install --upgrade pip
               python -m pip install --requirement requirements.txt
-          - name: "pyproject"
+          - name: "package"
             cmd: |
               python -m pip install --upgrade pip
               python -m pip install .
-    name: ${{ matrix.method.name }}
+    name: pip / ${{ matrix.method.name }}
     runs-on: "ubuntu-latest"
     timeout-minutes: 20
     steps:
