+++ requirements.txt
@@ -11,8 +11,8 @@
 
 # logging and serialization
 scilogger~=1.2.0
-scisave~=1.3.0
-jsonschema~=4.23
+scisave~=1.4.0
+jsonschema~=4.23.0
 
 # numerical libraries
 numpy~=1.26.4
