+++ conda_dev.yaml
@@ -27,7 +27,7 @@ dependencies:
   - python-build=1.2.2
   # logging and serialization
   - scilogger=1.2.0
-  - scisave=1.3.0
+  - scisave=1.4.0
   - jsonschema=4.23.0
   # numerical libraries
   - numpy=1.26.4
