+++ conda_base.yaml
@@ -16,7 +16,7 @@ dependencies:
   - python=3.11
   # logging and serialization
   - scilogger=1.2.0
-  - scisave=1.3.0
+  - scisave=1.4.0
   - jsonschema=4.23.0
   # numerical libraries
   - numpy=1.26.4
