+++ pyproject.toml
@@ -52,7 +52,7 @@ classifiers = [
 
 dependencies = [
     "scilogger >= 1.2",
-    "scisave >= 1.3",
+    "scisave >= 1.4",
     "jsonschema >= 4.23",
     "joblib >= 1.3",
     "numpy >= 1.24",
