+++ .github/workflows/run_hook.yml
@@ -6,7 +6,7 @@
 # Thomas Guillod - Dartmouth College
 # Mozilla Public License Version 2.0
 
-name: Run Standard Tests
+name: Run Hook Tests
 
 on:
   push:
