### 2024-11-13 f6b19ad
better check for plot tags
