### 2024-11-18 #6644 @otvam
Here is his IEEE Xplore profile:
https://ieeexplore.ieee.org/author/534993517644449
