### 2024-11-20 c266f1a
improved Biot-Savart field computation
