### 2024-11-21 81043ea
bug for Biot-Savart field computation
