### 2024-11-21 46e9505
biot savart computation option
