+++ examples/config/tolerance.yaml
@@ -19,6 +19,11 @@ parallel_sweep:
 #   - if the normalized voxel distance is larger than the threshold, numerical approximations are used
 "integral_simplify": 20.0
 
+# control of the magnetic field is computed for the point cloud
+#   - "face" is using the face currents to compute the magnetic field
+#   - "voxel" is using the voxel currents to compute the magnetic field
+"biot_savart": "face"
+
 # method for dense matrix multiplication
 #   - "fft" for doing the multiplication with circulant tensors and FFT
 #   - "direct" for standard matrix multiplication
