### 2024-11-21 796767d
biot savart computation option
