### 2024-11-22 d917dd6
partial inductance warning
