### 2024-11-24 #6644 @likeajumprope
> @likeajumprope Apologies for the delayed response to your email. I would be happy to review the paper and software.

Perfect, thank you so much @svenweihe
