### 2024-11-24 #6644 @svenweihe
@likeajumprope I have completed my review and am happy with the various items on the checklist. The only comment I would have for the authors is that in the case of most power electronics applications a square wave voltage is applied as a source and, as such, it would be very nice to have at least one example of a simulation where the most significant harmonics required to represent the square wave are swept, and the swept results are then combined such that they can then be visualized.
