+++ pypeec/run/solver.py
@@ -308,7 +308,7 @@ def _run_solver_sweep(data_solver, data_internal, data_param, sol_init):
         H_pts_c = extract_solution.get_magnetic_field_electric(n, d, idx_fc, A_net_c, I_fc, pts_net_c, pts_cloud)
 
         # get the cloud point magnetic field (contributions of the magnetic domains)
-        H_pts_m = extract_solution.get_magnetic_magnetic(A_net_m, I_fm, pts_net_m, pts_cloud)
+        H_pts_m = extract_solution.get_magnetic_field_magnetic(A_net_m, I_fm, pts_net_m, pts_cloud)
 
     # assemble solution
     var = {
