### 2024-11-26 a657dae
typo for utils type checks
