### 2024-11-27 #6644 @otvam
@svenweihe Thanks for accepting this review!

The comment about making Fourier analysis is interesting. The PEEC method can be used in the time-domain but it is rarely a good idea as the differential equations are **extremely stiff and unstable**. So the best workflow is indeed through Fourier analysis:
	- Transforming the excitation from the time-domain into the frequency-domain
	- Solving the PEEC problem in the frequency-domain
	- Transforming the solution back into time-domain

PyPEEC can support such a workflow and I agree that an example in the documentation would be useful.

I propose the following example:
	- Geometry: air-core inductor
	- Excitation: PWM voltage
	- Solution: impedance is affected by skin effect (frequency-dependent resistance)
	- Results: computation of the inductor current and the associated losses

Would you think that such an example would be good? I am open to any other power eletronics example (the only limitation being that the example should have a low memory footprint in order to be compatible with GitHub).
