### 2024-11-28 b52ad5e
new CI tests
