+++ .github/envs/pip.txt
@@ -1,0 +1,29 @@
+# Definition of the Python Virtual Environment (minimum versions).
+#
+# WARNING: Only the required packages are in the environment.
+#          The optional libraries are not installed.
+#
+# Thomas Guillod - Dartmouth College
+# Mozilla Public License Version 2.0
+
+# logging and serialization
+scilogger >= 1.2
+scisave >= 1.4
+jsonschema >= 4.23
+
+# numerical libraries
+joblib >= 1.3
+numpy >= 1.24
+scipy >= 1.12
+
+# geometry manipulation
+vtk >= 9.2
+pyvista >= 0.40
+shapely >= 2.0
+rasterio >= 1.3
+Pillow >= 10.0
+
+# plotting libraries
+PyQt5 >= 5.15
+pyvistaqt >= 0.11
+matplotlib >= 3.7
