+++ .github/workflows/run_version.yml
@@ -1,10 +1,10 @@
-# Run the "Repository tests" with the following conditions.
+# Run the "Version tests" with the following conditions.
 #   - dispatch
 #   - weekly
 #
 # The tests are run on the following repositories:
-#   - installing the package from the PyPi repository
-#   - installing the package from the Conda repository
+#   - installing the dependencies from "pyproject.toml"
+#   - installing the dependencies from "pyconda.yaml"
 #
 # The tests are run on the following systems:
 #   - Linux / Ubuntu / conda / pip
@@ -19,7 +19,7 @@
 # Thomas Guillod - Dartmouth College
 # Mozilla Public License Version 2.0
 
-name: Run Repository Tests
+name: Run Version Tests
 
 on:
   workflow_dispatch:
@@ -70,6 +70,7 @@ jobs:
     runs-on: ${{ matrix.system }}
     timeout-minutes: 20
     steps:
+      - uses: actions/checkout@v4
       - name: Set up Conda
         uses: conda-incubator/setup-miniconda@v3
         with:
@@ -77,11 +78,13 @@ jobs:
           python-version: ${{ matrix.version }}
           auto-activate-base: false
           activate-environment: pypeec
+          environment-file: pyconda.yaml
           use-mamba: true
       - name: Install dependencies
         shell: bash -el {0}
         run: |
-          conda install --yes conda-forge::pypeec
+          conda env update --file .github/envs/conda.yaml
+          python -m pip install --no-deps .
       - name: Run the tests
         shell: bash -el {0}
         run: |
@@ -104,6 +107,7 @@ jobs:
     runs-on: ${{ matrix.system }}
     timeout-minutes: 20
     steps:
+      - uses: actions/checkout@v4
       - name: Set up Python
         uses: actions/setup-python@v5
         with:
@@ -111,7 +115,8 @@ jobs:
       - name: Install dependencies
         shell: bash -el {0}
         run: |
-          python -m pip install pypeec
+          python -m pip install --requirement .github/envs/pip.txt
+          python -m pip install --no-deps .
       - name: Run the tests
         shell: bash -el {0}
         run: |
