+++ conda_base.yaml
@@ -1,34 +1,0 @@
-# Definition of the Conda Environment for PyPEEC.
-#
-# WARNING: Only the required packages are in the environment.
-#          The following optional libraries are not installed:
-#              - FFTW, UMFPACK, PyAMG, and MKL/PARDISO
-#              - Jupyter, IPyWidgets, and Trame
-#              - CuPy and CUDA
-#
-# Thomas Guillod - Dartmouth College
-# Mozilla Public License Version 2.0
-
-name: pypeec
-channels: [conda-forge]
-dependencies:
-  # base installation
-  - python=3.11
-  # logging and serialization
-  - scilogger=1.2.0
-  - scisave=1.4.0
-  - jsonschema=4.23.0
-  # numerical libraries
-  - numpy=1.26.4
-  - scipy=1.13.1
-  - joblib=1.4.2
-  # geometry libraries
-  - vtk=9.2.6
-  - pyvista=0.40.0
-  - shapely=2.0.1
-  - rasterio=1.3.8
-  - pillow=10.0.0
-  # plotting libraries
-  - pyqt=5.15.7
-  - pyvistaqt=0.11.0
-  - matplotlib=3.7.1
