+++ .github/envs/conda.yaml
@@ -1,0 +1,28 @@
+# Definition of the Conda Environment for PyPEEC (minimum versions).
+#
+# WARNING: Only the required packages are in the environment.
+#          The optional libraries are not installed.
+#
+# Thomas Guillod - Dartmouth College
+# Mozilla Public License Version 2.0
+
+channels: [conda-forge]
+dependencies:
+  # logging and serialization
+  - scilogger >=1.2
+  - scisave >=1.4
+  - jsonschema >=4.23
+  # numerical libraries
+  - numpy >=1.24
+  - scipy >=1.12
+  - joblib >=1.3
+  # geometry libraries
+  - vtk >=9.2, < 9.3
+  - pyvista >=0.40
+  - shapely >=2.0
+  - rasterio >=1.3
+  - pillow >=10.0
+  # plotting libraries
+  - pyqt >=5.15
+  - pyvistaqt >=0.11
+  - matplotlib >=3.7
