### 2024-11-28 5c64960
new CI tests
