+++ .github/workflows/run_version.yml
@@ -22,6 +22,7 @@
 name: Run Version Tests
 
 on:
+  push:
   workflow_dispatch:
   schedule:
     - cron: "0 2 * * 1"
