### 2024-11-28 8d49aa8
new CI tests
