### 2024-11-28 b6fb509
new CI tests
