### 2024-11-28 290adbb
script for CI
