+++ scripts/run_coverage.sh
@@ -1,15 +1,5 @@
 #!/bin/bash
-# Script for checking the code coverage:
-#   - run the tests with coverage analysis
-#   - generate an HTML report with the results
-#
-# The variable "ALLOW_PLOTTING" are allowing the tests to be run without a display server.
-# The variables "PYTHONIOENCODING" and "PYTHONUNBUFFERED" are setting up the Python interpreter.
-#
-# The following variables control the tests:
-#   - TEST_TOL: relative tolerance for the test results
-#   - TEST_CHECK: check (or not) the test results
-#   - TEST_SET: generate (or not) the test results
+# Script for checking the code coverage.
 #
 # Thomas Guillod - Dartmouth College
 # Mozilla Public License Version 2.0
@@ -48,15 +38,15 @@ function coverage_html {
   ret=$(( ret || $? ))
 }
 
-# global variables for the Python interpreter
-export ALLOW_PLOTTING="true"
-export PYTHONIOENCODING="utf8"
-export PYTHONUNBUFFERED="1"
+# set up the Python interpreter
+export ALLOW_PLOTTING="true"      # tests can run without a display server
+export PYTHONIOENCODING="utf8"    # set encoding for the console
+export PYTHONUNBUFFERED="1"       # disable buffering for the console
 
-# global variables for disabling the tests
-export TEST_TOL="nan"
-export TEST_CHECK="0"
-export TEST_SET="0"
+# options for the tests
+export TEST_TOL="nan"             # relative tolerance for the test results
+export TEST_CHECK="0"             # check (or not) the test results
+export TEST_SET="0"               # generate (or not) the test results
 
 # change to root directory
 cd "$(dirname "$0")" && cd ..
