+++ .github/envs/requirements.txt
@@ -1,0 +1,31 @@
+# Definition of the Python Virtual Environment.
+#
+# WARNING: Only tested with Linux x86/64.
+#
+# WARNING: Only the required packages are in the environment.
+#          The optional libraries are not installed.
+#
+# Thomas Guillod - Dartmouth College
+# Mozilla Public License Version 2.0
+
+# logging and serialization
+scilogger~=1.2.0
+scisave~=1.4.0
+jsonschema~=4.23.0
+
+# numerical libraries
+numpy~=1.26.4
+scipy~=1.14.1
+joblib~=1.4.2
+
+# geometry manipulation
+vtk~=9.3.1
+pyvista~=0.43.10
+shapely~=2.0.4
+rasterio~=1.3.10
+Pillow~=10.4.0
+
+# plotting libraries
+PyQt5~=5.15.10
+pyvistaqt~=0.11.1
+matplotlib~=3.9.0
