+++ .github/workflows/run_version.yml
@@ -11,14 +11,13 @@
 
 name: Run Version Tests
 
-### ===================== job triggers
 on:
+  # ===================== job triggers
   push:
   workflow_dispatch:
   schedule:
     - cron: "0 2 * * 1"
 
-### ===================== job definition
 jobs:
   # ===================== run with conda
   conda:
@@ -48,11 +47,10 @@ jobs:
           use-mamba: true
       - name: Install dependencies
         shell: bash -el {0}
-        run: |
-          conda env update --file .github/envs/conda.yaml
+        run: conda env update --file .github/envs/env_conda.yaml
       - name: Run the tests
         shell: bash -el {0}
-        run: ./scripts/run_tests.sh
+        run: ./.github/scripts/run_tests.sh
   # ===================== run with pypi
   pypi:
     strategy:
@@ -76,8 +74,7 @@ jobs:
           python-version: ${{ matrix.version }}
       - name: Install dependencies
         shell: bash -el {0}
-        run: |
-          python -m pip install --requirement .github/envs/pip.txt
+        run: python -m pip install --requirement .github/envs/env_pypi.txt
       - name: Run the tests
         shell: bash -el {0}
-        run: ./scripts/run_tests.sh
+        run: ./.github/scripts/run_tests.sh
