+++ .github/workflows/run_hook.yml
@@ -8,14 +8,14 @@
 
 name: Run Hook Tests
 
-### ===================== job triggers
 on:
+  # ===================== job triggers
   push:
   pull_request:
   workflow_dispatch:
 
-### ===================== job definition
 jobs:
+  # ===================== run the tests
   default:
     runs-on: "ubuntu-latest"
     timeout-minutes: 20
@@ -29,7 +29,7 @@ jobs:
         shell: bash -el {0}
         run: |
           python -m pip install --upgrade pip
-          python -m pip install --requirement requirements.txt
+          python -m pip install --requirement .github/envs/requirements.txt
       - name: Run the tests
         shell: bash -el {0}
-        run: ./scripts/run_tests.sh
+        run: ./.github/scripts/run_tests.sh

[…]
