+++ .github/workflows/run_repository.yml
@@ -36,6 +36,8 @@ jobs:
     runs-on: ${{ matrix.system }}
     timeout-minutes: 20
     steps:
+      - uses: actions/checkout@v4
+        sparse-checkout: .github
       - name: Set up Conda
         uses: conda-incubator/setup-miniconda@v3
         with:
@@ -66,6 +68,8 @@ jobs:
     runs-on: ${{ matrix.system }}
     timeout-minutes: 20
     steps:
+      - uses: actions/checkout@v4
+        sparse-checkout: .github
       - name: Set up Python
         uses: actions/setup-python@v5
         with:
@@ -76,3 +80,4 @@ jobs:
       - name: Run the script
         shell: bash -el {0}
         run: ./.github/scripts/run_script.sh
+
