### 2024-11-28 50941c7
script for CI
