+++ .github/workflows/run_repository.yml
@@ -38,6 +38,7 @@ jobs:
     steps:
       - uses: actions/checkout@v4
         with:
+          sparse-checkout-cone-mode: false
           sparse-checkout: .github
       - name: Set up Conda
         uses: conda-incubator/setup-miniconda@v3
@@ -71,6 +72,7 @@ jobs:
     steps:
       - uses: actions/checkout@v4
         with:
+          sparse-checkout-cone-mode: false
           sparse-checkout: .github
       - name: Set up Python
         uses: actions/setup-python@v5
