+++ .github/scripts/run_script.sh
@@ -12,8 +12,6 @@ export ALLOW_PLOTTING="true"      # tests can run without a display server
 export PYTHONIOENCODING="utf8"    # set encoding for the console
 export PYTHONUNBUFFERED="1"       # disable buffering for the console
 
-tree -a
-
 # check the version
 pypeec --version
 
