+++ .github/envs/env_conda.yaml
@@ -17,7 +17,7 @@ dependencies:
   - scipy >=1.12
   - joblib >=1.3
   # geometry libraries
-  - vtk >=9.2, < 9.3
+  - vtk >=9.2
   - pyvista >=0.40
   - shapely >=2.0
   - rasterio >=1.3
