+++ docs/content/warnings.rst
@@ -48,10 +48,10 @@ Library Issues
     GPU support is optional, PyPEEC is fully functional without GPU support.
 
 .. Warning::
-    **FFTW, UMFPACK, PyAMG, and MKL/PARDISO** are not included in the default package dependencies.
+    **FFTW, PyAMG, and MKL/PARDISO** are not included in the default package dependencies.
     These libraries can be tricky to install, especially on MS Windows or Apple MacOS.
     Make sure that these libraries are compiled with multithreading support.
-    FFTW, UMFPACK, PyAMG, and MKL/PARDISO are optional, PyPEEC is fully functional without them.
+    FFTW, PyAMG, and MKL/PARDISO are optional, PyPEEC is fully functional without them.
 
 General Issues
 --------------
