+++ pypeec/data/schema_tolerance.yaml
@@ -183,7 +183,6 @@
                 "type": "string"
                 "enum":
                     - "SuperLU"
-                    - "UMFPACK"
                     - "PARDISO"
                     - "PyAMG"
             "pyamg_options":
