### 2024-11-29 c2f7acc
lift the pinned version for conda.yaml
