+++ conda.yaml
@@ -19,9 +19,9 @@ dependencies:
   # base installation
   - python=3.11
   # dev tools
-  - pip=24.2
+  - pip=24.3.1
   - twine=5.1.1
-  - coverage=7.6.4
+  - coverage=7.6.8
   - setuptools_scm=8.1.0
   - sphinx_rtd_theme=3.0.1
   - sphinx=8.1.3
@@ -32,33 +32,30 @@ dependencies:
   - scisave=1.4.0
   - jsonschema=4.23.0
   # numerical libraries
-  - numpy=1.26.4
-  - scipy=1.13.1
+  - numpy=2.1.3
+  - scipy=1.14.1
   - joblib=1.4.2
   # geometry libraries
-  - vtk=9.2.6
-  - pyvista=0.40.0
-  - shapely=2.0.1
-  - rasterio=1.3.8
-  - pillow=10.0.0
+  - vtk=9.3.1
+  - pyvista=0.44.2
+  - shapely=2.0.6
+  - rasterio=1.4.2
+  - pillow=11.0.0
   # plotting libraries
-  - pyqt=5.15.7
-  - pyvistaqt=0.11.0
-  - matplotlib=3.7.1
+  - pyqt=5.15.9
+  - pyvistaqt=0.11.1
+  - matplotlib=3.9.2
   # optional packages for Jupyter
-  - jupyterlab=4.0.4
-  - ipywidgets=8.1.0
-  - trame=3.1.0
+  - jupyterlab=4.3.1
+  - ipywidgets=8.1.5
+  - trame-vtk=2.8.12
+  - trame-vuetify=2.7.2
+  - trame=3.7.0
   # optional optimized numerical libraries
-  - pyamg=5.0.1
-  - cupy=12.1.0
-  - scikit-umfpack=0.3.3
+  - pyamg=5.2.1
+  - cupy=13.3.0
+  - pyfftw=0.15.0
+  - pydiso=0.1.2
   # optional optimized numerical toolkits
-  - mkl-devel=2023.2.0
+  - mkl-devel=2024.2.2
   - cudatoolkit=11.8.0
-  # optional optimized numerical libraries
-  - pip:
-    # the conda package is not multithreaded
-    - pyFFTW==0.13.1
-    # the conda package is partially broken
-    - pydiso==0.0.5
