### 2024-11-29 17004bb
lift the pinned version for requirements.txt
