+++ .github/workflows/run_version.yml
@@ -13,7 +13,6 @@ name: Run Version Tests
 
 on:
   # ===================== job triggers
-  push:
   workflow_dispatch:
   schedule:
     - cron: "0 2 * * 1"
