+++ .github/workflows/run_repository.yml
@@ -13,7 +13,6 @@ name: Run Repository Tests
 
 on:
   # ===================== job triggers
-  push:
   workflow_dispatch:
   schedule:
     - cron: "0 2 * * 1"
