### 2024-11-29 c3c918b
voxel indices utils example
