### 2024-12-02 59d6b47
ac/dc for the stl transformer
