+++ examples/examples_stl/transformer/problem.yaml
@@ -43,8 +43,14 @@
     "sec_src": {"V_re": 0.0, "V_im": 0.0, "Z_re": 1.0e-3, "Z_im": 0.0}
     "sec_sink": {"V_re": 0.0, "V_im": 0.0, "Z_re": 1.0e-3, "Z_im": 0.0}
 "sweep_solver":
-    "sim_default":
+    "sim_dc":
         "init": null
+        "param":
+            "freq": 0.0
+            "material_val": *material_val
+            "source_val": *source_val
+    "sim_ac":
+        "init": "sim_dc"
         "param":
             "freq": 100.0e+3
             "material_val": *material_val
