### 2024-12-02 49ed78b
add the NumPy FFT (useful for MKL/NumPy)
