+++ examples/config/tolerance.yaml
@@ -41,6 +41,7 @@ parallel_sweep:
     "matrix_split": true
 
     # FFT library
+    #   - "NumPy" FFT library is always available (integrated with NumPy)
     #   - "SciPy" FFT library is always available (integrated with SciPy)
     #   - "FFTW" has to be installed separately (available through pyFFTW)
     #   - "CuPy" is extremely fast but require GPUs compatible with the CUDA toolkit.
