+++ examples/config/tolerance.yaml
@@ -44,6 +44,7 @@ parallel_sweep:
     #   - "NumPy" FFT library is always available (integrated with NumPy)
     #   - "SciPy" FFT library is always available (integrated with SciPy)
     #   - "FFTW" has to be installed separately (available through pyFFTW)
+    #   - "MKL" has to be installed separately (available through mkl_fft)
     #   - "CuPy" is extremely fast but require GPUs compatible with the CUDA toolkit.
     "library": "SciPy"
 
