+++ pypeec/data/schema_tolerance.yaml
@@ -157,6 +157,7 @@
                 "enum":
                     - "NumPy"
                     - "SciPy"
+                    - "MKL"
                     - "FFTW"
                     - "CuPy"
             "scipy_worker":
