+++ pypeec/utils/impedance_matrix.py
@@ -1,188 +1,0 @@
-"""
-Module for extracting the impedance matrix from the solver solution.
-
-The impedance matrix is extracted from the terminal currents and voltages.
-An equation system is formed with all the terminal quantities.
-The least-square solution of the system is computed.
-The impedance matrix is post-processed:
-    - computation of the resistance matrix
-    - computation of the inductance matrix
-    - computation of the quality factors
-    - computation of the couplings
-"""
-
-__author__ = "Thomas Guillod"
-__copyright__ = "Thomas Guillod - Dartmouth College"
-__license__ = "Mozilla Public License Version 2.0"
-
-import numpy as np
-import numpy.linalg as lna
-
-
-def _get_idx_matrix(n_winding):
-    """
-    Mapping of the impedance coefficients into a vector.
-    The indices (row and column) of the impedance coefficients are returned.
-    The impedance matrix is symmetric and only the independent coefficients are considered.
-
-    """
-
-    var_idx = []
-    for i in range(n_winding):
-        for j in range(i+1):
-            var_idx.append([i, j])
-
-    return var_idx
-
-
-def _get_assign_matrix(n_winding, var_idx, sol):
-    """
-    Assign the full impedance matrix from the coefficients.
-    The coefficients are given in a vector.
-    """
-
-    mat = np.zeros((n_winding, n_winding), dtype=np.complex128)
-    for i, (idx_i_1, idx_i_2) in enumerate(var_idx):
-        mat[idx_i_1, idx_i_2] = sol[i]
-        mat[idx_i_2, idx_i_1] = sol[i]
-
-    return mat
-
-
-def _get_eqn_matrix(n_winding, var_idx, I_vec):
-    """
-    Create the equation matrix for a given current excitation.
-    """
-
-    eqn_mat = np.zeros((n_winding, len(var_idx)), dtype=np.complex128)
-    for i, (idx_i_1, idx_i_2) in enumerate(var_idx):
-        eqn_mat[idx_i_1, i] = I_vec[idx_i_2]
-        eqn_mat[idx_i_2, i] = I_vec[idx_i_1]
-
-    return eqn_mat
-
-
-def _get_solve_matrix(terminal):
-    """
-    Extract the impedance matrix of the component.
-
-    The impedance matrix is computed with a linear equation system.
-    The equation system is built with the following variable:
-        - matrix: current excitations
-        - right-hand side: voltage excitations
-        - solution: impedance coefficients
-    """
-
-    # extract the data
-    n_solution = terminal["n_solution"]
-    n_winding = terminal["n_winding"]
-    freq = terminal["freq"]
-    V_mat = terminal["V_mat"]
-    I_mat = terminal["I_mat"]
-
-    # check size
-    assert I_mat.shape == (n_winding, n_solution), "invalid solution: current matrix shape"
-    assert V_mat.shape == (n_winding, n_solution), "invalid solution: voltage matrix shape"
-
-    # get the matrix size and indices of the coefficients
-    var_idx = _get_idx_matrix(n_winding)
-
-    # init the matrices
-    rhs_vec = np.zeros(0, dtype=np.complex128)
-    eqn_mat = np.zeros((0, len(var_idx)), dtype=np.complex128)
-
-    # get the matrices
-    for i in range(n_solution):
-        # extract the data
-        V_vec = V_mat[:, i]
-        I_vec = I_mat[:, i]
-
-        # get the equation matrix
[…]

[…]
