+++ examples/examples_shape/pwm/geometry.yaml
@@ -1,0 +1,75 @@
+# Definition of a "file_geometry" data.
+#
+# This file is used by the mesher.
+# Definition of the voxel structure.
+#
+# Thomas Guillod - Dartmouth College
+# Mozilla Public License Version 2.0
+
+"mesh_type": "shape"
+"data_voxelize":
+    "param":
+        "dx": 50.0e-6
+        "dy": 50.0e-6
+        "dz": 50.0e-6
+        "cz": 0.0e-6
+        "simplify": 1.0e-6
+        "construct": null
+        "xy_min": null
+        "xy_max": null
+    "layer_stack":
+        - {"n_layer": 1, "tag_layer": "copper"}
+    "geometry_shape":
+        "trace":
+            -
+                "shape_layer": ["copper"]
+                "shape_operation": "add"
+                "shape_type": "trace"
+                "shape_data":
+                    "width": 1.75e-3
+                    "coord":
+                        - [0.0e-3, 0.0e-3]
+                        - [14.0e-3, 0.0e-3]
+                        - [14.0e-3, 14.0e-3]
+                        - [0.0e-3, 14.0e-3]
+                        - [0.0e-3, 2.0e-3]
+                        - [12.0e-3, 2.0e-3]
+                        - [12.0e-3, 12.0e-3]
+                        - [2.0e-3, 12.0e-3]
+                        - [2.0e-3, 4.0e-3]
+                        - [10.0e-3, 4.0e-3]
+        "src":
+            -
+                "shape_layer": ["copper"]
+                "shape_operation": "add"
+                "shape_type": "pad"
+                "shape_data":
+                    "diameter": 1.0e-3
+                    "coord": [[0.0e-3, 0.0e-3]]
+        "sink":
+            -
+                "shape_layer": ["copper"]
+                "shape_operation": "add"
+                "shape_type": "pad"
+                "shape_data":
+                    "diameter": 1.0e-3
+                    "coord": [[10.0e-3, 4.0e-3]]
+"data_point":
+    "check_cloud": true
+    "full_cloud": true
+    "pts_cloud": []
+"data_resampling":
+    "use_reduce": false
+    "use_resample": false
+    "resampling_factor": [1, 1, 1]
+"data_conflict":
+    "resolve_rules": true
+    "resolve_random": false
+    "conflict_rules":
+        - {"domain_resolve": ["trace"], "domain_keep": ["src"]}
+        - {"domain_resolve": ["trace"], "domain_keep": ["sink"]}
+"data_integrity":
+    "domain_connected":
+        "conductor": {"domain_group": [["trace"], ["src"], ["sink"]], "connected": true}
+    "domain_adjacent":
+        "terminal": {"domain_group": [["src"], ["sink"]], "connected": false}
