### 2024-12-03 62e3012
add fourier.py for the utils
