+++ examples/examples_stl/transformer/run_matrix.py
@@ -1,5 +1,8 @@
 """
-Convert voxel indices (linear and tensor indices).
+Extract the impedance matrix from the solution:
+    - extract the terminal data (currents and voltages) from the solution
+    - expand the extracted terminal data with the given symmetries
+    - extract the impedance matrix from the terminal data
 """
 
 __author__ = "Thomas Guillod"
