### 2024-12-03 f4fe1ff
example for the Fourier analysis
