+++ examples/examples_shape/pwm/run_fourier.py
@@ -1,0 +1,262 @@
+"""
+Compute non-sinusoidal waveforms with Fourier analysis:
+    - extract the frequency domain impedance
+    - compute the waveforms in the frequency domain
+    - transform back the results in the time domain
+"""
+
+__author__ = "Thomas Guillod"
+__copyright__ = "Thomas Guillod - Dartmouth College"
+__license__ = "Mozilla Public License Version 2.0"
+
+import sys
+import scisave
+import numpy as np
+import matplotlib.pyplot as plt
+from pypeec.utils import fourier
+
+
+def _get_impedance_peec(data_solution):
+    """
+    Extract the resistance and inductance values (from the solution).
+    """
+
+    # extract the data
+    data_init = data_solution["data_init"]
+    data_sweep = data_solution["data_sweep"]
+
+    # check solution
+    assert isinstance(data_init, dict), "invalid solution"
+    assert isinstance(data_sweep, dict), "invalid solution"
+
+    # init the impedance vector
+    f_sim = np.zeros(len(data_sweep), dtype=np.float64)
+    Z_sim = np.zeros(len(data_sweep), dtype=np.complex128)
+
+    # find the impedance for the different frequencies
+    for idx, data_sweep_tmp in enumerate(data_sweep.values()):
+        # extract the data
+        freq = data_sweep_tmp["freq"]
+        source = data_sweep_tmp["source"]
+
+        # extract the impedance
+        voltage = source["src"]["V"]-source["sink"]["V"]
+        current = (source["src"]["I"]-source["sink"]["I"])/2
+        impedance = voltage/current
+
+        # assign
+        f_sim[idx] = freq
+        Z_sim[idx] = impedance
+
+    # extract the equivalent circuit
+    R_sim = np.real(Z_sim)
+    L_sim = np.imag(Z_sim)/(2*np.pi*f_sim)
+
+    return f_sim, R_sim, L_sim
+
+
+def _get_impedance_fourier(f_vec, f_sim, R_sim, L_sim):
+    """
+    Interpolate the impedance at the Fourier frequencies.
+    """
+
+    # clip to prevent extrapolation
+    f_vec = np.clip(f_vec, np.min(f_sim), np.max(f_sim))
+
+    # interpolate the circuit parameters (log-scale frequency)
+    R_freq = np.interp(np.log10(f_vec), np.log10(f_sim), R_sim)
+    L_freq = np.interp(np.log10(f_vec), np.log10(f_sim), L_sim)
+
+    # assemble the impedance
+    Z_freq = R_freq+(1j*2*np.pi*f_vec)*L_freq
+
+    return Z_freq
+
+
+def _plot_circuit(f_sim, R_sim, L_sim):
+    """
+    Plot the extracted resistance and inductance values.
+    """
+
+    (fig, ax) = plt.subplots(2, 1)
+
+    ax[0].semilogx(f_sim, 1e3*R_sim, "x-")
+    ax[0].set_xlabel("f (Hz)")
+    ax[0].set_ylabel("R (mOhm)")
+    ax[0].set_title("Resistance")
+    ax[0].grid()
+
+    ax[1].semilogx(f_sim, 1e9*L_sim, "x-")
+    ax[1].set_xlabel("f (Hz)")
+    ax[1].set_ylabel("L (nH)")
+    ax[1].set_title("Inductance")
+    ax[1].grid()
+
+    fig.tight_layout()
+
+
+def _plot_frequency(f_vec, V_freq, I_freq, n_plot):
+    """
+    Plot the voltage and current spectrums.
+    """
+
+    f_vec = f_vec[:n_plot]
+    V_freq = V_freq[:n_plot]
+    I_freq = I_freq[:n_plot]
+
[…]
