### 2024-12-03 78a7427
update the documentation with the examples
