+++ docs/content/examples.rst
@@ -33,32 +33,34 @@ STL Mesher Examples
 examples_stl/inductor_air
 ^^^^^^^^^^^^^^^^^^^^^^^^^
 
-* **3D air-core inductor**
 * Defined with **STL files**
+* **3D air-core spiral inductor**
 
 .. image:: ../examples/examples_stl_inductor_air.png
 
 examples_stl/inductor_core
 ^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-* **3D inductor with a magnetic E-core**
 * Defined with **STL files**
+* **3D inductor with a magnetic E-core**
 
 .. image:: ../examples/examples_stl_inductor_core.png
 
 examples_stl/inductor_toroid
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-* **3D inductor with a toroidal core**
 * Defined with **STL files**
+* **3D inductor with a toroidal core**
 
 .. image:: ../examples/examples_stl_inductor_toroid.png
 
 examples_stl/transformer
 ^^^^^^^^^^^^^^^^^^^^^^^^
 
-* **Planar transformer with two windings**
 * Defined with **STL files**
+* **Planar transformer with two windings**
+* Extraction of the **impedance matrix** from the solution
+* Example for the extraction of the impedance matrix: ``run_matrix.py``
 
 .. image:: ../examples/examples_stl_transformer.png
 
@@ -68,42 +70,53 @@ Shape Mesher Examples
 examples_shape/busbar
 ^^^^^^^^^^^^^^^^^^^^^
 
-* **Coplanar L-shaped busbar**
 * Defined with **2D shapes**
+* **Coplanar L-shaped busbar**
 
 .. image:: ../examples/examples_shape_busbar.png
 
 examples_shape/coplanar
 ^^^^^^^^^^^^^^^^^^^^^^^
 
-* **PCB with coplanar traces**
 * Defined with **2D shapes**
+* **PCB with coplanar traces**
 
 .. image:: ../examples/examples_shape_coplanar.png
 
 examples_shape/parallel
 ^^^^^^^^^^^^^^^^^^^^^^^
 
-* **Parallel wires with opposing currents**
 * Defined with **2D shapes**
+* **Parallel wires with opposing currents**
 
 .. image:: ../examples/examples_shape_parallel.png
 
 examples_shape/hole
 ^^^^^^^^^^^^^^^^^^^
 
-* **Trace with a hole**
 * Defined with **2D shapes**
+* **Single PCB trace with a hole**
 
 .. image:: ../examples/examples_shape_hole.png
 
 examples_shape/wire
 ^^^^^^^^^^^^^^^^^^^
 
-* **Straight round wire**
 * Defined with **2D shapes**
+* **Single straight round wire**
 
 .. image:: ../examples/examples_shape_wire.png
+
+examples_shape/pwm
+^^^^^^^^^^^^^^^^^^
+
+* Defined with **2D shapes**
+* **Planar spiral inductor**
+* **Non-sinusoidal** analysis with **Fourier** series
+* Example for the frequency domain analysis: ``run_fourier.py``
+
+.. image:: ../examples/examples_shape_pwm.png
+
 
 PNG Mesher Examples
 -------------------
@@ -111,48 +124,50 @@ PNG Mesher Examples
 examples_png/inductor_spiral
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-* **Planar spiral inductor**
 * Defined with **PNG files**
+* **Planar spiral inductor**
 
 .. image:: ../examples/examples_png_inductor_spiral.png
 
 examples_png/inductor_gap
 ^^^^^^^^^^^^^^^^^^^^^^^^^
 
-* **Gapped inductor with a magnetic E-core**
 * Defined with **PNG files**
[…]
