### 2024-12-03 87cdb4f
refactoring and better documentation
