### 2024-12-03 b872b50
bug for the Fourier analysis
