+++ examples/run_solver.py
@@ -6,7 +6,6 @@ __author__ = "Thomas Guillod"
 __copyright__ = "Thomas Guillod - Dartmouth College"
 __license__ = "Mozilla Public License Version 2.0"
 
-import sys
 import os.path
 import pypeec
 import examples_config
