### 2024-12-04 8a5da9e
better Fourier example
