+++ examples/examples_shape/pwm/run_fourier.py
@@ -80,13 +80,13 @@ def _plot_circuit(f_sim, R_sim, L_sim):
 
     (fig, ax) = plt.subplots(2, 1)
 
-    ax[0].semilogx(f_sim, 1e3*R_sim, "x-")
+    ax[0].semilogx(f_sim, 1e3*R_sim, "o-")
     ax[0].set_xlabel("f (Hz)")
     ax[0].set_ylabel("R (mOhm)")
     ax[0].set_title("Resistance")
     ax[0].grid()
 
-    ax[1].semilogx(f_sim, 1e9*L_sim, "x-")
+    ax[1].semilogx(f_sim, 1e9*L_sim, "o-")
     ax[1].set_xlabel("f (Hz)")
     ax[1].set_ylabel("L (nH)")
     ax[1].set_title("Inductance")
@@ -106,16 +106,16 @@ def _plot_frequency(f_vec, V_freq, I_freq, n_plot):
 
     (fig, ax) = plt.subplots(2, 1)
 
-    ax[0].plot(1e-3*f_vec, np.abs(V_freq), "-")
+    ax[0].plot(1e-3*f_vec, np.abs(V_freq), "o-")
     ax[0].set_xlabel("f (kHz)")
     ax[0].set_ylabel("V (V)")
-    ax[0].set_title("Voltage")
+    ax[0].set_title("Voltage - Frequency Domain")
     ax[0].grid()
 
-    ax[1].plot(1e-3*f_vec, np.abs(I_freq), "-")
+    ax[1].plot(1e-3*f_vec, np.abs(I_freq), "o-")
     ax[1].set_xlabel("f (kHz)")
     ax[1].set_ylabel("I (A)")
-    ax[1].set_title("Current")
+    ax[1].set_title("Current - Frequency Domain")
     ax[1].grid()
 
     fig.tight_layout()
@@ -132,16 +132,16 @@ def _plot_power(f_vec, P_vec, Q_vec, n_plot):
 
     (fig, ax) = plt.subplots(2, 1)
 
-    ax[0].plot(1e-3*f_vec, np.abs(P_vec), "-")
+    ax[0].plot(1e-3*f_vec, np.abs(P_vec), "o-")
     ax[0].set_xlabel("f (kHz)")
     ax[0].set_ylabel("P (W)")
-    ax[0].set_title("Active Power")
+    ax[0].set_title("Active Power - Frequency Domain")
     ax[0].grid()
 
-    ax[1].plot(1e-3*f_vec, np.abs(Q_vec), "-")
+    ax[1].plot(1e-3*f_vec, np.abs(Q_vec), "o-")
     ax[1].set_xlabel("f (kHz)")
     ax[1].set_ylabel("Q (VA)")
-    ax[1].set_title("Reative Power")
+    ax[1].set_title("Reative Power - Frequency Domain")
     ax[1].grid()
 
     fig.tight_layout()
@@ -157,13 +157,13 @@ def _plot_waveform(t_vec, V_time, I_time):
     ax[0].plot(1e6*t_vec, V_time, "-")
     ax[0].set_xlabel("t (us)")
     ax[0].set_ylabel("V (V)")
-    ax[0].set_title("Voltage")
+    ax[0].set_title("Voltage - Time Domain")
     ax[0].grid()
 
     ax[1].plot(1e6*t_vec, I_time, "-")
     ax[1].set_xlabel("t (us)")
     ax[1].set_ylabel("I (A)")
-    ax[1].set_title("Current")
+    ax[1].set_title("Current - Time Domain")
     ax[1].grid()
 
     fig.tight_layout()
@@ -188,21 +188,22 @@ if __name__ == "__main__":
     # ########################################################
 
     # parameters for the signal
-    d_pwm = 0.1     # duty cycle for the PWM voltage
[…]
