### 2024-12-04 0381052
instructions for running the utils examples / better Fourier example
