### 2024-12-04 09b71f3
instructions for running the utils examples
