### 2024-12-04 bbe4384
documentation for the geometry preparation
