+++ docs/content/workflow.rst
@@ -85,13 +85,31 @@ Input/Output File Description
 
 The following input files (JSON or YAML format) are used:
 
-* ``file_geometry`` (JSON or YAML, input/configuration file): description of the geometry
-* ``file_problem`` (JSON or YAML, input/configuration file): description of the magnetic problem
-* ``file_tolerance`` (JSON or YAML, input/configuration file): description of the solver numerical options
-* ``file_viewer`` (JSON or YAML, input/configuration file): options for the plots generated by the viewer
-* ``file_plotter`` (JSON or YAML, input/configuration file): options for the plots generated by the plotter
+* ``file_geometry``: description of the geometry
+* ``file_problem``: description of the magnetic problem
+* ``file_tolerance``: description of the solver numerical options
+* ``file_viewer``: options for the plots generated by the viewer
+* ``file_plotter``: options for the plots generated by the plotter
 
 The following files (JSON or Pickle format) are generated:
 
-* ``file_voxel`` (JSON or Pickle, output/data file): file with the definition of the voxel structure
-* ``file_solution`` (JSON or Pickle, output/data file): file with the solution of the magnetic problem
+* ``file_voxel``: file with the definition of the voxel structure
+* ``file_solution``: file with the solution of the magnetic problem
+
+Geometry Preparation
+--------------------
+
+The following **open-source tools** can be used to generate **PNG** files:
+
+* Interactive software: ``GIMP`` or ``Inkscape``
+* Scripting software: ``ImageMagick`` or ``Pillow``
+
+The following **open-source tools** can be used to generate **STL** files:
+
+* Interactive software: ``FreeCAD`` or ``Blender``
+* Scripting software: ``CadQuery`` or ``OpenSCAD``
+
+The following **open-source tools** can be used to generate **GERBER** files:
+
+* Interactive software: ``KiCAD``
+* Scripting software: ``gerber-writer``
