### 2024-12-04 #6644 @otvam
@svenweihe The new **time-domain example** has been merged into the `main` branch:

* `examples/examples_shape/pwm/run_fourier.py`
* extract the frequency domain impedance of a spiral inductor
* apply a PWM voltage to the spiral inductor terminals
* compute the current waveform in the frequency domain
* transform back the results in the time domain

Here are the results (frequency domain power spectrums and time domain waveforms):

![fig_freq](https://github.com/user-attachments/assets/fa78f80c-ae73-46fa-a9da-bc6ea9b3ac82)
![fig_time](https://github.com/user-attachments/assets/e4db4b86-21e5-44c5-962f-b2955e1f034b)

Another typical power electronics example has been added for the extraction of transformer **impedance matrices**:

* `examples/examples_stl/transformer/run_matrix.py`
* extract the terminal data (currents and voltages) from the solution
* expand the extracted terminal data with the given symmetries
* extract the impedance matrix from the terminal data
