### 2024-12-04 e3a05eb
better conda file for CI
