+++ .github/envs/env_conda.txt
@@ -1,4 +1,4 @@
-# Definition of the Python Virtual Environment (minimum versions).
+# Definition of the Conda Environment for PyPEEC (minimum versions).
 #
 # Thomas Guillod - Dartmouth College
 # Mozilla Public License Version 2.0
@@ -9,18 +9,18 @@ scisave>=1.4
 jsonschema>=4.23
 
 # numerical libraries
-joblib>=1.3
 numpy>=1.24
 scipy>=1.12
+joblib>=1.3
 
-# geometry manipulation
+# geometry libraries
 vtk>=9.2
 pyvista>=0.40
 shapely>=2.0
 rasterio>=1.3
-Pillow>=10.0
+pillow>=10.0
 
-# plottinglibraries
-PyQt5>=5.15
+# plotting libraries
+pyqt>=5.15
 pyvistaqt>=0.11
 matplotlib>=3.7
