### 2024-12-04 4e61e63
better conda file for CI
