+++ scripts/run_tests.sh
@@ -34,7 +34,7 @@ export PYTHONIOENCODING="utf8"    # set encoding for the console
 export PYTHONUNBUFFERED="1"       # disable buffering for the console
 
 # options for the tests
-export TEST_TOL="1e-5"            # relative tolerance for the test results
+export TEST_TOL="1e-4"            # relative tolerance for the test results
 export TEST_CHECK="1"             # check (or not) the test results
 export TEST_SET="0"               # generate (or not) the test results
 
