+++ .github/envs/requirements.txt
@@ -13,7 +13,7 @@ numpy==2.1.3
 scipy==1.14.1
 joblib==1.4.2
 
-# geometry manipulation
+# geometry libraries
 vtk==9.3.1
 pyvista==0.44.2
 shapely==2.0.6
