+++ .github/envs/env_pypi.txt
@@ -9,18 +9,18 @@ scisave>=1.4
 jsonschema>=4.23
 
 # numerical libraries
-joblib>=1.3
 numpy>=1.24
 scipy>=1.12
+joblib>=1.3
 
-# geometry manipulation
+# geometry libraries
 vtk>=9.2
 pyvista>=0.40
 shapely>=2.0
 rasterio>=1.3
-Pillow>=10.0
+pillow>=10.0
 
-# plottinglibraries
+# plotting libraries
 PyQt5>=5.15
 pyvistaqt>=0.11
 matplotlib>=3.7
