### 2024-12-05 beb65db
consistent CI files
