### 2024-12-13 #6644 @svenweihe
@otvam Thanks for providing the updated examples. Apologies for the slow response, I will check them out this weekend.
I wanted to update pypeec but not super familiar with python. Is there a function I can call to update the software to the latest version? If so, it would be great if this could also be included as an additional line on the installation page of the website.
