### 2024-12-14 #6644 @otvam
Great! For updating the package, this is depending which package manager is being used:
* with pip: `pip install pypeec==5.3.3`
* with conda: `mamba install conda-forge:pypeec==5.3.3`

However, the Python packaging ecosystem is quite famous to be messy and prone to issues... So, very often, the simplest variant is to delete the old environment and reinstall from scratch:
* with pip: removing the virtual environment folder is sufficient
* with conda: using the `mamba remove` command with the name of the environment
