### 2025-03-31 #6644 @likeajumprope
> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> ✅ OK DOIs
>
> - None
>
> 🟡 SKIP DOIs
>
> - No DOI given, and none found for title: Free-Shape Optimization of VHF Air-Core Inductors ...
> - No DOI given, and none found for title: Circuit Oriented Electromagnetic Modeling Using th...
> - No DOI given, and none found for title: FFT-PEEC
> - No DOI given, and none found for title: FFT-NLIE
> - No DOI given, and none found for title: VoxHenry
> - No DOI given, and none found for title: MARIE
>
> ❌ MISSING DOIs
>
> - 10.6028/jres.069c.016 may be a valid DOI for title: Exact Inductance Equations for Rectangular Conduct...
> - 10.2528/pierm12080314 may be a valid DOI for title: Exact Closed Form Formula for Self Inductance of C...
> - 10.1109/tpel.2021.3092431 may be a valid DOI for title: FFT-PEEC: A Fast Tool From CAD to Power Electronic...
> - 10.1016/j.jcp.2014.03.026 may be a valid DOI for title: Stable FFT-JVIE Solvers for Fast Analysis of Highl...
> - 10.1016/0885-064x(86)90007-5 may be a valid DOI for title: Fast Multiplication of a Recursive Block Toeplitz ...
> - 10.1109/temc.2023.3238394 may be a valid DOI for title: Efficient PEEC Iterative Solver for Power Electron...
> - 10.1109/memc.2024.10834417 may be a valid DOI for title: Equivalent Circuit Models for Three-Dimensional Mu...
> - 10.1109/tap.2019.2927789 may be a valid DOI for title: A Volume PEEC Formulation Based on the Cell Method...
> - 10.1109/tmtt.2017.2785842 may be a valid DOI for title: VoxHenry: FFT-Accelerated Inductance Extraction fo...
> - 10.1088/1361-6587/abce8f may be a valid DOI for title: Fast Fourier Transform-Volume Integral: a Smart Ap...
> - 10.1016/j.cmpb.2021.106543 may be a valid DOI for title: A Fast Tool for the Parametric Analysis of Human B...
> - 10.1109/iceccme52200.2021.9590864 may be a valid DOI for title: Application of FFT-PEEC Method for Nonlinear Induc...
> - 10.1109/icemic.2003.1287800 may be a valid DOI for title: A Review on Computational EMI Modelling Techniques
>
> ❌ INVALID DOIs
>
> - None
> ```

References are correct, just in a different format (starting with https://doi.org) which is why editorialbot won't recognize them
