### 2025-03-31 #6644 @otvam
@likeajumprope

The authors, affiliations, licence, and title are correct.
The missings DOIs have been added in the bib file.

The zenodo archive for the software is:
https://doi.org/10.5281/zenodo.15028917

The version number is: 5.6.0
