### 2025-04-13 #6644 @otvam
@likeajumprope

I've just published version **5.6.1** that correct a small bug in the command line interface. The new version is also archived:
https://doi.org/10.5281/zenodo.15206714
