### 2025-04-27 #6644 @otvam
@likeajumprope

Version **5.7.0** is out; offering a better control for the preconditioner and better voxelization of vector shapes. The DOI of this version is: https://doi.org/10.5281/zenodo.15276514
