### 2025-05-08 #6644 @likeajumprope
> 🔴 Failed to discover a valid open source license

I will check that
