### 2025-05-12 #6644 @otvam
The paper is located is a separate git branch (called "paper_joss"). So the bot does not find the license that is located in the "main" branch.

> @editorialbot` check repository from branch main
