### 2025-07-22 #6644 @otvam
@likeajumprope

Just wanted to ask if there is anything left preventing the paper to be published?
