### 2025-07-22 #6644 @likeajumprope
> [@likeajumprope](https://github.com/likeajumprope)
>
> Just wanted to ask if there is anything left preventing the paper to be published?

thank you for the ping @otvam  - I am back from holiday and will have a look!
