### 2025-08-07 #6644 @otvam
@likeajumprope

Version **5.7.3** is making PyPEEC compatible with the latest VTK version.
https://doi.org/10.5281/zenodo.16764265
