### 2025-09-01 #6644 @otvam
@likeajumprope
I checked the proof and found two small typos (sorry!). The corrected version is in the branch "paper_joss". Otherwise, everything is OK (links, archive, version, etc.).
